Cardano.CLI.Compatible.Exception

Documentation

type CIO e a = HasCallStack => RIO e a

Type alias that enforces the presence of a call stack.

data CustomCliException where

Custom exception type for CLI commands. Any custom errors created in `cardano-cl` should be wrapped in this exception type.

Constructors

CustomCliException :: forall err. (HasCallStack, Show err, Typeable err, Error err) => err -> CustomCliException

fromEitherCIOCli :: (Show err, Typeable err, Error err) => CIO e (Either err a) -> CIO e a

fromMaybeCli :: (Show err, Typeable err, Error err) => err -> Maybe a -> CIO e a

throwCliError :: (HasCallStack, Show e, Typeable e, Error e, MonadIO m) => e -> m a

Wrapper function which allows throwing of types of instance Error, attaching call stack from the call site

fromEitherCli :: forall e m a. (HasCallStack, MonadIO m, Show e, Typeable e, Error e) => Either e a -> m a

fromEitherIOCli :: forall e m a. (HasCallStack, MonadIO m, Show e, Typeable e, Error e) => IO (Either e a) -> m a

fromExceptTCli :: forall e m a. (HasCallStack, MonadIO m, Show e, Typeable e, Error e) => ExceptT e IO a -> m a