Cardano.CLI.Compatible.Governance.Types

Documentation

data GovernanceActionProtocolParametersUpdateCmdArgs era

Constructors

GovernanceActionProtocolParametersUpdateCmdArgs

Fields

uppShelleyBasedEra :: !(ShelleyBasedEra era)

uppPreConway :: !(Maybe (UpdateProtocolParametersPreConway era))

uppConwayOnwards :: !(Maybe (UpdateProtocolParametersConwayOnwards era))

uppNewPParams :: !(EraBasedProtocolParametersUpdate era)

New parameters to be proposed. From Alonzo onwards, the type EraBasedProtocolParametersUpdate also contains cost models. Since all other protocol parameters are read from command line arguments, whereas the cost models are read from a file, we separate the cost models from the rest of the protocol parameters to ease parsing.

uppCostModelsFile :: !(Maybe (CostModelsFile era))

The new cost models proposed. See the comment at uppNewPParams for why this is a separate field.

uppFilePath :: !(File () 'Out)

data UpdateProtocolParametersPreConway era

Constructors

UpdateProtocolParametersPreConway

Fields

eon :: !(ShelleyToBabbageEra era)

expiryEpoch :: !EpochNo

genesisVerificationKeys :: ![VerificationKeyFile 'In]

data UpdateProtocolParametersConwayOnwards era

Constructors

UpdateProtocolParametersConwayOnwards

Fields

era :: !(Era era)

networkId :: !Network

deposit :: !Lovelace

returnAddr :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

governanceActionId :: !(Maybe GovActionId)

constitutionScriptHash :: !(Maybe ScriptHash)

data CostModelsFile era

Constructors

CostModelsFile

Fields

eon :: !(AlonzoEraOnwards era)

costModelsFile :: !(File CostModels 'In)