Cardano.CLI.EraBased.Commands.Governance.Hash

Documentation

data GovernanceHashCmds era

Constructors

GovernanceHashAnchorDataCmd !(GovernanceHashAnchorDataCmdArgs era)

GovernanceHashScriptCmd !(GovernanceHashScriptCmdArgs era)

data GovernanceHashAnchorDataCmdArgs era

Constructors

GovernanceHashAnchorDataCmdArgs

Fields

eon ∷ !(ConwayEraOnwards era)

toHash ∷ !GovernanceAnchorDataHashSource

moutFile ∷ !(Maybe (File () Out))

The output file to which the hash is written

data GovernanceHashScriptCmdArgs era

Constructors

GovernanceHashScriptCmdArgs

Fields

eon ∷ !(ConwayEraOnwards era)

toHash ∷ !ScriptFile

moutFile ∷ !(Maybe (File () Out))

The output file to which the hash is written

data GovernanceAnchorDataHashSource

Constructors

GovernanceAnchorDataHashSourceBinaryFile (File ProposalBinary In)

GovernanceAnchorDataHashSourceTextFile (File ProposalText In)

GovernanceAnchorDataHashSourceText Text

renderGovernanceHashCmds ∷ GovernanceHashCmds era → Text