Cardano.CLI.EraBased.Commands.Transaction

Documentation

data TransactionCmds era

Constructors

TransactionBuildRawCmd !(TransactionBuildRawCmdArgs era)

TransactionBuildCmd !(TransactionBuildCmdArgs era)

TransactionBuildEstimateCmd !(TransactionBuildEstimateCmdArgs era)

TransactionSignCmd !TransactionSignCmdArgs

TransactionWitnessCmd !TransactionWitnessCmdArgs

TransactionSignWitnessCmd !TransactionSignWitnessCmdArgs

TransactionSubmitCmd !TransactionSubmitCmdArgs

TransactionPolicyIdCmd !TransactionPolicyIdCmdArgs

TransactionCalculateMinFeeCmd !TransactionCalculateMinFeeCmdArgs

TransactionCalculateMinValueCmd !(TransactionCalculateMinValueCmdArgs era)

TransactionHashScriptDataCmd !TransactionHashScriptDataCmdArgs

TransactionTxIdCmd !TransactionTxIdCmdArgs

data TransactionBuildRawCmdArgs era

Constructors

TransactionBuildRawCmdArgs

Fields

eon :: !(ShelleyBasedEra era)

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

txIns :: ![(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

readOnlyRefIns :: ![TxIn]

Read only reference inputs

txInsCollateral :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

mTotalCollateral :: !(Maybe Coin)

Total collateral

requiredSigners :: ![RequiredSigner]

Required signers

txouts :: ![TxOutAnyEra]

mValue :: !(Maybe (Value, [CliMintScriptRequirements]))

Multi-Asset value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

fee :: !Coin

Transaction fee

certificates :: ![(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))]

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

mProtocolParamsFile :: !(Maybe ProtocolParamsFile)

mUpdateProprosalFile :: !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

voteFiles :: ![(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

currentTreasuryValueAndDonation :: !(Maybe (TxCurrentTreasuryValue, TxTreasuryDonation))

txBodyOutFile :: !(TxBodyFile 'Out)

data TransactionBuildCmdArgs era

Like TransactionBuildRaw but without the fee, and with a change output.

Constructors

TransactionBuildCmdArgs

Fields

currentEra :: !(Era era)

nodeSocketPath :: !SocketPath

consensusModeParams :: !ConsensusModeParams

networkId :: !NetworkId

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

mOverrideWitnesses :: !(Maybe Word)

Override the required number of tx witnesses

txins :: ![(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

readOnlyReferenceInputs :: ![TxIn]

Read only reference inputs

requiredSigners :: ![RequiredSigner]

Required signers

txinsc :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

mTotalCollateral :: !(Maybe Coin)

Total collateral

txouts :: ![TxOutAnyEra]

Normal outputs

changeAddresses :: !TxOutChangeAddress

A change output

mValue :: !(Maybe (Value, [CliMintScriptRequirements]))

Multi-Asset value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

certificates :: ![(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))]

Withdrawals with potential script witness

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

mUpdateProposalFile :: !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

voteFiles :: ![(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

treasuryDonation :: !(Maybe TxTreasuryDonation)

buildOutputOptions :: !TxBuildOutputOptions

data TransactionBuildEstimateCmdArgs era

Like TransactionBuildCmd but does not require explicit access to a running node

Constructors

TransactionBuildEstimateCmdArgs

Fields

currentEra :: !(Era era)

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

shelleyWitnesses :: !Int

Number of shelley witnesses to be added

mByronWitnesses :: !(Maybe Int)

protocolParamsFile :: !ProtocolParamsFile

totalUTxOValue :: !Value

txins :: ![(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

readOnlyReferenceInputs :: ![TxIn]

Read only reference inputs

requiredSigners :: ![RequiredSigner]

Required signers

txinsc :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

txouts :: ![TxOutAnyEra]

Normal outputs

changeAddress :: !TxOutChangeAddress

A change output

mValue :: !(Maybe (Value, [CliMintScriptRequirements]))

Multi-Asset value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

certificates :: ![(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))]

Withdrawals with potential script witness

plutusCollateral :: !(Maybe Coin)

Total collateral

totalReferenceScriptSize :: !(Maybe ReferenceScriptSize)

Size of all reference scripts in bytes

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

mUpdateProposalFile :: !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

voteFiles :: ![(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

currentTreasuryValueAndDonation :: !(Maybe (TxCurrentTreasuryValue, TxTreasuryDonation))

txBodyOutFile :: !(TxBodyFile 'Out)

data TransactionSignCmdArgs

Constructors

TransactionSignCmdArgs

Fields

txOrTxBodyFile :: !InputTxBodyOrTxFile

witnessSigningData :: ![WitnessSigningData]

mNetworkId :: !(Maybe NetworkId)

outTxFile :: !(TxFile 'Out)

data TransactionWitnessCmdArgs

Constructors

TransactionWitnessCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

witnessSigningData :: !WitnessSigningData

mNetworkId :: !(Maybe NetworkId)

outFile :: !(File () 'Out)

data TransactionSignWitnessCmdArgs

Constructors

TransactionSignWitnessCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

witnessFiles :: ![WitnessFile]

outFile :: !(File () 'Out)

data TransactionSubmitCmdArgs

Constructors

TransactionSubmitCmdArgs

Fields

nodeSocketPath :: !SocketPath

consensusModeParams :: !ConsensusModeParams

networkId :: !NetworkId

txFile :: !FilePath

newtype TransactionPolicyIdCmdArgs

Constructors

TransactionPolicyIdCmdArgs

Fields

scriptFile :: ScriptFile

data TransactionCalculateMinFeeCmdArgs

Constructors

TransactionCalculateMinFeeCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

protocolParamsFile :: !ProtocolParamsFile

txShelleyWitnessCount :: !TxShelleyWitnessCount

txByronWitnessCount :: !TxByronWitnessCount

referenceScriptSize :: !ReferenceScriptSize

The total size in bytes of the transaction reference scripts.

outputFormat :: !(Maybe OutputFormatJsonOrText)

outFile :: !(Maybe (File () 'Out))

data TransactionCalculateMinValueCmdArgs era

Constructors

TransactionCalculateMinValueCmdArgs

Fields

eon :: !(ShelleyBasedEra era)

protocolParamsFile :: !ProtocolParamsFile

txOut :: !TxOutShelleyBasedEra

newtype TransactionHashScriptDataCmdArgs

Constructors

TransactionHashScriptDataCmdArgs

Fields

scriptDataOrFile :: ScriptDataOrFile

newtype TransactionTxIdCmdArgs

Constructors

TransactionTxIdCmdArgs

Fields

inputTxBodyOrTxFile :: InputTxBodyOrTxFile

data TransactionViewCmdArgs

Constructors

TransactionViewCmdArgs

renderTransactionCmds :: TransactionCmds era -> Text