Cardano.CLI.EraBased.Genesis.Command

Documentation

data GenesisCmds era

Constructors

GenesisCreate !(GenesisCreateCmdArgs era)

GenesisCreateCardano !(GenesisCreateCardanoCmdArgs era)

GenesisCreateStaked !(GenesisCreateStakedCmdArgs era)

GenesisCreateTestNetData !GenesisCreateTestNetDataCmdArgs

GenesisKeyGenGenesis !GenesisKeyGenGenesisCmdArgs

GenesisKeyGenDelegate !GenesisKeyGenDelegateCmdArgs

GenesisKeyGenUTxO !GenesisKeyGenUTxOCmdArgs

GenesisCmdKeyHash !(VerificationKeyFile 'In)

GenesisVerKey !GenesisVerKeyCmdArgs

GenesisTxIn !GenesisTxInCmdArgs

GenesisAddr !GenesisAddrCmdArgs

GenesisHashFile !GenesisFile

data GenesisCreateCmdArgs era

Constructors

GenesisCreateCmdArgs

Fields

eon :: !AnyShelleyBasedEra

keyOutputFormat :: !(Vary '[FormatBech32, FormatTextEnvelope])

genesisDir :: !GenesisDir

numGenesisKeys :: !Word

numUTxOKeys :: !Word

mSystemStart :: !(Maybe SystemStart)

mSupply :: !(Maybe Coin)

network :: !NetworkId

data GenesisCreateCardanoCmdArgs era

Constructors

GenesisCreateCardanoCmdArgs

Fields

eon :: !(Era era)

genesisDir :: !GenesisDir

numGenesisKeys :: !Word

numUTxOKeys :: !Word

mSystemStart :: !(Maybe SystemStart)

mSupply :: !(Maybe Coin)

security :: !(NonZero Word64)

slotLength :: !Word

slotCoeff :: !Rational

network :: !NetworkId

byronGenesisTemplate :: !FilePath

shelleyGenesisTemplate :: !FilePath

alonzoGenesisTemplate :: !FilePath

conwayGenesisTemplate :: !FilePath

mNodeConfigTemplate :: !(Maybe FilePath)

data GenesisCreateStakedCmdArgs era

Constructors

GenesisCreateStakedCmdArgs

Fields

eon :: !(Era era)

keyOutputFormat :: !(Vary '[FormatBech32, FormatTextEnvelope])

genesisDir :: !GenesisDir

numGenesisKeys :: !Word

numUTxOKeys :: !Word

numPools :: !Word

numStakeDelegators :: !Word

mSystemStart :: !(Maybe SystemStart)

mNonDelegatedSupply :: !(Maybe Coin)

delegatedSupply :: !Coin

network :: !NetworkId

numBulkPoolCredFiles :: !Word

numBulkPoolsPerFile :: !Word

numStuffedUtxo :: !Word

mStakePoolRelaySpecFile :: !(Maybe FilePath)

Relay specification filepath

data GenesisCreateTestNetDataCmdArgs

Constructors

GenesisCreateTestNetDataCmdArgs

Fields

eon :: !(Era era)

specShelley :: !(Maybe FilePath)

Path to the genesis-shelley file to use. If unspecified, a default one will be used.

specAlonzo :: !(Maybe FilePath)

Path to the genesis-alonzo file to use. If unspecified, a default one will be used.

specConway :: !(Maybe FilePath)

Path to the genesis-conway file to use. If unspecified, a default one will be used.

specDijkstra :: !(Maybe FilePath)

Path to the genesis-dijkstra file to use. If unspecified, a default one will be used.

numGenesisKeys :: !Word

The number of genesis keys credentials to create and write to disk.

numPools :: !Word

The number of stake pools credentials to create and write to disk.

stakeDelegators :: !StakeDelegators

The number of members of the constitutional committee

numCommitteeKeys :: !Word

The number of delegators to pools and DReps to create.

numDRepKeys :: !DRepCredentials

The number of DRep keys to create. They are registered and get delegated to by stake delegators

numStuffedUtxo :: !Word

The number of UTxO accounts to make. They are "stuffed" because the credentials are not written to disk.

numUtxoKeys :: !Word

The number of UTxO credentials to create and write to disk.

totalSupply :: !(Maybe Coin)

The total number of Lovelace

delegatedSupply :: !(Maybe Coin)

The number of Lovelace being delegated

networkId :: !(Maybe NetworkId)

The network ID to use. Overrides the network id supplied in the spec file.

relays :: !(Maybe FilePath)

Filepath of the specification of relays

systemStart :: !(Maybe SystemStart)

The genesis start time.

outputDir :: !FilePath

Directory where to write credentials and files.

data GenesisKeyGenGenesisCmdArgs

Constructors

GenesisKeyGenGenesisCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'Out)

signingKeyPath :: !(SigningKeyFile 'Out)

data GenesisKeyGenDelegateCmdArgs

Constructors

GenesisKeyGenDelegateCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'Out)

signingKeyPath :: !(SigningKeyFile 'Out)

opCertCounterPath :: !(OpCertCounterFile 'Out)

data GenesisKeyGenUTxOCmdArgs

Constructors

GenesisKeyGenUTxOCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'Out)

signingKeyPath :: !(SigningKeyFile 'Out)

data GenesisVerKeyCmdArgs

Constructors

GenesisVerKeyCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'Out)

signingKeyPath :: !(SigningKeyFile 'In)

data GenesisTxInCmdArgs

Constructors

GenesisTxInCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'In)

network :: !NetworkId

mOutFile :: !(Maybe (File () 'Out))

data GenesisAddrCmdArgs

Constructors

GenesisAddrCmdArgs

Fields

verificationKeyPath :: !(VerificationKeyFile 'In)

network :: !NetworkId

mOutFile :: !(Maybe (File () 'Out))

renderGenesisCmds :: GenesisCmds era -> Text