Cardano.CLI.EraBased.Governance.Actions.Option

Documentation

pGovActionProtocolParametersUpdate :: ShelleyBasedEra era -> Parser (EraBasedProtocolParametersUpdate era)

pCostModelsFile :: ShelleyBasedEra era -> Parser (Maybe (CostModelsFile era))

Cost models only makes sense in eras from Alonzo onwards. For earlier eras, this parser doesn't show up in the command line and returns Nothing.

pGovernanceActionCmds :: IsEra era => Maybe (Parser (GovernanceActionCmds era))

pProtocolParametersUpdateGenesisKeys :: Parser [VerificationKeyFile 'In]

pUpdateProtocolParametersPostConway :: IsEra era => Parser (UpdateProtocolParametersConwayOnwards era)