Cardano.CLI.EraBased.Governance.Actions.Run

Documentation

runGovernanceActionCmds :: GovernanceActionCmds era -> CIO e ()

data GovernanceActionsError

Constructors

GovernanceActionsCmdCostModelsError CostModelsError

GovernanceActionsCmdWriteFileError (FileError ())

GovernanceActionsValueUpdateProtocolParametersNotFound AnyShelleyBasedEra

GovernanceActionsMismatchedHashError

Fields

AnchorDataTypeCheck

Type of anchor data that we were checking

!(SafeHash AnchorData)

Expected hash

!(SafeHash AnchorData)

Actual hash

GovernanceActionsProposalFetchURLError

Fields

AnchorDataTypeCheck

Type of anchor data that we were checking

FetchURLError

Error that occurred while fetching the anchor data

addCostModelsToEraBasedProtocolParametersUpdate :: AlonzoEraOnwards era -> CostModels -> EraBasedProtocolParametersUpdate era -> EraBasedProtocolParametersUpdate era

carryHashChecks

Arguments

:: MustCheckHash a
Whether to check the hash or not (CheckHash for checking or TrustHash for not checking)

-> Anchor
The anchor data whose hash is to be checked

-> AnchorDataTypeCheck
The type of anchor data to check (for error reporting purpouses)

-> ExceptT GovernanceActionsError IO ()

Check the hash of the anchor data against the hash in the anchor if checkHash is set to CheckHash.