Cardano.CLI.EraBased.Options.Common

Documentation

defaultShelleyBasedEra ∷ EraInEon ShelleyBasedEra

defaultShelleyToBabbageEra ∷ EraInEon ShelleyToBabbageEra

command' ∷ String → String → Parser a → Mod CommandFields a

prefixFlag ∷ Maybe String → String → String

prefixFlag Nothing bar is bar, while prefixFlag (Just "foo") bar is foo-bar. This function is used to optionally prefix some long flags

pNetworkIdDeprecated ∷ Parser NetworkId

pNetworkId ∷ EnvCli → Parser NetworkId

pTarget ∷ CardanoEra era → Parser (Target ChainPoint)

toUnitIntervalOrErr ∷ Rational → UnitInterval

pConsensusModeParams ∷ Parser ConsensusModeParams

defaultByronEpochSlots ∷ Word64

pEpochSlots ∷ Parser EpochSlots

pSocketPath ∷ EnvCli → Parser SocketPath

readerFromParsecParser ∷ Parser a → ReadM a

parseTxIn ∷ Parser TxIn

parseTxId ∷ Parser TxId

parseTxIx ∷ Parser TxIx

decimal ∷ Parser Integer

pStakeIdentifier ∷ Maybe String → Parser StakeIdentifier

pStakeVerifier ∷ Maybe String → Parser StakeVerifier

pStakeAddress ∷ Maybe String → Parser StakeAddress

parseStakeAddress ∷ Parser StakeAddress

pStakeVerificationKeyOrFile ∷ Maybe String → Parser (VerificationKeyOrFile StakeKey)

First argument is the optional prefix

pScriptFor ∷ String → Maybe String → String → Parser ScriptFile

pStakeVerificationKey ∷ Maybe String → Parser (VerificationKey StakeKey)

The first argument is the optional prefix.

readVerificationKey ∷ ∀ keyrole. SerialiseAsBech32 (VerificationKey keyrole) ⇒ AsType keyrole → ReadM (VerificationKey keyrole)

Read a Bech32 or hex-encoded verification key.

pStakeVerificationKeyFile ∷ Maybe String → Parser (VerificationKeyFile In)

The first argument is the optional prefix.

subParser ∷ String → ParserInfo a → Parser a

subInfoParser ∷ String → InfoMod a → [Maybe (Parser a)] → Maybe (Parser a)

pAnyShelleyBasedEra ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pAnyShelleyToBabbageEra ∷ EnvCli → Parser (EraInEon ShelleyToBabbageEra)

pShelleyBasedShelley ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pShelleyBasedAllegra ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pShelleyBasedMary ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pShelleyBasedAlonzo ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pShelleyBasedBabbage ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pShelleyBasedConway ∷ EnvCli → Parser (EraInEon ShelleyBasedEra)

pFileOutDirection ∷ String → String → Parser (File a Out)

pFileInDirection ∷ String → String → Parser (File a In)

parseLovelace ∷ Parser Coin

pStakePoolVerificationKeyOrFile ∷ Maybe String → Parser (VerificationKeyOrFile StakePoolKey)

The first argument is the optional prefix.

pStakePoolVerificationKey ∷ Maybe String → Parser (VerificationKey StakePoolKey)

The first argument is the optional prefix.

pStakePoolVerificationKeyFile ∷ Maybe String → Parser (VerificationKeyFile In)

The first argument is the optional prefix.

pOutputFile ∷ Parser (File content Out)

pMIRPot ∷ Parser MIRPot

pRewardAmt ∷ Parser Coin

pTransferAmt ∷ Parser Coin

pHexHash ∷ SerialiseAsRawBytes (Hash a) ⇒ AsType a → ReadM (Hash a)

pBech32KeyHash ∷ SerialiseAsBech32 (Hash a) ⇒ AsType a → ReadM (Hash a)

pGenesisDelegateVerificationKey ∷ Parser (VerificationKey GenesisDelegateKey)

pColdVerificationKeyOrFile ∷ Maybe String → Parser ColdVerificationKeyOrFile

The first argument is the optional prefix.

pColdVerificationKeyFile ∷ Parser (VerificationKeyFile direction)

pColdSigningKeyFile ∷ Parser (File (SigningKey keyrole) direction)

pVerificationKeyFileOut ∷ Parser (File (VerificationKey keyrole) Out)

pSigningKeyFileOut ∷ Parser (File (SigningKey keyrole) Out)

pOperatorCertIssueCounterFile ∷ Parser (File OpCertCounter direction)

pAddCommitteeColdVerificationKeySource ∷ Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey)

pAddCommitteeColdVerificationKeyHash ∷ Parser (Hash CommitteeColdKey)

pAddCommitteeColdVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile CommitteeColdKey)

pAddCommitteeColdVerificationKey ∷ Parser (VerificationKey CommitteeColdKey)

pAddCommitteeColdVerificationKeyFile ∷ Parser (File (VerificationKey keyrole) In)

pRemoveCommitteeColdVerificationKeySource ∷ Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey)

pScriptHash

Arguments

∷ String
long option name

→ String
help text

→ Parser ScriptHash

pRemoveCommitteeColdVerificationKeyHash ∷ Parser (Hash CommitteeColdKey)

pRemoveCommitteeColdVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile CommitteeColdKey)

pRemoveCommitteeColdVerificationKey ∷ Parser (VerificationKey CommitteeColdKey)

pRemoveCommitteeColdVerificationKeyFile ∷ Parser (File (VerificationKey keyrole) In)

pCommitteeColdVerificationKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile CommitteeColdKey)

pCommitteeColdVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile CommitteeColdKey)

pCommitteeColdVerificationKey ∷ Parser (VerificationKey CommitteeColdKey)

pCommitteeColdVerificationKeyHash ∷ Parser (Hash CommitteeColdKey)

pCommitteeColdVerificationKeyFile ∷ Parser (File (VerificationKey keyrole) In)

pVerificationKeyFileIn ∷ Parser (VerificationKeyFile In)

pAnyVerificationKeyFileIn ∷ String → Parser (VerificationKeyFile In)

pAnyVerificationKeyText ∷ String → Parser AnyVerificationKeyText

pAnyVerificationKeySource ∷ String → Parser AnyVerificationKeySource

pCommitteeHotKey ∷ Parser (VerificationKey CommitteeHotKey)

pCommitteeHotVerificationKey ∷ Parser (VerificationKey CommitteeHotKey)

deserialiseHotCCKeyFromHex ∷ String → Either String (VerificationKey CommitteeHotKey)

pCommitteeHotKeyFile ∷ Parser (VerificationKeyFile In)

pCommitteeHotVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pCommitteeHotKeyHash ∷ Maybe String → Parser (Hash CommitteeHotKey)

The first argument is the optional prefix.

pCommitteeHotKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile CommitteeHotKey)

pCommitteeHotVerificationKeyOrHashOrVerificationFile ∷ Parser (VerificationKeyOrHashOrFile CommitteeHotKey)

pCommitteeHotVerificationKeyOrHashOrVerificationFileOrScriptHash ∷ Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeHotKey)

catCommands ∷ [Parser a] → Maybe (Parser a)

pConstitutionUrl ∷ Parser ConstitutionUrl

pConstitutionHash ∷ Parser (SafeHash StandardCrypto AnchorData)

pUrl ∷ String → String → Parser Url

pGovActionDeposit ∷ Parser Coin

pNewGovActionDeposit ∷ Parser Coin

pStakeVerificationKeyOrHashOrFile ∷ Maybe String → Parser (VerificationKeyOrHashOrFile StakeKey)

First argument is the optional prefix

pStakeVerificationKeyHash ∷ Maybe String → Parser (Hash StakeKey)

First argument is the optional prefix

pStakePoolVerificationKeyOrHashOrFile ∷ Maybe String → Parser (VerificationKeyOrHashOrFile StakePoolKey)

The first argument is the optional prefix.

pCBORInFile ∷ Parser FilePath

pPollQuestion ∷ Parser Text

pPollAnswer ∷ Parser Text

pPollAnswerIndex ∷ Parser Word

pPollFile ∷ Parser (File GovernancePoll In)

pPollTxFile ∷ Parser (TxFile In)

pPollNonce ∷ Parser Word

pScriptWitnessFiles

Arguments

∷ ∀ witctx. WitCtx witctx

→ BalanceTxExecUnits
Use the execution-units flag.

→ String
Script flag prefix

→ Maybe String

→ String

→ Parser (ScriptWitnessFiles witctx)

pExecutionUnits ∷ String → Parser ExecutionUnits

pScriptRedeemerOrFile ∷ String → Parser ScriptDataOrFile

pScriptDatumOrFile ∷ String → WitCtx witctx → Parser (ScriptDatumOrFile witctx)

pScriptDataOrFile ∷ String → String → String → Parser ScriptDataOrFile

pVoteFiles ∷ ShelleyBasedEra era → BalanceTxExecUnits → Parser [(VoteFile In, Maybe (ScriptWitnessFiles WitCtxStake))]

pVoteFile ∷ BalanceTxExecUnits → Parser (VoteFile In, Maybe (ScriptWitnessFiles WitCtxStake))

pProposalFiles ∷ ShelleyBasedEra era → BalanceTxExecUnits → Parser [(ProposalFile In, Maybe (ScriptWitnessFiles WitCtxStake))]

pProposalFile ∷ BalanceTxExecUnits → Parser (ProposalFile In, Maybe (ScriptWitnessFiles WitCtxStake))

pPaymentVerifier ∷ Parser PaymentVerifier

pPaymentVerificationKeyTextOrFile ∷ Parser VerificationKeyTextOrFile

pPaymentVerificationKeyText ∷ Parser Text

pPaymentVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pScript ∷ Parser ScriptFile

pReferenceTxIn ∷ String → String → Parser TxIn

pReadOnlyReferenceTxIn ∷ Parser TxIn

pAddressKeyType ∷ Parser AddressKeyType

pProtocolParamsFile ∷ Parser ProtocolParamsFile

pCalculatePlutusScriptCost ∷ Parser TxBuildOutputOptions

pCertificateFile ∷ BalanceTxExecUnits → Parser (CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))

pPoolMetadataFile ∷ Parser (StakePoolMetadataFile In)

pTxMetadataJsonSchema ∷ Parser TxMetadataJsonSchema

convertTime ∷ String → UTCTime

pMetadataFile ∷ Parser MetadataFile

pWithdrawal ∷ BalanceTxExecUnits → Parser (StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))

pPlutusStakeReferenceScriptWitnessFiles

Arguments

∷ String

→ BalanceTxExecUnits
Use the execution-units flag.

→ Parser (ScriptWitnessFiles WitCtxStake)

pPlutusScriptLanguage ∷ String → Parser AnyScriptLanguage

pUpdateProposalFile ∷ Parser UpdateProposalFile

pRequiredSigner ∷ Parser RequiredSigner

pVrfSigningKeyFile ∷ Parser (SigningKeyFile In)

pWhichLeadershipSchedule ∷ Parser EpochLeadershipSchedule

pWitnessSigningData ∷ Parser WitnessSigningData

pSigningKeyFileIn ∷ Parser (SigningKeyFile In)

pKesPeriod ∷ Parser KESPeriod

pEpochNo ∷ String → Parser EpochNo

pEpochNoUpdateProp ∷ Parser EpochNo

pGenesisFile ∷ String → Parser GenesisFile

pOperationalCertificateFile ∷ Parser (File () direction)

pKeyOutputFormat ∷ Parser KeyOutputFormat

pPoolIdOutputFormat ∷ Parser IdOutputFormat

pQueryOutputFormat ∷ String → Parser QueryOutputFormat

pQueryOutputFormat kind is a parser to specify in which format to view some data (json or text). kind is the kind of data considered.

pTxViewOutputFormat ∷ Parser ViewOutputFormat

pGovernanceActionViewOutputFormat ∷ Parser ViewOutputFormat

pGovernanceVoteViewOutputFormat ∷ Parser ViewOutputFormat

pViewOutputFormat ∷ String → Parser ViewOutputFormat

pViewOutputFormat kind is a parser to specify in which format to view some data (json or yaml). what is the kind of data considered.

pMaybeOutputFile ∷ Parser (Maybe (File content Out))

pVerificationKey ∷ ∀ keyrole. SerialiseAsBech32 (VerificationKey keyrole) ⇒ AsType keyrole → Parser (VerificationKey keyrole)

pVerificationKeyOrFileIn ∷ SerialiseAsBech32 (VerificationKey keyrole) ⇒ AsType keyrole → Parser (VerificationKeyOrFile keyrole)

pExtendedVerificationKeyFileIn ∷ Parser (VerificationKeyFile In)

pGenesisVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pGenesisVerificationKeyHash ∷ Parser (Hash GenesisKey)

pGenesisVerificationKey ∷ Parser (VerificationKey GenesisKey)

pGenesisVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile GenesisKey)

pGenesisVerificationKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile GenesisKey)

pGenesisDelegateVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pGenesisDelegateVerificationKeyHash ∷ Parser (Hash GenesisDelegateKey)

pGenesisDelegateVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile GenesisDelegateKey)

pGenesisDelegateVerificationKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile GenesisDelegateKey)

pKesVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile KesKey)

pKesVerificationKey ∷ Parser (VerificationKey KesKey)

pKesVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pTxSubmitFile ∷ Parser FilePath

pTxIn ∷ BalanceTxExecUnits → Parser (TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))

pTxInCollateral ∷ Parser TxIn

pReturnCollateral ∷ Parser TxOutShelleyBasedEra

pTotalCollateral ∷ Parser Coin

pWitnessOverride ∷ Parser Word

pTxOut ∷ Parser TxOutAnyEra

pTxOutShelleyBased ∷ Parser TxOutShelleyBasedEra

pTxOutDatum ∷ Parser TxOutDatumAnyEra

pRefScriptFp ∷ Parser ReferenceScriptAnyEra

pMintMultiAsset ∷ BalanceTxExecUnits → Parser (Value, [ScriptWitnessFiles WitCtxMint])

pPolicyId ∷ Parser PolicyId

pInvalidBefore ∷ Parser SlotNo

pLegacyInvalidHereafter ∷ Parser SlotNo

pInvalidHereafter ∷ ShelleyBasedEra era → Parser (TxValidityUpperBound era)

pTxFee ∷ Parser Coin

pWitnessFile ∷ Parser WitnessFile

pTxBodyFileIn ∷ Parser (TxBodyFile In)

pTxBodyFileOut ∷ Parser (TxBodyFile Out)

pTxFileIn ∷ Parser (TxFile In)

pTxFileOut ∷ Parser (TxFile Out)

pInputTxOrTxBodyFile ∷ Parser InputTxBodyOrTxFile

pTxInCountDeprecated ∷ Parser TxInCount

pTxOutCountDeprecated ∷ Parser TxOutCount

pTxShelleyWitnessCount ∷ Parser TxShelleyWitnessCount

pTxByronWitnessCount ∷ Parser TxByronWitnessCount

pQueryUTxOFilter ∷ Parser QueryUTxOFilter

pFilterByStakeAddress ∷ Parser StakeAddress

pByronAddress ∷ Parser (Address ByronAddr)

pAddress ∷ Parser Text

pStakePoolVerificationKeyHash ∷ Maybe String → Parser (Hash StakePoolKey)

First argument is the prefix to use

pVrfVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pVrfVerificationKeyHash ∷ Parser (Hash VrfKey)

pVrfVerificationKey ∷ Parser (VerificationKey VrfKey)

pVrfVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile VrfKey)

pVrfVerificationKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile VrfKey)

pRewardAcctVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pRewardAcctVerificationKey ∷ Parser (VerificationKey StakeKey)

pRewardAcctVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile StakeKey)

pPoolOwnerVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pPoolOwnerVerificationKey ∷ Parser (VerificationKey StakeKey)

pPoolOwnerVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile StakeKey)

pPoolPledge ∷ Parser Coin

pPoolCost ∷ Parser Coin

pRational ∷ String → String → Parser Rational

pPoolMargin ∷ Parser Rational

pPoolRelay ∷ Parser StakePoolRelay

pMultiHostName ∷ Parser StakePoolRelay

pSingleHostName ∷ Parser StakePoolRelay

pSingleHostAddress ∷ Parser StakePoolRelay

pIpV4 ∷ Parser IPv4

pIpV6 ∷ Parser IPv6

pPort ∷ Parser PortNumber

pStakePoolMetadataReference ∷ Parser (Maybe StakePoolMetadataReference)

pStakePoolMetadataUrl ∷ Parser Text

pStakePoolMetadataHash ∷ Parser (Hash StakePoolMetadata)

pStakePoolRegistrationParserRequirements ∷ EnvCli → Parser StakePoolRegistrationParserRequirements

pProtocolParametersUpdate ∷ Parser ProtocolParametersUpdate

pCostModels ∷ Parser FilePath

pMinFeePerByteFactor ∷ Parser Coin

pMinFeeConstantFactor ∷ Parser Coin

pMinUTxOValue ∷ Parser Coin

pMinPoolCost ∷ Parser Coin

pMaxBodySize ∷ Parser Word32

pMaxTransactionSize ∷ Parser Word32

pMaxBlockHeaderSize ∷ Parser Word16

pKeyRegistDeposit ∷ Parser Coin

pDrepDeposit ∷ Parser Coin

pPoolDeposit ∷ Parser Coin

pEpochBoundRetirement ∷ Parser EpochInterval

pNumberOfPools ∷ Parser Natural

pPoolInfluence ∷ Parser Rational

pTreasuryExpansion ∷ Parser Rational

pMonetaryExpansion ∷ Parser Rational

pDecentralParam ∷ Parser Rational

pExtraEntropy ∷ Parser (Maybe PraosNonce)

pUTxOCostPerByte ∷ Parser Coin

pExecutionUnitPrices ∷ Parser ExecutionUnitPrices

pMaxTxExecutionUnits ∷ Parser ExecutionUnits

pMaxBlockExecutionUnits ∷ Parser ExecutionUnits

pMaxValueSize ∷ Parser Natural

pCollateralPercent ∷ Parser Natural

pMaxCollateralInputs ∷ Parser Natural

pProtocolVersion ∷ Parser (Natural, Natural)

pPoolVotingThresholds ∷ Parser PoolVotingThresholds

pDRepVotingThresholds ∷ Parser DRepVotingThresholds

pMinCommitteeSize ∷ Parser Natural

pCommitteeTermLength ∷ Parser EpochInterval

pGovActionLifetime ∷ Parser EpochInterval

pDRepDeposit ∷ Parser Coin

pDRepActivity ∷ Parser EpochInterval

parseTxOutShelleyBasedEra ∷ Parser (TxOutDatumAnyEra → ReferenceScriptAnyEra → TxOutShelleyBasedEra)

parseShelleyAddress ∷ Parser (Address ShelleyAddr)

parseTxOutAnyEra ∷ Parser (TxOutDatumAnyEra → ReferenceScriptAnyEra → TxOutAnyEra)

pVoteChoice ∷ Parser Vote

pVoterType ∷ Parser VType

pVotingCredential ∷ Parser (VerificationKeyOrFile StakePoolKey)

pVoteDelegationTarget ∷ Parser VoteDelegationTarget

pAlwaysAbstain ∷ Parser ()

pVoteAnchor ∷ Parser (VoteUrl, SafeHash StandardCrypto AnchorData)

pVoteAnchorDataHash ∷ Parser (SafeHash StandardCrypto AnchorData)

pAlwaysNoConfidence ∷ Parser ()

pDRepHashSource ∷ Parser DRepHashSource

pDRepScriptHash ∷ Parser ScriptHash

pConstitutionScriptHash ∷ Parser ScriptHash

pDRepVerificationKeyOrHashOrFile ∷ Parser (VerificationKeyOrHashOrFile DRepKey)

pDRepVerificationKeyOrHashOrFileOrScriptHash ∷ Parser (VerificationKeyOrHashOrFileOrScriptHash DRepKey)

pAllOrOnlyDRepHashSoure ∷ Parser (AllOrOnly DRepHashSource)

pDRepVerificationKeyHash ∷ Parser (Hash DRepKey)

pDRepVerificationKey ∷ Parser (VerificationKey DRepKey)

pDRepVerificationKeyOrFile ∷ Parser (VerificationKeyOrFile DRepKey)

pDRepVerificationKeyFile ∷ Parser (VerificationKeyFile In)

pAnchorUrl ∷ Parser ProposalUrl

pAnchorDataHash ∷ Parser (SafeHash StandardCrypto AnchorData)

pPreviousGovernanceAction ∷ Parser (Maybe (TxId, Word32))

pGovernanceActionId ∷ Parser (TxId, Word32)

pWord32 ∷ String → String → Parser Word32

pTxId ∷ String → String → Parser TxId

pNetworkIdForTestnetData ∷ EnvCli → Parser NetworkId

pFeatured ∷ Eon eon ⇒ ToCardanoEra peon ⇒ peon era → Parser a → Parser (Maybe (Featured eon era a))

hiddenSubParser ∷ String → ParserInfo a → Parser a