Cardano.CLI.EraBased.Run.Query

Documentation

runQueryCmds ∷ QueryCmds era → ExceptT QueryCmdError IO ()

runQueryConstitutionHashCmd ∷ QueryConstitutionHashCmdArgs → ExceptT QueryCmdError IO ()

runQueryKesPeriodInfoCmd ∷ QueryKesPeriodInfoCmdArgs → ExceptT QueryCmdError IO ()

runQueryLeadershipScheduleCmd ∷ QueryLeadershipScheduleCmdArgs → ExceptT QueryCmdError IO ()

runQueryLedgerStateCmd ∷ QueryLedgerStateCmdArgs → ExceptT QueryCmdError IO ()

runQueryPoolStateCmd ∷ QueryPoolStateCmdArgs → ExceptT QueryCmdError IO ()

Query the current and future parameters for a stake pool, including the retirement date. Any of these may be empty (in which case a null will be displayed).

runQueryProtocolParametersCmd ∷ QueryProtocolParametersCmdArgs → ExceptT QueryCmdError IO ()

runQueryProtocolStateCmd ∷ QueryProtocolStateCmdArgs → ExceptT QueryCmdError IO ()

runQuerySlotNumberCmd ∷ QuerySlotNumberCmdArgs → ExceptT QueryCmdError IO ()

runQueryStakeAddressInfoCmd ∷ QueryStakeAddressInfoCmdArgs → ExceptT QueryCmdError IO ()

Query the current delegations and reward accounts, filtered by a given set of addresses, from a Shelley node via the local state query protocol.

runQueryStakeDistributionCmd ∷ QueryStakeDistributionCmdArgs → ExceptT QueryCmdError IO ()

runQueryStakePoolsCmd ∷ QueryStakePoolsCmdArgs → ExceptT QueryCmdError IO ()

runQueryStakeSnapshotCmd ∷ QueryStakeSnapshotCmdArgs → ExceptT QueryCmdError IO ()

Obtain stake snapshot information for a pool, plus information about the total active stake. This information can be used for leader slot calculation, for example, and has been requested by SPOs. Obtaining the information directly is significantly more time and memory efficient than using a full ledger state dump.

runQueryTipCmd ∷ QueryTipCmdArgs → ExceptT QueryCmdError IO ()

runQueryTxMempoolCmd ∷ QueryTxMempoolCmdArgs → ExceptT QueryCmdError IO ()

Query the local mempool state

runQueryUTxOCmd ∷ QueryUTxOCmdArgs → ExceptT QueryCmdError IO ()

Query the UTxO, filtered by a given set of addresses, from a Shelley node via the local state query protocol.

newtype DelegationsAndRewards

A mapping of Shelley reward accounts to both the stake pool that they delegate to and their reward account balance. TODO: Move to cardano-api

Constructors

DelegationsAndRewards (Map StakeAddress Coin, Map StakeAddress PoolId)

renderQueryCmdError ∷ QueryCmdError → Doc ann

renderOpCertIntervalInformation ∷ FilePath → OpCertIntervalInformation → Doc AnsiStyle

percentage

Arguments

∷ RelativeTime
tolerance. If b - a < tolerance, then 100% is reported. This even if we are tolerance seconds behind, we are still considered fully synced.

→ RelativeTime
nowTime. The time of the most recently synced block.

→ RelativeTime
tipTime. The time of the tip of the block chain to which we need to sync.

→ Text

Calculate the percentage sync rendered as text.