Cardano.CLI.EraBased.Run.StakeAddress

Documentation

runStakeAddressCmds ∷ StakeAddressCmds era → ExceptT StakeAddressCmdError IO ()

runStakeAddressBuildCmd ∷ StakeVerifier → NetworkId → Maybe (File () Out) → ExceptT StakeAddressCmdError IO ()

runStakeAddressKeyGenCmd ∷ KeyOutputFormat → VerificationKeyFile Out → SigningKeyFile Out → ExceptT StakeAddressCmdError IO (VerificationKey StakeKey, SigningKey StakeKey)

runStakeAddressKeyHashCmd ∷ VerificationKeyOrFile StakeKey → Maybe (File () Out) → ExceptT StakeAddressCmdError IO ()

runStakeAddressStakeDelegationCertificateCmd

Arguments

∷ ShelleyBasedEra era

→ StakeIdentifier
Delegator stake verification key, verification key file or script file.

→ VerificationKeyOrHashOrFile StakePoolKey
Delegatee stake pool verification key or verification key file or verification key hash.

→ File () Out

→ ExceptT StakeAddressCmdError IO ()

runStakeAddressDeregistrationCertificateCmd

Arguments

∷ ShelleyBasedEra era

→ StakeIdentifier

→ Maybe Coin
Deposit required in conway era

→ File () Out

→ ExceptT StakeAddressCmdError IO ()

runStakeAddressRegistrationCertificateCmd

Arguments

∷ ShelleyBasedEra era

→ StakeIdentifier

→ Maybe Coin
Deposit required in conway era

→ File () Out

→ ExceptT StakeAddressCmdError IO ()