Cardano.CLI.EraBased.StakeAddress.Run

Documentation

runStakeAddressCmds :: StakeAddressCmds era -> ExceptT StakeAddressCmdError IO ()

runStakeAddressBuildCmd :: StakeVerifier -> NetworkId -> Maybe (File () 'Out) -> ExceptT StakeAddressCmdError IO ()

runStakeAddressKeyGenCmd :: KeyOutputFormat -> VerificationKeyFile 'Out -> SigningKeyFile 'Out -> ExceptT StakeAddressCmdError IO (VerificationKey StakeKey, SigningKey StakeKey)

runStakeAddressKeyHashCmd :: VerificationKeyOrFile StakeKey -> Maybe (File () 'Out) -> ExceptT StakeAddressCmdError IO ()

runStakeAddressStakeDelegationCertificateCmd

Arguments

:: ShelleyBasedEra era

-> StakeIdentifier
Delegator stake verification key, verification key file or script file.

-> VerificationKeyOrHashOrFile StakePoolKey
Delegatee stake pool verification key or verification key file or verification key hash.

-> File () 'Out

-> ExceptT StakeAddressCmdError IO ()

runStakeAddressDeregistrationCertificateCmd

Arguments

:: ShelleyBasedEra era

-> StakeIdentifier

-> Maybe Lovelace
Deposit required in conway era

-> File () 'Out

-> ExceptT StakeAddressCmdError IO ()

runStakeAddressRegistrationCertificateCmd

Arguments

:: ShelleyBasedEra era

-> StakeIdentifier

-> Maybe Lovelace
Deposit required in conway era

-> File () 'Out

-> ExceptT StakeAddressCmdError IO ()

runStakeAddressRegistrationAndDelegationCertificateCmd

Arguments

:: ConwayEraOnwards era

-> StakeIdentifier

-> VerificationKeyOrHashOrFile StakePoolKey
Delegatee stake pool verification key or verification key file or id

-> Lovelace

-> File () 'Out

-> ExceptT StakeAddressCmdError IO ()

runStakeAddressRegistrationStakeAndVoteDelegationCertificateCmd :: ConwayEraOnwards era -> StakeIdentifier -> VerificationKeyOrHashOrFile StakePoolKey -> VoteDelegationTarget -> Lovelace -> File () 'Out -> ExceptT StakeAddressCmdError IO ()