Cardano.CLI.EraBased.Transaction.Command

Documentation

data MustExtendSafeZone

Allow overriding the validity of the era history past the safe zone. The safe zone is a period of time during which we are sure there won't be any era transition (hard fork), and we are confident that the slot duration will not change, thus the conversion from slot numbers to POSIX times using the era history will be correct.

This safe zone is conservative. Even if we are past the safe zone, if there hasn't been any era transition (hard fork) since we obtained it, we can continue safely using the era history.

MustExtendSafeZone essentially disables the safe zone check. This allows the user to use the era history past the safe zone, at the user's discretion.

Constructors

MustExtendSafeZone

DoNotExtendSafeZone

data NodeContextInfoSource era

Either information about the context in which the transaction command is run, or information required to obtain it (information to connect to the node).

Constructors

NodeConnectionInfo !LocalNodeConnectInfo

ProvidedTransactionContextInfo !(TransactionContext era)

data SystemStartOrGenesisFileSource

The system start time or the genesis file from which to get it

Constructors

SystemStartLiteral !SystemStart

SystemStartFromGenesisFile !GenesisFile

data TransactionBuildCmdArgs era

Like TransactionBuildRaw but without the fee, and with a change output.

Constructors

TransactionBuildCmdArgs

Fields

currentEra :: !(Era era)

nodeConnInfo :: !LocalNodeConnectInfo

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

mOverrideWitnesses :: !(Maybe Word)

Override the required number of tx witnesses

txins :: ![(TxIn, Maybe (ScriptRequirements 'TxInItem))]

Transaction inputs with optional spending scripts

readOnlyReferenceInputs :: ![TxIn]

Read only reference inputs

requiredSigners :: ![RequiredSigner]

Required signers

txinsc :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

mTotalCollateral :: !(Maybe Coin)

Total collateral

txouts :: ![TxOutAnyEra]

Normal outputs

changeAddresses :: !TxOutChangeAddress

A change output

mMintedAssets :: !(Maybe (MultiAsset, [ScriptRequirements 'MintItem]))

Multi-Asset minted value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

certificates :: ![(CertificateFile, Maybe (ScriptRequirements 'CertItem))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptRequirements 'WithdrawalItem))]

Withdrawals with potential script witness

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

mUpdateProposalFile :: !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

voteFiles :: ![(VoteFile 'In, Maybe (ScriptRequirements 'VoterItem))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptRequirements 'ProposalItem))]

treasuryDonation :: !(Maybe TxTreasuryDonation)

isCborOutCanonical :: !TxCborFormat

buildOutputOptions :: !TxBuildOutputOptions

data TransactionBuildEstimateCmdArgs era

Like TransactionBuildCmd but does not require explicit access to a running node

Constructors

TransactionBuildEstimateCmdArgs

Fields

currentEra :: !(Era era)

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

shelleyWitnesses :: !Int

Number of shelley witnesses to be added

mByronWitnesses :: !(Maybe Int)

protocolParamsFile :: !ProtocolParamsFile

totalUTxOValue :: !Value

txins :: ![(TxIn, Maybe (ScriptRequirements 'TxInItem))]

Transaction inputs with optional spending scripts

readOnlyReferenceInputs :: ![TxIn]

Read only reference inputs

requiredSigners :: ![RequiredSigner]

Required signers

txinsc :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

txouts :: ![TxOutAnyEra]

Normal outputs

changeAddress :: !TxOutChangeAddress

A change output

mMintedAssets :: !(Maybe (MultiAsset, [ScriptRequirements 'MintItem]))

Multi-Asset value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

certificates :: ![(CertificateFile, Maybe (ScriptRequirements 'CertItem))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptRequirements 'WithdrawalItem))]

Withdrawals with potential script witness

plutusCollateral :: !(Maybe Coin)

Total collateral

totalReferenceScriptSize :: !(Maybe ReferenceScriptSize)

Size of all reference scripts in bytes

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

voteFiles :: ![(VoteFile 'In, Maybe (ScriptRequirements 'VoterItem))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptRequirements 'ProposalItem))]

currentTreasuryValueAndDonation :: !(Maybe (TxCurrentTreasuryValue, TxTreasuryDonation))

isCborOutCanonical :: !TxCborFormat

txBodyOutFile :: !(TxBodyFile 'Out)

data TransactionBuildRawCmdArgs era

Constructors

TransactionBuildRawCmdArgs

Fields

eon :: !(Era era)

mScriptValidity :: !(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

txIns :: ![(TxIn, Maybe (ScriptRequirements 'TxInItem))]

Transaction inputs with optional spending scripts

readOnlyRefIns :: ![TxIn]

Read only reference inputs

txInsCollateral :: ![TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

mReturnCollateral :: !(Maybe TxOutShelleyBasedEra)

Return collateral

mTotalCollateral :: !(Maybe Coin)

Total collateral

requiredSigners :: ![RequiredSigner]

Required signers

txouts :: ![TxOutAnyEra]

mMintedAssets :: !(Maybe (MultiAsset, [ScriptRequirements 'MintItem]))

Multi-Asset minted value with script witness

mValidityLowerBound :: !(Maybe SlotNo)

Transaction validity lower bound

mValidityUpperBound :: !(TxValidityUpperBound era)

Transaction validity upper bound

fee :: !Coin

Transaction fee

certificates :: ![(CertificateFile, Maybe (ScriptRequirements 'CertItem))]

Certificates with potential script witness

withdrawals :: ![(StakeAddress, Coin, Maybe (ScriptRequirements 'WithdrawalItem))]

metadataSchema :: !TxMetadataJsonSchema

scriptFiles :: ![ScriptFile]

Auxiliary scripts

metadataFiles :: ![MetadataFile]

mProtocolParamsFile :: !(Maybe ProtocolParamsFile)

mUpdateProprosalFile :: !(Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

voteFiles :: ![(VoteFile 'In, Maybe (ScriptRequirements 'VoterItem))]

proposalFiles :: ![(ProposalFile 'In, Maybe (ScriptRequirements 'ProposalItem))]

currentTreasuryValueAndDonation :: !(Maybe (TxCurrentTreasuryValue, TxTreasuryDonation))

isCborOutCanonical :: !TxCborFormat

txBodyOutFile :: !(TxBodyFile 'Out)

data TransactionCalculateMinFeeCmdArgs

Constructors

TransactionCalculateMinFeeCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

protocolParamsFile :: !ProtocolParamsFile

txShelleyWitnessCount :: !TxShelleyWitnessCount

txByronWitnessCount :: !TxByronWitnessCount

referenceScriptSize :: !ReferenceScriptSize

The total size in bytes of the transaction reference scripts.

outputFormat :: !(Vary '[FormatJson, FormatText, FormatYaml])

outFile :: !(Maybe (File () 'Out))

data TransactionCalculateMinValueCmdArgs era

Constructors

TransactionCalculateMinValueCmdArgs

Fields

era :: !(Era era)

protocolParamsFile :: !ProtocolParamsFile

txOut :: !TxOutShelleyBasedEra

data TransactionCalculatePlutusScriptCostCmdArgs era

Constructors

TransactionCalculatePlutusScriptCostCmdArgs

Fields

nodeContextInfoSource :: !(NodeContextInfoSource era)

txFileIn :: FilePath

outputFile :: !(Maybe (File () 'Out))

data TransactionCmds era

Constructors

TransactionBuildRawCmd !(TransactionBuildRawCmdArgs era)

TransactionBuildCmd !(TransactionBuildCmdArgs era)

TransactionBuildEstimateCmd !(TransactionBuildEstimateCmdArgs era)

TransactionSignCmd !TransactionSignCmdArgs

TransactionWitnessCmd !TransactionWitnessCmdArgs

TransactionSignWitnessCmd !TransactionSignWitnessCmdArgs

TransactionSubmitCmd !TransactionSubmitCmdArgs

TransactionPolicyIdCmd !TransactionPolicyIdCmdArgs

TransactionCalculateMinFeeCmd !TransactionCalculateMinFeeCmdArgs

TransactionCalculateMinValueCmd !(TransactionCalculateMinValueCmdArgs era)

TransactionCalculatePlutusScriptCostCmd !(TransactionCalculatePlutusScriptCostCmdArgs era)

TransactionHashScriptDataCmd !TransactionHashScriptDataCmdArgs

TransactionTxIdCmd !TransactionTxIdCmdArgs

data TransactionContext era

Transaction context, required to evaluate the execution costs of the plutus scripts in the transaction.

Constructors

TransactionContext

Fields

systemStartSource :: SystemStartOrGenesisFileSource

mustExtendSafeZone :: MustExtendSafeZone

eraHistoryFile :: File EraHistory 'In

utxoFile :: File (Some UTxO) 'In

protocolParamsFile :: ProtocolParamsFile

newtype TransactionHashScriptDataCmdArgs

Constructors

TransactionHashScriptDataCmdArgs

Fields

scriptDataOrFile :: ScriptDataOrFile

newtype TransactionPolicyIdCmdArgs

Constructors

TransactionPolicyIdCmdArgs

Fields

scriptFile :: ScriptFile

data TransactionSignCmdArgs

Constructors

TransactionSignCmdArgs

Fields

txOrTxBodyFile :: !InputTxBodyOrTxFile

witnessSigningData :: ![WitnessSigningData]

mNetworkId :: !(Maybe NetworkId)

isCborOutCanonical :: !TxCborFormat

outTxFile :: !(TxFile 'Out)

data TransactionSignWitnessCmdArgs

Constructors

TransactionSignWitnessCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

witnessFiles :: ![WitnessFile]

isCborOutCanonical :: !TxCborFormat

outFile :: !(File () 'Out)

data TransactionSubmitCmdArgs

Constructors

TransactionSubmitCmdArgs

Fields

nodeConnInfo :: !LocalNodeConnectInfo

txFile :: !FilePath

data TransactionTxIdCmdArgs

Constructors

TransactionTxIdCmdArgs

Fields

inputTxBodyOrTxFile :: InputTxBodyOrTxFile

outputFormat :: !(Vary '[FormatJson, FormatText, FormatYaml])

data TransactionViewCmdArgs

Constructors

TransactionViewCmdArgs

data TransactionWitnessCmdArgs

Constructors

TransactionWitnessCmdArgs

Fields

txBodyFile :: !(TxBodyFile 'In)

witnessSigningData :: !WitnessSigningData

mNetworkId :: !(Maybe NetworkId)

outFile :: !(File () 'Out)

data TxCborFormat

Whether output transaction is in CBOR canonical format according to RFC7049 section 3.9.

https://datatracker.ietf.org/doc/html/rfc7049#section-3.9

https://github.com/cardano-foundation/CIPs/blob/master/CIP-0021/README.md#canonical-cbor-serialization-format

Constructors

TxCborCanonical

TxCborNotCanonical

renderTransactionCmds :: TransactionCmds era -> Text