Cardano.CLI.EraBased.Transaction.HashCheck

Documentation

checkCertificateHashes :: Certificate era -> ExceptT TxCmdError IO ()

Find references to anchor data and check the hashes are valid and they match the linked data.

checkVotingProcedureHashes :: ShelleyBasedEra era -> VotingProcedures era -> ExceptT TxCmdError IO ()

Find references to anchor data in voting procedures and check the hashes are valid and they match the linked data.

checkProposalHashes :: ShelleyBasedEra era -> Proposal era -> ExceptT TxCmdError IO ()

Find references to anchor data in proposals and check the hashes are valid and they match the linked data.