Cardano.CLI.EraBased.Transaction.Internal.HashCheck

Documentation

checkCertificateHashes :: IsEra era => Certificate (LedgerEra era) -> ExceptT TxCmdError IO ()

Find references to anchor data and check the hashes are valid and they match the linked data.

checkVotingProcedureHashes :: IsEra era => VotingProcedures era -> ExceptT TxCmdError IO ()

Find references to anchor data in voting procedures and check the hashes are valid and they match the linked data.

checkProposalHashes :: IsEra era => Proposal era -> ExceptT TxCmdError IO ()

Find references to anchor data in proposals and check the hashes are valid and they match the linked data.