Cardano.CLI.EraIndependent.Hash.Command

Documentation

data HashCmds

Constructors

HashAnchorDataCmd !HashAnchorDataCmdArgs

HashScriptCmd !HashScriptCmdArgs

HashGenesisFile !GenesisFile

data HashGoal hash

Constructors

HashToStdout
The hash is written to stdout

CheckHash !hash
The hash to check against

HashToFile !(File () 'Out)
The output file to which the hash is written

data HashAnchorDataCmdArgs

Constructors

HashAnchorDataCmdArgs

Fields

toHash :: !AnchorDataHashSource

hashGoal :: !(HashGoal (SafeHash AnchorData))

data HashScriptCmdArgs

Constructors

HashScriptCmdArgs

Fields

toHash :: !ScriptFile

mOutFile :: !(Maybe (File () 'Out))

The output file to which the hash is written

data AnchorDataHashSource

Constructors

AnchorDataHashSourceBinaryFile (File ProposalBinary 'In)

AnchorDataHashSourceTextFile (File ProposalText 'In)

AnchorDataHashSourceText Text

AnchorDataHashSourceURL Url

renderHashCmds :: HashCmds -> Text