Cardano.CLI.EraIndependent.Key.Command

Documentation

data KeyCmds

Constructors

KeyVerificationKeyCmd !KeyVerificationKeyCmdArgs

KeyNonExtendedKeyCmd !KeyNonExtendedKeyCmdArgs

KeyGenerateMnemonicCmd !KeyGenerateMnemonicCmdArgs

KeyExtendedSigningKeyFromMnemonicCmd !KeyExtendedSigningKeyFromMnemonicArgs

KeyConvertByronKeyCmd !KeyConvertByronKeyCmdArgs

KeyConvertByronGenesisVKeyCmd !KeyConvertByronGenesisVKeyCmdArgs

KeyConvertITNKeyCmd !KeyConvertITNKeyCmdArgs

KeyConvertITNExtendedKeyCmd !KeyConvertITNExtendedKeyCmdArgs

KeyConvertITNBip32KeyCmd !KeyConvertITNBip32KeyCmdArgs

KeyConvertCardanoAddressKeyCmd !KeyConvertCardanoAddressKeyCmdArgs

data KeyVerificationKeyCmdArgs

Get a verification key from a signing key. This supports all key types

Constructors

KeyVerificationKeyCmdArgs

Fields

skeyFile :: !(SigningKeyFile 'In)

Input filepath of the signing key

vkeyFile :: !(VerificationKeyFile 'Out)

Output filepath of the verification key

data KeyNonExtendedKeyCmdArgs

Get a non-extended verification key from an extended verification key. This supports all extended key types.

Constructors

KeyNonExtendedKeyCmdArgs

Fields

extendedVkeyFileIn :: !(VerificationKeyFile 'In)

Input filepath of the ed25519-bip32 verification key

nonExtendedVkeyFileOut :: !(VerificationKeyFile 'Out)

Output filepath of the verification key

data KeyGenerateMnemonicCmdArgs

Generate a mnemonic phrase that can be used to derive signing keys.

Constructors

KeyGenerateMnemonicCmdArgs

Fields

mnemonicOutputFormat :: !(Maybe (File () 'Out))

Output format for the mnemonic phrase

mnemonicWords :: !MnemonicSize

Number of mnemonic words to generate it must be one of: 12, 15, 18, 21, or 24.

data KeyExtendedSigningKeyFromMnemonicArgs

Get an extended signing key from a mnemonic.

Constructors

KeyExtendedSigningKeyFromMnemonicArgs

Fields

keyOutputFormat :: !(Vary '[FormatBech32, FormatTextEnvelope])

derivedExtendedSigningKeyType :: !ExtendedSigningType

derivationAccountNo :: !Word32

mnemonicSource :: !MnemonicSource

signingKeyFileOut :: !(SigningKeyFile 'Out)

data ExtendedSigningType

Type of the key derived from a mnemonic together with the payment key number in the derivation path for cases where it is applicable.

Constructors

ExtendedSigningPaymentKey !Word32

ExtendedSigningStakeKey !Word32

ExtendedSigningDRepKey

ExtendedSigningCCColdKey

ExtendedSigningCCHotKey

data MnemonicSource

Constructors

MnemonicFromFile !(File () 'In)

MnemonicFromInteractivePrompt

data KeyConvertByronKeyCmdArgs

Convert a Byron payment, genesis or genesis delegate key (signing or verification) to a corresponding Shelley-format key.

Constructors

KeyConvertByronKeyCmdArgs

Fields

mPassword :: !(Maybe Text)

Password for signing key (if applicable)

byronKeyType :: !ByronKeyType

The byron key type of the input file

someKeyFileIn :: !(SomeKeyFile 'In)

Input file containing the byron key

someKeyFileOut :: !(File () 'Out)

The output file to which the Shelley-format key will be written

data KeyConvertByronGenesisVKeyCmdArgs

Constructors

KeyConvertByronGenesisVKeyCmdArgs

Fields

vkey :: !VerificationKeyBase64

Base64 string for the Byron genesis verification key

vkeyFileOut :: !(File () 'Out)

The output file

data KeyConvertITNKeyCmdArgs

Convert an Incentivized Testnet (ITN) non-extended (Ed25519) signing or verification key to a corresponding Shelley stake key

Constructors

KeyConvertITNKeyCmdArgs

Fields

itnKeyFile :: !(SomeKeyFile 'In)

Filepath of the ITN key (signing or verification)

outFile :: !(File () 'Out)

The output file

data KeyConvertITNExtendedKeyCmdArgs

Convert an Incentivized Testnet (ITN) extended (Ed25519Extended) signing key to a corresponding Shelley stake signing key

Constructors

KeyConvertITNExtendedKeyCmdArgs

Fields

itnPrivKeyFile :: !(SomeKeyFile 'In)

Filepath of the ITN signing key

outFile :: !(File () 'Out)

The output file

data KeyConvertITNBip32KeyCmdArgs

Convert an Incentivized Testnet (ITN) BIP32 (Ed25519Bip32) signing key to a corresponding Shelley stake signing key

Constructors

KeyConvertITNBip32KeyCmdArgs

Fields

itnPrivKeyFile :: !(SomeKeyFile 'In)

Filepath of the ITN signing key

outFile :: !(File () 'Out)

The output file

data KeyConvertCardanoAddressKeyCmdArgs

Convert a cardano-address extended signing key to a corresponding Shelley-format key

Constructors

KeyConvertCardanoAddressKeyCmdArgs

Fields

cardanoAddressKeyType :: !CardanoAddressKeyType

Address key type of th signing key input file

skeyFileIn :: !(SigningKeyFile 'In)

Input filepath of the signing key

skeyFileOut :: !(File () 'Out)

The output file

renderKeyCmds :: KeyCmds -> Text