Cardano.CLI.Mary.RenderValue

Documentation

data RenderAdaAssetId

Whether the ADA asset ID should be rendered.

Constructors

RenderAdaAssetId
Render the ADA asset ID.

RenderNoAdaAssetId
Do not render the ADA asset ID.

data RenderIndentation

How a "prettified" representation of a Value should be indented.

Constructors

IndentTab
Indent with a tab character ('\t').

IndentSpaces !Int
Indent with a provided number of spaces.

data RenderPrettyValueOptions

Options which detail how a "prettified" representation of a Value should be rendered.

Constructors

RenderPrettyValueOptions

Fields

rpvoIndentation ∷ !RenderIndentation

rpvoRenderAdaAssetId ∷ !RenderAdaAssetId

newtype RenderValueOptions

Options which detail how a representation of a Value should be rendered.

Constructors

RenderValueOptions RenderAdaAssetId

defaultRenderPrettyValueOptions ∷ RenderPrettyValueOptions

defaultRenderValueOptions ∷ RenderValueOptions

renderPrettyValue ∷ RenderPrettyValueOptions → Value → Text

Render a "prettified" textual representation of a Value.

renderValue ∷ RenderValueOptions → Value → Text

Render a textual representation of a Value.

Note that this textual representation can be parsed by parseValue.