Cardano.CLI.Option.Flag.Type

Documentation

data Flag a

A flag that can be used in the command line interface.

It contains information about how to render the flag, its long name, its content, and its value. The type variable a represents the type of the value that the flag holds.

The reason for this type instead of using 'Parser a' directly is to allow for more complex parsing logic, such as handling default values.

Constructors

Flag

Fields

longName :: String

format :: String

options :: FlagOptions

value :: a

data Defaultness

Constructors

IsDefault

NonDefault

newtype FlagOptions

Options for a flag that control how it is rendered and parsed.

Constructors

FlagOptions

Fields

isDefault :: Defaultness

Whether the flag is a default value.

defaultFlagOptions :: FlagOptions