Cardano.CLI.Render

Documentation

customRenderHelp :: Int -> ParserHelp -> String

Convert a help text to String. When the CLI_HELP_TRACE environment variable is set to '1', the output will be in HTML so that it can be viewed in a browser where developer tools can be used to inspect tracing that aids in describing the structure of the output document.

renderAnyCmdError :: Text -> (a -> Doc ann) -> a -> Doc ann