Cardano.CLI.Run.Mnemonic

Documentation

generateMnemonic

Arguments

:: forall (m :: Type -> Type). MonadIO m

=> MnemonicSize
The number of words in the mnemonic.

-> Maybe (File () 'Out)
The file to write the mnemonic to. If Nothing, write to stdout.

-> ExceptT KeyCmdError m ()

Generate a mnemonic and write it to a file or stdout.

extendedSigningKeyFromMnemonicImpl

Arguments

:: Vary '[FormatBech32, FormatTextEnvelope]
The format in which to write the signing key.

-> ExtendedSigningType
The type of the extended signing key to derive with an optional payment key index.

-> Word32
The account index.

-> MnemonicSource
The source of the mnemonic (either file or stdin).

-> SigningKeyFile 'Out
The file to write the signing key to.

-> ExceptT KeyCmdError IO ()

Derive an extended signing key from a mnemonic and write it to a file.