Cardano.CLI.Shelley.Errors

Documentation

data ShelleyQueryCmdLocalStateQueryError

An error that can occur while querying a node's local state.

Constructors

AcquireFailureError !AcquireFailure

EraMismatchError !EraMismatch
A query from a certain era was applied to a ledger from a different era.

ByronProtocolNotSupportedError
The query does not support the Byron protocol.

ShelleyProtocolEraMismatch
The Shelley protocol only supports the Shelley era.

renderLocalStateQueryError ∷ ShelleyQueryCmdLocalStateQueryError → Text

data ShelleyQueryCmdError

Constructors

ShelleyQueryCmdEnvVarSocketErr !EnvSocketError

ShelleyQueryCmdLocalStateQueryError !ShelleyQueryCmdLocalStateQueryError

ShelleyQueryCmdWriteFileError !(FileError ())

ShelleyQueryCmdHelpersError !HelpersError

ShelleyQueryCmdAcquireFailure !AcquiringFailure

ShelleyQueryCmdEraConsensusModeMismatch !AnyConsensusMode !AnyCardanoEra

ShelleyQueryCmdByronEra

ShelleyQueryCmdPoolIdError (Hash StakePoolKey)

ShelleyQueryCmdEraMismatch !EraMismatch

ShelleyQueryCmdUnsupportedMode !AnyConsensusMode

ShelleyQueryCmdPastHorizon !PastHorizonException

ShelleyQueryCmdSystemStartUnavailable

ShelleyQueryCmdGenesisReadError !ShelleyGenesisCmdError

ShelleyQueryCmdLeaderShipError !LeadershipError

ShelleyQueryCmdTextEnvelopeReadError !(FileError TextEnvelopeError)

ShelleyQueryCmdTextReadError !(FileError InputDecodeError)

ShelleyQueryCmdColdKeyReadFileError !(FileError InputDecodeError)

ShelleyQueryCmdOpCertCounterReadError !(FileError TextEnvelopeError)

ShelleyQueryCmdProtocolStateDecodeFailure !(ByteString, DecoderError)

ShelleyQueryCmdSlotToUtcError Text

ShelleyQueryCmdNodeUnknownStakePool FilePath
Operational certificate of the unknown stake pool.

ShelleyQueryCmdPoolStateDecodeError DecoderError

renderShelleyQueryCmdError ∷ ShelleyQueryCmdError → Text

data ShelleyGenesisCmdError

Constructors

ShelleyGenesisCmdAesonDecodeError !FilePath !Text

ShelleyGenesisCmdGenesisFileReadError !(FileError IOException)

ShelleyGenesisCmdGenesisFileDecodeError !FilePath !Text

ShelleyGenesisCmdGenesisFileError !(FileError ())

ShelleyGenesisCmdFileError !(FileError ())

ShelleyGenesisCmdMismatchedGenesisKeyFiles [Int] [Int] [Int]

ShelleyGenesisCmdFilesNoIndex [FilePath]

ShelleyGenesisCmdFilesDupIndex [FilePath]

ShelleyGenesisCmdTextEnvReadFileError !(FileError TextEnvelopeError)

ShelleyGenesisCmdUnexpectedAddressVerificationKey !VerificationKeyFile !Text !SomeAddressVerificationKey

ShelleyGenesisCmdTooFewPoolsForBulkCreds !Word !Word !Word

ShelleyGenesisCmdAddressCmdError !ShelleyAddressCmdError

ShelleyGenesisCmdNodeCmdError !ShelleyNodeCmdError

ShelleyGenesisCmdPoolCmdError !ShelleyPoolCmdError

ShelleyGenesisCmdStakeAddressCmdError !ShelleyStakeAddressCmdError

ShelleyGenesisCmdCostModelsError !FilePath

ShelleyGenesisCmdByronError !ByronGenesisError

ShelleyGenesisStakePoolRelayFileError !FilePath !IOException

ShelleyGenesisStakePoolRelayJsonDecodeError !FilePath !String

renderShelleyNodeCmdError ∷ ShelleyNodeCmdError → Text

data ShelleyNodeCmdError

Constructors

ShelleyNodeCmdReadFileError !(FileError TextEnvelopeError)

ShelleyNodeCmdReadKeyFileError !(FileError InputDecodeError)

ShelleyNodeCmdWriteFileError !(FileError ())

ShelleyNodeCmdOperationalCertificateIssueError !OperationalCertIssueError

ShelleyNodeCmdVrfSigningKeyCreationError

Fields

FilePath

Target path

FilePath

Temp path

ShelleyNodeCmdQuery !ShelleyQueryCmdError

ShelleyNodeCmdOperationalCertificateKESPeriodInFuture

ShelleyNodeCmdKESKeyEvolutionFailed

data HelpersError

Constructors

CBORPrettyPrintError !DeserialiseFailure

CBORDecodingError !DeserialiseFailure

IOError' !FilePath !IOException

OutputMustNotAlreadyExist FilePath

ReadCBORFileFailure !FilePath !Text

renderHelpersError ∷ HelpersError → Text