Cardano.CLI.Shelley.Key

Description

Shelley CLI option data types and functions for cryptographic keys.

Documentation

data InputFormat a where

Input format/encoding.

Constructors

InputFormatBech32 ∷ SerialiseAsBech32 a ⇒ InputFormat a
Bech32 encoding.

InputFormatHex ∷ SerialiseAsRawBytes a ⇒ InputFormat a
Hex/Base16 encoding.

InputFormatTextEnvelope ∷ HasTextEnvelope a ⇒ InputFormat a
Text envelope format.

data InputDecodeError

Input decoding error.

Constructors

InputTextEnvelopeError !TextEnvelopeError
The provided data seems to be a valid text envelope, but some error occurred in deserialising it.

InputBech32DecodeError !Bech32DecodeError
The provided data is valid Bech32, but some error occurred in deserialising it.

InputInvalidError
The provided data does not represent a valid value of the provided type.

deserialiseInput ∷ ∀ a. AsType a → NonEmpty (InputFormat a) → ByteString → Either InputDecodeError a

Deserialise an input of some type that is formatted in some way.

deserialiseInputAnyOf ∷ ∀ b. [FromSomeType SerialiseAsBech32 b] → [FromSomeType HasTextEnvelope b] → ByteString → Either InputDecodeError b

Deserialise an input of some type that is formatted in some way.

The provided ByteString can either be Bech32-encoded or in the text envelope format.

renderInputDecodeError ∷ InputDecodeError → Text

Render an error message for a InputDecodeError.

readKeyFile ∷ AsType a → NonEmpty (InputFormat a) → FilePath → IO (Either (FileError InputDecodeError) a)

Read a cryptographic key from a file.

The contents of the file can either be Bech32-encoded, hex-encoded, or in the text envelope format.

readKeyFileAnyOf ∷ ∀ b. [FromSomeType SerialiseAsBech32 b] → [FromSomeType HasTextEnvelope b] → FilePath → IO (Either (FileError InputDecodeError) b)

Read a cryptographic key from a file given that it is one of the provided types.

The contents of the file can either be Bech32-encoded or in the text envelope format.

readKeyFileTextEnvelope ∷ HasTextEnvelope a ⇒ AsType a → FilePath → IO (Either (FileError InputDecodeError) a)

Read a cryptographic key from a file.

The contents of the file must be in the text envelope format.

readSigningKeyFile ∷ ∀ keyrole. (HasTextEnvelope (SigningKey keyrole), SerialiseAsBech32 (SigningKey keyrole)) ⇒ AsType keyrole → SigningKeyFile → IO (Either (FileError InputDecodeError) (SigningKey keyrole))

Read a signing key from a file.

The contents of the file can either be Bech32-encoded, hex-encoded, or in the text envelope format.

readSigningKeyFileAnyOf ∷ ∀ b. [FromSomeType SerialiseAsBech32 b] → [FromSomeType HasTextEnvelope b] → SigningKeyFile → IO (Either (FileError InputDecodeError) b)

Read a signing key from a file given that it is one of the provided types of signing key.

The contents of the file can either be Bech32-encoded or in the text envelope format.

data VerificationKeyOrFile keyrole

Either a verification key or path to a verification key file.

Constructors

VerificationKeyValue !(VerificationKey keyrole)
A verification key.

VerificationKeyFilePath !VerificationKeyFile
A path to a verification key file. Note that this file hasn't been validated at all (whether it exists, contains a key of the correct type, etc.)

readVerificationKeyOrFile ∷ (HasTextEnvelope (VerificationKey keyrole), SerialiseAsBech32 (VerificationKey keyrole)) ⇒ AsType keyrole → VerificationKeyOrFile keyrole → IO (Either (FileError InputDecodeError) (VerificationKey keyrole))

Read a verification key or verification key file and return a verification key.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrTextEnvFile ∷ HasTextEnvelope (VerificationKey keyrole) ⇒ AsType keyrole → VerificationKeyOrFile keyrole → IO (Either (FileError InputDecodeError) (VerificationKey keyrole))

Read a verification key or verification key file and return a verification key.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data VerificationKeyTextOrFile

Either an unvalidated text representation of a verification key or a path to a verification key file.

Constructors

VktofVerificationKeyText !Text

VktofVerificationKeyFile !VerificationKeyFile

data VerificationKeyTextOrFileError

An error in deserialising a VerificationKeyTextOrFile to a VerificationKey.

Constructors

VerificationKeyTextError !InputDecodeError

VerificationKeyFileError !(FileError InputDecodeError)

readVerificationKeyTextOrFileAnyOf ∷ ∀ b. [FromSomeType SerialiseAsBech32 b] → [FromSomeType HasTextEnvelope b] → VerificationKeyTextOrFile → IO (Either VerificationKeyTextOrFileError b)

Deserialise a verification key from text or a verification key file given that it is one of the provided types.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

renderVerificationKeyTextOrFileError ∷ VerificationKeyTextOrFileError → Text

Render an error message for a VerificationKeyTextOrFileError.

data VerificationKeyOrHashOrFile keyrole

Verification key, verification key hash, or path to a verification key file.

Constructors

VerificationKeyOrFile !(VerificationKeyOrFile keyrole)
Either a verification key or path to a verification key file.

VerificationKeyHash !(Hash keyrole)
A verification key hash.

readVerificationKeyOrHashOrFile ∷ (Key keyrole, SerialiseAsBech32 (VerificationKey keyrole)) ⇒ AsType keyrole → VerificationKeyOrHashOrFile keyrole → IO (Either (FileError InputDecodeError) (Hash keyrole))

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrHashOrTextEnvFile ∷ Key keyrole ⇒ AsType keyrole → VerificationKeyOrHashOrFile keyrole → IO (Either (FileError InputDecodeError) (Hash keyrole))

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data PaymentVerifier

Constructors

PaymentVerifierKey VerificationKeyTextOrFile

PaymentVerifierScriptFile ScriptFile

data StakeVerifier

Constructors

StakeVerifierKey (VerificationKeyOrFile StakeKey)

StakeVerifierScriptFile ScriptFile