Cardano.CLI.Shelley.Run.Node

Documentation

data ShelleyNodeCmdError

Constructors

ShelleyNodeCmdReadFileError !(FileError TextEnvelopeError)

renderShelleyNodeCmdError ∷ ShelleyNodeCmdError → Text

runNodeCmd ∷ NodeCmd → ExceptT ShelleyNodeCmdError IO ()

runNodeIssueOpCert

Arguments

∷ VerificationKeyOrFile KesKey
This is the hot KES verification key.

→ SigningKeyFile
This is the cold signing key.

→ OpCertCounterFile
Counter that establishes the precedence of the operational certificate.

→ KESPeriod
Start of the validity period for this certificate.

→ OutputFile

→ ExceptT ShelleyNodeCmdError IO ()

runNodeKeyGenCold ∷ VerificationKeyFile → SigningKeyFile → OpCertCounterFile → ExceptT ShelleyNodeCmdError IO ()

runNodeKeyGenKES ∷ VerificationKeyFile → SigningKeyFile → ExceptT ShelleyNodeCmdError IO ()

runNodeKeyGenVRF ∷ VerificationKeyFile → SigningKeyFile → ExceptT ShelleyNodeCmdError IO ()

readColdVerificationKeyOrFile ∷ ColdVerificationKeyOrFile → IO (Either (FileError TextEnvelopeError) (VerificationKey StakePoolKey))

Read a cold verification key or file.

If a filepath is provided, it will be interpreted as a text envelope formatted file.