Cardano.CLI.Shelley.Run.Query

Documentation

newtype DelegationsAndRewards

A mapping of Shelley reward accounts to both the stake pool that they delegate to and their reward account balance. TODO: Move to cardano-api

Constructors

DelegationsAndRewards (Map StakeAddress Lovelace, Map StakeAddress PoolId)

data ShelleyQueryCmdError

data ShelleyQueryCmdLocalStateQueryError

An error that can occur while querying a node's local state.

Constructors

AcquireFailureError !AcquireFailure

EraMismatchError !EraMismatch
A query from a certain era was applied to a ledger from a different era.

ByronProtocolNotSupportedError
The query does not support the Byron protocol.

ShelleyProtocolEraMismatch
The Shelley protocol only supports the Shelley era.

renderOpCertIntervalInformation :: FilePath -> OpCertIntervalInformation -> String

renderShelleyQueryCmdError :: ShelleyQueryCmdError -> Text

renderLocalStateQueryError :: ShelleyQueryCmdLocalStateQueryError -> Text

runQueryCmd :: QueryCmd -> ExceptT ShelleyQueryCmdError IO ()

mergeDelegsAndRewards :: DelegationsAndRewards -> [(StakeAddress, Maybe Lovelace, Maybe PoolId)]

percentage

Arguments

:: RelativeTime
tolerance. If b - a < tolerance, then 100% is reported. This even if we are tolerance seconds behind, we are still considered fully synced.

-> RelativeTime
nowTime. The time of the most recently synced block.

-> RelativeTime
tipTime. The time of the tip of the block chain to which we need to sync.

-> Text

Calculate the percentage sync rendered as text.

executeQuery :: forall result era mode. CardanoEra era -> ConsensusModeParams mode -> LocalNodeConnectInfo mode -> QueryInMode mode (Either EraMismatch result) -> ExceptT ShelleyQueryCmdError IO result