Cardano.CLI.Shelley.Run.Read

Metadata

data MetadataError

Constructors

MetadataErrorFile (FileError ())

MetadataErrorJsonParseError !FilePath !String

MetadataErrorConversionError !FilePath !TxMetadataJsonError

MetadataErrorValidationError !FilePath ![(Word64, TxMetadataRangeError)]

MetadataErrorDecodeError !FilePath !DecoderError

MetadataErrorNotAvailableInEra AnyCardanoEra

renderMetadataError ∷ MetadataError → Text

readFileTxMetadata ∷ TxMetadataJsonSchema → MetadataFile → ExceptT MetadataError IO TxMetadata

readTxMetadata ∷ CardanoEra era → TxMetadataJsonSchema → [MetadataFile] → IO (Either MetadataError (TxMetadataInEra era))

Script

data ScriptWitnessError

Constructors

ScriptWitnessErrorFile (FileError ScriptDecodeError)

ScriptWitnessErrorScriptLanguageNotSupportedInEra AnyScriptLanguage AnyCardanoEra

ScriptWitnessErrorExpectedSimple !FilePath !AnyScriptLanguage

ScriptWitnessErrorExpectedPlutus !FilePath !AnyScriptLanguage

ScriptWitnessErrorReferenceScriptsNotSupportedInEra !AnyCardanoEra

ScriptWitnessErrorScriptData ScriptDataError

renderScriptWitnessError ∷ ScriptWitnessError → Text

readScriptDataOrFile ∷ ScriptDataOrFile → ExceptT ScriptDataError IO HashableScriptData

readScriptWitness ∷ CardanoEra era → ScriptWitnessFiles witctx → ExceptT ScriptWitnessError IO (ScriptWitness witctx era)

readScriptWitnessFiles ∷ CardanoEra era → [(a, Maybe (ScriptWitnessFiles ctx))] → ExceptT ScriptWitnessError IO [(a, Maybe (ScriptWitness ctx era))]

readScriptWitnessFilesThruple ∷ CardanoEra era → [(a, b, Maybe (ScriptWitnessFiles ctx))] → ExceptT ScriptWitnessError IO [(a, b, Maybe (ScriptWitness ctx era))]

data ScriptDecodeError

Constructors

ScriptDecodeTextEnvelopeError TextEnvelopeError

ScriptDecodeSimpleScriptError JsonDecodeError

deserialiseScriptInAnyLang ∷ ByteString → Either ScriptDecodeError ScriptInAnyLang

readFileScriptInAnyLang ∷ FilePath → ExceptT (FileError ScriptDecodeError) IO ScriptInAnyLang

Read a script file. The file can either be in the text envelope format wrapping the binary representation of any of the supported script languages, or alternatively it can be a JSON format file for one of the simple script language versions.

Script data (datums and redeemers)

data ScriptDataError

Constructors

ScriptDataErrorFile (FileError ())

ScriptDataErrorJsonParse !FilePath !String

ScriptDataErrorConversion !FilePath !ScriptDataJsonError

ScriptDataErrorValidation !FilePath !ScriptDataRangeError

ScriptDataErrorMetadataDecode !FilePath !DecoderError

ScriptDataErrorJsonBytes !ScriptDataJsonBytesError

readScriptDatumOrFile ∷ ScriptDatumOrFile witctx → ExceptT ScriptDataError IO (ScriptDatum witctx)

readScriptRedeemerOrFile ∷ ScriptRedeemerOrFile → ExceptT ScriptDataError IO ScriptRedeemer

renderScriptDataError ∷ ScriptDataError → Text

Tx

data CddlError

newtype CddlTx

Constructors

CddlTx

Fields

unCddlTx ∷ InAnyCardanoEra Tx

data IncompleteTx

Constructors

UnwitnessedCliFormattedTxBody (InAnyCardanoEra TxBody)

IncompleteCddlFormattedTx (InAnyCardanoEra Tx)

readFileTx ∷ FileOrPipe → IO (Either CddlError (InAnyCardanoEra Tx))

readFileTxBody ∷ FileOrPipe → IO (Either CddlError IncompleteTx)

readCddlTx ∷ FileOrPipe → IO (Either (FileError TextEnvelopeCddlError) CddlTx)

Tx witnesses

data ReadWitnessSigningDataError

Constructors

ReadWitnessSigningDataSigningKeyDecodeError !(FileError InputDecodeError)

ReadWitnessSigningDataScriptError !(FileError JsonDecodeError)

ReadWitnessSigningDataSigningKeyAndAddressMismatch
A Byron address was specified alongside a non-Byron signing key.

renderReadWitnessSigningDataError ∷ ReadWitnessSigningDataError → Text

Render an error message for a ReadWitnessSigningDataError.

data SomeWitness

Constructors

AByronSigningKey (SigningKey ByronKey) (Maybe (Address ByronAddr))

APaymentSigningKey (SigningKey PaymentKey)

APaymentExtendedSigningKey (SigningKey PaymentExtendedKey)

AStakeSigningKey (SigningKey StakeKey)

AStakeExtendedSigningKey (SigningKey StakeExtendedKey)

AStakePoolSigningKey (SigningKey StakePoolKey)

AGenesisSigningKey (SigningKey GenesisKey)

AGenesisExtendedSigningKey (SigningKey GenesisExtendedKey)

AGenesisDelegateSigningKey (SigningKey GenesisDelegateKey)

AGenesisDelegateExtendedSigningKey (SigningKey GenesisDelegateExtendedKey)

AGenesisUTxOSigningKey (SigningKey GenesisUTxOKey)

data ByronOrShelleyWitness

Some kind of Byron or Shelley witness.

Constructors

AByronWitness !ShelleyBootstrapWitnessSigningKeyData

AShelleyKeyWitness !ShelleyWitnessSigningKey

data ShelleyBootstrapWitnessSigningKeyData

Data required for constructing a Shelley bootstrap witness.

Constructors

ShelleyBootstrapWitnessSigningKeyData

Fields

!(SigningKey ByronKey)

Byron signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

data CddlWitnessError

Constructors

CddlWitnessErrorTextEnv (FileError TextEnvelopeError) (FileError TextEnvelopeCddlError)

CddlWitnessIOError (FileError TextEnvelopeError)

readFileTxKeyWitness ∷ FilePath → IO (Either CddlWitnessError (InAnyCardanoEra KeyWitness))

readWitnessSigningData ∷ WitnessSigningData → IO (Either ReadWitnessSigningDataError SomeWitness)

Required signer

data RequiredSignerError

Constructors

RequiredSignerErrorFile (FileError InputDecodeError)

RequiredSignerErrorByronKey SigningKeyFile

categoriseSomeWitness ∷ SomeWitness → ByronOrShelleyWitness

readRequiredSigner ∷ RequiredSigner → IO (Either RequiredSignerError (Hash PaymentKey))

FileOrPipe

data FileOrPipe

We need a type for handling files that may be actually be things like pipes. Currently the CLI makes no guarantee that a "file" will only be read once. This is a problem for a user who who expects to be able to pass a pipe. To handle this, we have a type for representing either files or pipes where the contents will be saved in memory if what we're reading is a pipe (so it can be re-read later). Unfortunately this means we can't easily stream data from pipes, but at present that's not an issue.

fileOrPipe ∷ FilePath → IO FileOrPipe

fileOrPipePath ∷ FileOrPipe → FilePath

Get the path backing a FileOrPipe. This should primarily be used when generating error messages for a user. A user should not call directly call a function like readFile on the result of this function

fileOrPipeCache ∷ FileOrPipe → IO (Maybe ByteString)

readFileOrPipe ∷ FileOrPipe → IO ByteString

Get the contents of a file or pipe. This function reads the entire contents of the file or pipe, and is blocking.