Cardano.CLI.Shelley.Script

Documentation

data ScriptDecodeError

Constructors

ScriptDecodeTextEnvelopeError TextEnvelopeError

ScriptDecodeSimpleScriptError JsonDecodeError

deserialiseScriptInAnyLang :: ByteString -> Either ScriptDecodeError ScriptInAnyLang

readFileScriptInAnyLang :: FilePath -> ExceptT (FileError ScriptDecodeError) IO ScriptInAnyLang

Read a script file. The file can either be in the text envelope format wrapping the binary representation of any of the supported script languages, or alternatively it can be a JSON format file for one of the simple script language versions.