Cardano.CLI.Type.Common

Documentation

data AllOrOnly a

Constructors

All

Only [a]

data AddressKeyType

Constructors

AddressKeyShelley

AddressKeyShelleyExtended

AddressKeyByron

data AnchorScheme

The different schemes that can be used to fetch anchor data.

Constructors

FileScheme

HttpScheme

HttpsScheme

IpfsScheme

data AnyPlutusScriptVersion where

Constructors

AnyPlutusScriptVersion :: forall lang. IsPlutusScriptLanguage lang => PlutusScriptVersion lang -> AnyPlutusScriptVersion

data AnySLanguage where

Constructors

AnySLanguage :: forall (lang :: Language). PlutusLanguage lang => SLanguage lang -> AnySLanguage

data BalanceTxExecUnits

A flag that differentiates between automatically and manually balancing a tx.

Constructors

AutoBalance

ManualBalance

newtype BlockId

Constructors

BlockId String

data ByronKeyFormat

Constructors

NonLegacyByronKeyFormat

LegacyByronKeyFormat

data ByronKeyType

Constructors

ByronPaymentKey ByronKeyFormat

ByronGenesisKey ByronKeyFormat

ByronDelegateKey ByronKeyFormat

data CardanoAddressKeyType

The type of cardano-address key.

Constructors

CardanoAddressShelleyPaymentKey

CardanoAddressShelleyStakeKey

CardanoAddressIcarusPaymentKey

CardanoAddressByronPaymentKey

CardanoAddressCommitteeColdKey

CardanoAddressCommitteeHotKey

CardanoAddressDRepKey

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile :: FilePath

data ConstitutionHashSource

Constructors

ConstitutionHashSourceFile (File ConstitutionText 'In)

ConstitutionHashSourceText Text

ConstitutionHashSourceHash (SafeHash AnchorData)

newtype ConstitutionText

Constructors

ConstitutionText

Fields

unConstitutionText :: Text

newtype ConstitutionUrl

Constructors

ConstitutionUrl

Fields

unConstitutionUrl :: Url

data CredentialGenerationMode

Constructors

OnDisk
Write credentials to disk

Transient
Don't write them to disk (process them in memory)

newtype CurrentKesPeriod

Constructors

CurrentKesPeriod

Fields

unCurrentKesPeriod :: Word64

data DRepCredentials

Constructors

DRepCredentials

Fields

dRepCredentialGenerationMode :: !CredentialGenerationMode

Whether to write them to disk

numOfDRepCredentials :: !Word

The number of DRep credentials to generate

data EpochLeadershipSchedule

Which leadership schedule we are interested in. TODO: Implement Previous and Next epochs

Constructors

CurrentEpoch

NextEpoch

newtype File content (direction :: FileDirection)

A file path with additional type information to indicate what the file is meant to contain and whether it is to be used for reading or writing.

Constructors

File

Fields

unFile :: FilePath

data FileDirection

Constructors

In
Indicate the file is to be used for reading.

Out
Indicate the file is to be used for writing.

InOut
Indicate the file is to be used for both reading and writing.

data FormatBech32

Constructors

FormatBech32

data FormatCborBin

Constructors

FormatCborBin

data FormatCborHex

Constructors

FormatCborHex

data FormatCip129

Constructors

FormatCip129

data FormatHex

Constructors

FormatHex

data FormatJson

Constructors

FormatJson

data FormatText

Constructors

FormatText

data FormatTextEnvelope

Constructors

FormatTextEnvelope

data FormatYaml

Constructors

FormatYaml

newtype GenesisDir

Constructors

GenesisDir FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile :: FilePath

newtype GenesisKeyFile

Constructors

GenesisKeyFile FilePath

data IncludeStake

Whether to include the stake, as queried by drep-stake-distribution, in the output of drep-state. This is (computationally) expensive, but sometimes convenient.

Constructors

WithStake

NoStake

data InputTxBodyOrTxFile

Constructors

InputTxBodyFile (TxBodyFile 'In)

InputTxFile (TxFile 'In)

data MetadataFile

Constructors

MetadataFileJSON (File () 'In)

MetadataFileCBOR (File () 'In)

data MustCheckHash a

Constructors

CheckHash

TrustHash

data OpCertCounter

type OpCertCounterFile = File OpCertCounter

newtype OpCertEndingKesPeriod

Constructors

OpCertEndingKesPeriod

Fields

unOpCertEndingKesPeriod :: Word64

data OpCertIntervalInformation

Constructors

OpCertWithinInterval OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod SlotsTillKesKeyExpiry

OpCertStartingKesPeriodIsInTheFuture OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertExpired OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertSomeOtherError OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod
Shouldn't be possible

data OpCertNodeAndOnDiskCounterInformation

Constructors

OpCertOnDiskCounterEqualToNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is equal to its corresponding counter in the node state. The on disk operational certificate therefore has a valid counter.

OpCertOnDiskCounterAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by 1. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterTooFarAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is less than the counter in the node state. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterBehindNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The corresponding counter for operational certificate was not found in the node state. This means the relevant stake pool has not minted a block yet. When the stake pool has minted a block the corresponding operational certificate's counter will be present in the node state.

OpCertNoBlocksMintedYet OpCertOnDiskCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by more than 1. The on disk operational certificate is invalid in this case.

newtype OpCertNodeStateCounter

Constructors

OpCertNodeStateCounter

Fields

unOpCertNodeStateCounter :: Word64

newtype OpCertOnDiskCounter

Constructors

OpCertOnDiskCounter

Fields

unOpCertOnDiskCounter :: Word64

newtype OpCertStartingKesPeriod

Constructors

OpCertStartingKesPeriod

Fields

unOpCertStartingKesPeriod :: Word64

data PoolParams

This data structure is used to allow nicely formatted output in the query pool-params command. params are the current pool parameter settings, futureparams are new parameters, retiringEpoch is the epoch that has been set for pool retirement. Any of these may be Nothing.

Constructors

PoolParams

Fields

poolParameters :: Maybe StakePoolState

futurePoolParameters :: Maybe StakePoolState

retiringEpoch :: Maybe EpochNo

mkPoolStates :: PoolState era -> Map (KeyHash 'StakePool) PoolParams

data ParserFileDirection

Constructors

Input

Output

newtype PrivKeyFile

Constructors

PrivKeyFile FilePath

data ProposalBinary

Tag for tracking proposals submitted as Bytestring

type ProposalFile = File ()

data ProposalText

Tag for tracking proposals submitted as Text

newtype ProposalUrl

Constructors

ProposalUrl

Fields

unProposalUrl :: Url

newtype ProtocolParamsFile

Constructors

ProtocolParamsFile FilePath

data ReferenceScriptAnyEra

Constructors

ReferenceScriptAnyEraNone

ReferenceScriptAnyEra FilePath

newtype ReferenceScriptSize

Constructors

ReferenceScriptSize

Fields

unReferenceScriptSize :: Int

data RequiredSigner

Plutus script required signers

Constructors

RequiredSignerSkeyFile (SigningKeyFile 'In)

RequiredSignerHash (Hash PaymentKey)

data ScriptDataOrFile

Constructors

ScriptDataCborFile FilePath
By reference to a CBOR file

ScriptDataJsonFile FilePath
By reference to a JSON file

ScriptDataValue HashableScriptData
By value

type ScriptFile = File ScriptInAnyLang 'In

type ScriptRedeemerOrFile = ScriptDataOrFile

type SigningKeyFile = File (SigningKey ())

newtype SlotsTillKesKeyExpiry

Constructors

SlotsTillKesKeyExpiry

Fields

unSlotsTillKesKeyExpiry :: SlotNo

data SomeKeyFile (direction :: FileDirection)

Either a verification or signing key, used for conversions and other commands that make sense for both.

Constructors

AVerificationKeyFile (VerificationKeyFile direction)

ASigningKeyFile (SigningKeyFile direction)

data StakeDelegators

Constructors

StakeDelegators

Fields

stakeDelegatorsGenerationMode :: !CredentialGenerationMode

Whether to write them to disk

numOfStakeDelegators :: !Word

The number of stake credentials to generate

type StakePoolMetadataFile = File StakePoolMetadata

type SupportedSchemes = [AnchorScheme]

Specifies the schemes that are allowed to fetch anchor data.

data TransferDirection

Determines the direction in which the MIR certificate will transfer ADA.

Constructors

TransferToReserves

TransferToTreasury

type TxBodyFile = File (TxBody ())

data TxBuildOutputOptions

Specify whether to render the script cost as JSON in the cli's build command.

Constructors

OutputScriptCostOnly (File () 'Out)

OutputTxBodyOnly (TxBodyFile 'Out)

newtype TxByronWitnessCount

Constructors

TxByronWitnessCount Int

type TxFile = File (Tx ())

newtype TxSubmissionResult

Type used for serialization when printing the hash of a transaction after having submitted it. Also used for printing JSON output of "transaction txid".

Constructors

TxSubmissionResult

Fields

txhash :: TxId

newtype TxTreasuryDonation

Constructors

TxTreasuryDonation

Fields

unTxTreasuryDonation :: Lovelace

newtype TxInCount

Constructors

TxInCount Int

data TxMempoolQuery

Constructors

TxMempoolQueryTxExists TxId

TxMempoolQueryNextTx

TxMempoolQueryInfo

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value TxOutDatumAnyEra ReferenceScriptAnyEra

data TxOutShelleyBasedEra

Constructors

TxOutShelleyBasedEra !(Address ShelleyAddr) Value TxOutDatumAnyEra ReferenceScriptAnyEra

newtype TxOutChangeAddress

A partially-specified transaction output indented to use as a change output.

It does not specify a value, since this will be worked out automatically.

It does not use any script data hash, since that's generally not used for change outputs.

Constructors

TxOutChangeAddress AddressAny

newtype TxOutCount

Constructors

TxOutCount Int

data TxOutDatumAnyEra

Constructors

TxOutDatumByHashOnly (Hash ScriptData)

TxOutDatumByHashOf ScriptDataOrFile

TxOutDatumByValue ScriptDataOrFile

TxOutInlineDatumByValue ScriptDataOrFile

TxOutDatumByNone

newtype TxShelleyWitnessCount

Constructors

TxShelleyWitnessCount Int

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile :: FilePath

newtype VerificationKeyBase64

A raw verification key given in Base64, and decoded into a ByteString.

Constructors

VerificationKeyBase64 String

type VerificationKeyFile = File (VerificationKey ())

newtype VoteUrl

Constructors

VoteUrl

Fields

unVoteUrl :: Url

newtype VoteText

Constructors

VoteText

Fields

unVoteText :: Text

data VoteHashSource

Constructors

VoteHashSourceFile (File VoteText 'In)

VoteHashSourceText Text

VoteHashSourceHash (SafeHash AnchorData)

newtype WitnessFile

Constructors

WitnessFile FilePath

data WitnessSigningData

Data required to construct a witness.

Constructors

KeyWitnessSigningData

Fields

!(SigningKeyFile 'In)

Path to a file that should contain a signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

type DRepMetadataFile = File DRepMetadata

data DRepMetadataUrl

Tag for differentiating between DRep metadata sources and sources for other types of anchor data

data ResignationMetadataUrl

Tag for differentiating between resignation metadatata sources and sources for other types of anchor data

data PotentiallyCheckedAnchor anchorType anchor

Constructors

PotentiallyCheckedAnchor

Fields

pcaAnchor :: anchor

The anchor data whose hash is to be checked

pcaMustCheck :: MustCheckHash anchorType

Whether to check the hash or not (CheckHash for checking or TrustHash for not checking)