Cardano.CLI.Type.Error.CardanoAddressSigningKeyConversionError

Documentation

data CardanoAddressSigningKeyConversionError

An error that can occur while converting a cardano-address extended signing key.

Constructors

CardanoAddressSigningKeyBech32DecodeError !Bech32DecodeError
There was an error in decoding the string as Bech32.

CardanoAddressSigningKeyDeserialisationError !ByteString
There was an error in converting the cardano-address extended signing key.