Cardano.CLI.Types

Documentation

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile ∷ FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile ∷ FilePath

data OutputFormat

The desired output format.

Constructors

OutputFormatHex

OutputFormatBech32

data QueryFilter

UTxO query filtering options.

Constructors

FilterByAddress !(Set AddressAny)

NoFilter

newtype SigningKeyFile

Constructors

SigningKeyFile

Fields

unSigningKeyFile ∷ FilePath

newtype SocketPath

Constructors

SocketPath

Fields

unSocketPath ∷ FilePath

newtype ScriptFile

Constructors

ScriptFile

Fields

unScriptFile ∷ FilePath

data TransferDirection

Determines the direction in which the MIR certificate will transfer ADA.

Constructors

TransferToReserves

TransferToTreasury

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile ∷ FilePath

newtype VerificationKeyFile

Constructors

VerificationKeyFile

Fields

unVerificationKeyFile ∷ FilePath

data Stakes

This data structure is used to allow nicely formatted output within the query stake-snapshot command.

"markPool", "setPool", "goPool" are the three ledger state stake snapshots (from most recent to least recent) go is the snapshot that is used for the current epoch, set will be used in the next epoch, mark for the epoch after that. "markTotal", "setTotal", "goTotal" record the total active stake for each snapshot.

This information can be used by community tools to calculate upcoming leader schedules.

Constructors

Stakes

Fields

markPool ∷ Integer

setPool ∷ Integer

goPool ∷ Integer

markTotal ∷ Integer

setTotal ∷ Integer

goTotal ∷ Integer

data Params crypto

This data structure is used to allow nicely formatted output in the query pool-params command. params are the current pool parameter settings, futureparams are new parameters, retiringEpoch is the epoch that has been set for pool retirement. Any of these may be Nothing.

Constructors

Params

Fields

poolParameters ∷ Maybe (PoolParams crypto)

futurePoolParameters ∷ Maybe (PoolParams crypto)

retiringEpoch ∷ Maybe EpochNo