Cardano.CLI.Types.Common

Documentation

data AllOrOnly a

Constructors

All

Only [a]

data AddressKeyType

Constructors

AddressKeyShelley

AddressKeyShelleyExtended

AddressKeyByron

data BalanceTxExecUnits

A flag that differentiates between automatically and manually balancing a tx.

Constructors

AutoBalance

ManualBalance

newtype BlockId

Constructors

BlockId String

data ByronKeyFormat

Constructors

NonLegacyByronKeyFormat

LegacyByronKeyFormat

data ByronKeyType

Constructors

ByronPaymentKey ByronKeyFormat

ByronGenesisKey ByronKeyFormat

ByronDelegateKey ByronKeyFormat

data CardanoAddressKeyType

The type of cardano-address key.

Constructors

CardanoAddressShelleyPaymentKey

CardanoAddressShelleyStakeKey

CardanoAddressIcarusPaymentKey

CardanoAddressByronPaymentKey

CardanoAddressCommitteeColdKey

CardanoAddressCommitteeHotKey

CardanoAddressDRepKey

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile ∷ FilePath

data ConstitutionHashSource

Constructors

ConstitutionHashSourceFile (File ConstitutionText In)

ConstitutionHashSourceText Text

ConstitutionHashSourceHash (SafeHash StandardCrypto AnchorData)

newtype ConstitutionText

Constructors

ConstitutionText

Fields

unConstitutionText ∷ Text

newtype ConstitutionUrl

Constructors

ConstitutionUrl

Fields

unConstitutionUrl ∷ Url

data CredentialGenerationMode

Constructors

OnDisk
Write credentials to disk

Transient
Don't write them to disk (process them in memory)

newtype CurrentKesPeriod

Constructors

CurrentKesPeriod

Fields

unCurrentKesPeriod ∷ Word64

data DRepCredentials

Constructors

DRepCredentials

Fields

dRepCredentialGenerationMode ∷ !CredentialGenerationMode

Whether to write them to disk

numOfDRepCredentials ∷ !Word

The number of DRep credentials to generate

data EpochLeadershipSchedule

Which leadership schedule we are interested in. TODO: Implement Previous and Next epochs

Constructors

CurrentEpoch

NextEpoch

newtype File content (direction ∷ FileDirection)

A file path with additional type information to indicate what the file is meant to contain and whether it is to be used for reading or writing.

Constructors

File

Fields

unFile ∷ FilePath

data FileDirection

Constructors

In
Indicate the file is to be used for reading.

Out
Indicate the file is to be used for writing.

InOut
Indicate the file is to be used for both reading and writing.

newtype GenesisDir

Constructors

GenesisDir FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile ∷ FilePath

newtype GenesisKeyFile

Constructors

GenesisKeyFile FilePath

data InputTxBodyOrTxFile

Constructors

InputTxBodyFile (TxBodyFile In)

InputTxFile (TxFile In)

data KeyOutputFormat

Constructors

KeyOutputFormatTextEnvelope

KeyOutputFormatBech32

data MetadataFile

Constructors

MetadataFileJSON (File () In)

MetadataFileCBOR (File () In)

data OpCertCounter

type OpCertCounterFile = File OpCertCounter

newtype OpCertEndingKesPeriod

Constructors

OpCertEndingKesPeriod

Fields

unOpCertEndingKesPeriod ∷ Word64

data OpCertIntervalInformation

Constructors

OpCertWithinInterval OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod SlotsTillKesKeyExpiry

OpCertStartingKesPeriodIsInTheFuture OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertExpired OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertSomeOtherError OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod
Shouldn't be possible

data OpCertNodeAndOnDiskCounterInformation

Constructors

OpCertOnDiskCounterEqualToNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is equal to its corresponding counter in the node state. The on disk operational certificate therefore has a valid counter.

OpCertOnDiskCounterAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by 1. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterTooFarAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is less than the counter in the node state. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterBehindNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The corresponding counter for operational certificate was not found in the node state. This means the relevant stake pool has not minted a block yet. When the stake pool has minted a block the corresponding operational certificate's counter will be present in the node state.

OpCertNoBlocksMintedYet OpCertOnDiskCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by more than 1. The on disk operational certificate is invalid in this case.

newtype OpCertNodeStateCounter

Constructors

OpCertNodeStateCounter

Fields

unOpCertNodeStateCounter ∷ Word64

newtype OpCertOnDiskCounter

Constructors

OpCertOnDiskCounter

Fields

unOpCertOnDiskCounter ∷ Word64

newtype OpCertStartingKesPeriod

Constructors

OpCertStartingKesPeriod

Fields

unOpCertStartingKesPeriod ∷ Word64

data Params crypto

This data structure is used to allow nicely formatted output in the query pool-params command. params are the current pool parameter settings, futureparams are new parameters, retiringEpoch is the epoch that has been set for pool retirement. Any of these may be Nothing.

Constructors

Params

Fields

poolParameters ∷ Maybe (PoolParams crypto)

futurePoolParameters ∷ Maybe (PoolParams crypto)

retiringEpoch ∷ Maybe EpochNo

data ParserFileDirection

Constructors

Input

Output

data IdOutputFormat

Some entities such as stake pools and dreps have a notion of an ID and that id can be encoded as either a bech32 or hex string. This type is used to specify which encoding to use.

Constructors

IdOutputFormatHex

IdOutputFormatBech32

newtype PrivKeyFile

Constructors

PrivKeyFile FilePath

data ProposalBinary

Tag for tracking proposals submitted as Bytestring

type ProposalFile = File ()

data ProposalText

Tag for tracking proposals submitted as Text

newtype ProposalUrl

Constructors

ProposalUrl

Fields

unProposalUrl ∷ Url

newtype ProtocolParamsFile

Constructors

ProtocolParamsFile FilePath

data QueryOutputFormat

Constructors

QueryOutputFormatJson

QueryOutputFormatText

data ReferenceScriptAnyEra

Constructors

ReferenceScriptAnyEraNone

ReferenceScriptAnyEra FilePath

newtype ReferenceScriptSize

Constructors

ReferenceScriptSize

Fields

unReferenceScriptSize ∷ Int

data RequiredSigner

Plutus script required signers

Constructors

RequiredSignerSkeyFile (SigningKeyFile In)

RequiredSignerHash (Hash PaymentKey)

data ScriptDataOrFile

Constructors

ScriptDataCborFile FilePath
By reference to a CBOR file

ScriptDataJsonFile FilePath
By reference to a JSON file

ScriptDataValue HashableScriptData
By value

data ScriptDatumOrFile witctx where

Constructors

ScriptDatumOrFileForTxIn ∷ ScriptDataOrFile → ScriptDatumOrFile WitCtxTxIn

InlineDatumPresentAtTxIn ∷ ScriptDatumOrFile WitCtxTxIn

NoScriptDatumOrFileForMint ∷ ScriptDatumOrFile WitCtxMint

NoScriptDatumOrFileForStake ∷ ScriptDatumOrFile WitCtxStake

type ScriptFile = File ScriptInAnyLang In

type ScriptRedeemerOrFile = ScriptDataOrFile

data ScriptWitnessFiles witctx where

This type is like ScriptWitness, but the file paths from which to load the script witness data representation.

It is era-independent, but witness context-dependent.

TODO: Potentially update to WitnessFiles so we can get rid of Maybe (ScriptWitnessFiles). This will be clearer in conveying that we either expect a script witness or a key witness is provided at the signing stage.

Constructors

SimpleScriptWitnessFile ∷ ScriptFile → ScriptWitnessFiles witctx

PlutusScriptWitnessFiles ∷ ScriptFile → ScriptDatumOrFile witctx → ScriptRedeemerOrFile → ExecutionUnits → ScriptWitnessFiles witctx

PlutusReferenceScriptWitnessFiles

Fields

∷ TxIn

→ AnyScriptLanguage

→ ScriptDatumOrFile witctx

→ ScriptRedeemerOrFile

→ ExecutionUnits

→ Maybe PolicyId

For minting reference scripts

→ ScriptWitnessFiles witctx

SimpleReferenceScriptWitnessFiles

Fields

∷ TxIn

→ AnyScriptLanguage

→ Maybe PolicyId

For minting reference scripts

→ ScriptWitnessFiles witctx

type SigningKeyFile = File (SigningKey ())

newtype SlotsTillKesKeyExpiry

Constructors

SlotsTillKesKeyExpiry

Fields

unSlotsTillKesKeyExpiry ∷ SlotNo

data SomeKeyFile direction

Either a verification or signing key, used for conversions and other commands that make sense for both.

Constructors

AVerificationKeyFile (VerificationKeyFile direction)

ASigningKeyFile (SigningKeyFile direction)

data StakeDelegators

Constructors

StakeDelegators

Fields

stakeDelegatorsGenerationMode ∷ !CredentialGenerationMode

Whether to write them to disk

numOfStakeDelegators ∷ !Word

The number of stake credentials to generate

type StakePoolMetadataFile = File StakePoolMetadata

data TransferDirection

Determines the direction in which the MIR certificate will transfer ADA.

Constructors

TransferToReserves

TransferToTreasury

type TxBodyFile = File (TxBody ())

data TxBuildOutputOptions

Specify whether to render the script cost as JSON in the cli's build command.

Constructors

OutputScriptCostOnly (File () Out)

OutputTxBodyOnly (TxBodyFile Out)

newtype TxByronWitnessCount

Constructors

TxByronWitnessCount Int

type TxFile = File (Tx ())

newtype TxTreasuryDonation

Constructors

TxTreasuryDonation

Fields

unTxTreasuryDonation ∷ Coin

newtype TxInCount

Constructors

TxInCount Int

data TxMempoolQuery

Constructors

TxMempoolQueryTxExists TxId

TxMempoolQueryNextTx

TxMempoolQueryInfo

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value TxOutDatumAnyEra ReferenceScriptAnyEra

data TxOutShelleyBasedEra

Constructors

TxOutShelleyBasedEra !(Address ShelleyAddr) Value TxOutDatumAnyEra ReferenceScriptAnyEra

newtype TxOutChangeAddress

A partially-specified transaction output indented to use as a change output.

It does not specify a value, since this will be worked out automatically.

It does not use any script data hash, since that's generally not used for change outputs.

Constructors

TxOutChangeAddress AddressAny

newtype TxOutCount

Constructors

TxOutCount Int

data TxOutDatumAnyEra

Constructors

TxOutDatumByHashOnly (Hash ScriptData)

TxOutDatumByHashOf ScriptDataOrFile

TxOutDatumByValue ScriptDataOrFile

TxOutInlineDatumByValue ScriptDataOrFile

TxOutDatumByNone

newtype TxShelleyWitnessCount

Constructors

TxShelleyWitnessCount Int

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile ∷ FilePath

newtype VerificationKeyBase64

A raw verification key given in Base64, and decoded into a ByteString.

Constructors

VerificationKeyBase64 String

type VerificationKeyFile = File (VerificationKey ())

data ViewOutputFormat

Constructors

ViewOutputFormatJson

ViewOutputFormatYaml

newtype VoteUrl

Constructors

VoteUrl

Fields

unVoteUrl ∷ Url

newtype VoteText

Constructors

VoteText

Fields

unVoteText ∷ Text

data VoteHashSource

Constructors

VoteHashSourceFile (File VoteText In)

VoteHashSourceText Text

VoteHashSourceHash (SafeHash StandardCrypto AnchorData)

newtype WitnessFile

Constructors

WitnessFile FilePath

data WitnessSigningData

Data required to construct a witness.

Constructors

KeyWitnessSigningData

Fields

!(SigningKeyFile In)

Path to a file that should contain a signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

type DRepMetadataFile = File DRepMetadata