Cardano.CLI.Types.Key

Description

Shelley CLI option data types and functions for cryptographic keys.

Documentation

data VerificationKeyOrFile keyrole

Either a verification key or path to a verification key file.

Constructors

VerificationKeyValue !(VerificationKey keyrole)
A verification key.

VerificationKeyFilePath !(VerificationKeyFile 'In)
A path to a verification key file. Note that this file hasn't been validated at all (whether it exists, contains a key of the correct type, etc.)

readVerificationKeyOrFile :: forall t (m :: Type -> Type) keyrole. (MonadIOTransError (FileError InputDecodeError) t m, HasTextEnvelope (VerificationKey keyrole), SerialiseAsBech32 (VerificationKey keyrole)) => AsType keyrole -> VerificationKeyOrFile keyrole -> t m (VerificationKey keyrole)

Read a verification key or verification key file and return a verification key.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrTextEnvFile :: forall t (m :: Type -> Type) keyrole. (MonadIOTransError (FileError InputDecodeError) t m, HasTextEnvelope (VerificationKey keyrole)) => AsType keyrole -> VerificationKeyOrFile keyrole -> t m (VerificationKey keyrole)

Read a verification key or verification key file and return a verification key.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data VerificationKeyTextOrFile

Either an unvalidated text representation of a verification key or a path to a verification key file.

Constructors

VktofVerificationKeyText !Text

VktofVerificationKeyFile !(VerificationKeyFile 'In)

data VerificationKeyTextOrFileError

An error in deserialising a VerificationKeyTextOrFile to a VerificationKey.

Constructors

VerificationKeyTextError !InputDecodeError

VerificationKeyFileError !(FileError InputDecodeError)

readVerificationKeyTextOrFileAnyOf :: VerificationKeyTextOrFile -> IO (Either VerificationKeyTextOrFileError SomeAddressVerificationKey)

Deserialise a verification key from text or a verification key file. If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

renderVerificationKeyTextOrFileError :: VerificationKeyTextOrFileError -> Doc ann

Render an error message for a VerificationKeyTextOrFileError.

data VerificationKeyOrHashOrFile keyrole

Verification key, verification key hash, or path to a verification key file.

Constructors

VerificationKeyOrFile !(VerificationKeyOrFile keyrole)
Either a verification key or path to a verification key file.

VerificationKeyHash !(Hash keyrole)
A verification key hash.

readVerificationKeyOrHashOrFile :: forall t (m :: Type -> Type) keyrole. (MonadIOTransError (FileError InputDecodeError) t m, Key keyrole, SerialiseAsBech32 (VerificationKey keyrole)) => AsType keyrole -> VerificationKeyOrHashOrFile keyrole -> t m (Hash keyrole)

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrHashOrTextEnvFile :: forall t (m :: Type -> Type) keyrole. (MonadIOTransError (FileError InputDecodeError) t m, Key keyrole) => AsType keyrole -> VerificationKeyOrHashOrFile keyrole -> t m (Hash keyrole)

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data VerificationKeyOrHashOrFileOrScript keyrole

Constructors

VkhfsKeyHashFile !(VerificationKeyOrHashOrFile keyrole)

VkhfsScript !(File ScriptInAnyLang 'In)

data VerificationKeyOrHashOrFileOrScriptHash keyrole

Constructors

VkhfshKeyHashFile !(VerificationKeyOrHashOrFile keyrole)

VkhfshScriptHash !ScriptHash

data VerificationKeySource keyrole

Constructors

VksKeyHashFile !(VerificationKeyOrHashOrFile keyrole)

VksScript !(File ScriptInAnyLang 'In)

VksScriptHash !ScriptHash

readVerificationKeyOrHashOrFileOrScriptHash :: forall t (m :: Type -> Type) keyrole (kr :: KeyRole). (MonadIOTransError (FileError InputDecodeError) t m, Key keyrole) => AsType keyrole -> (Hash keyrole -> KeyHash kr StandardCrypto) -> VerificationKeyOrHashOrFileOrScriptHash keyrole -> t m (Credential kr StandardCrypto)

data PaymentVerifier

Constructors

PaymentVerifierKey VerificationKeyTextOrFile

PaymentVerifierScriptFile ScriptFile

data StakeIdentifier

Constructors

StakeIdentifierVerifier StakeVerifier

StakeIdentifierAddress StakeAddress

data StakeVerifier

Constructors

StakeVerifierKey (VerificationKeyOrHashOrFile StakeKey)

StakeVerifierScriptFile ScriptFile

generateKeyPair :: (MonadIO m, Key keyrole, HasTypeProxy keyrole) => AsType keyrole -> m (VerificationKey keyrole, SigningKey keyrole)

data StakePoolRegistrationParserRequirements

Constructors

StakePoolRegistrationParserRequirements

Fields

sprStakePoolKey :: VerificationKeyOrFile StakePoolKey

Stake pool verification key.

sprVrfKey :: VerificationKeyOrFile VrfKey

VRF Verification key.

sprPoolPledge :: Lovelace

Pool pledge.

sprPoolCost :: Lovelace

Pool cost.

sprPoolMargin :: Rational

Pool margin.

sprRewardAccountKey :: VerificationKeyOrFile StakeKey

Reward account verification staking key.

spoPoolOwnerKeys :: [VerificationKeyOrFile StakeKey]

Pool owner verification staking key(s).

sprRelays :: [StakePoolRelay]

Stake pool relays.

sprMetadata :: Maybe (PotentiallyCheckedAnchor StakePoolMetadataReference StakePoolMetadataReference)

Stake pool metadata.

sprNetworkId :: NetworkId

data AnyDelegationTarget where

A resource that identifies the delegation target. We can delegate our stake for two reasons: 1. To gain rewards. This is limited to choosing a stake pool 2. To delegate voting power. We can delegate this to a DRep, always abstain our vote or vote no confidence

Constructors

ShelleyToBabbageDelegTarget

Fields

:: forall era. ShelleyToBabbageEra era

-> VerificationKeyOrHashOrFile StakePoolKey

Stake pool target

-> AnyDelegationTarget

ConwayOnwardDelegTarget :: forall era. ConwayEraOnwards era -> StakeTarget era -> AnyDelegationTarget

data StakeTarget era where

Constructors

TargetStakePool :: forall era. ConwayEraOnwards era -> VerificationKeyOrHashOrFile StakePoolKey -> StakeTarget era

TargetVotingDrep :: forall era. ConwayEraOnwards era -> VerificationKeyOrHashOrFile DRepKey -> StakeTarget era

TargetVotingDrepAndStakePool :: forall era. ConwayEraOnwards era -> VerificationKeyOrHashOrFile DRepKey -> VerificationKeyOrHashOrFile StakePoolKey -> StakeTarget era

TargetAlwaysAbstain :: forall era. ConwayEraOnwards era -> StakeTarget era

TargetAlwaysNoConfidence :: forall era. ConwayEraOnwards era -> StakeTarget era

TargetVotingDRepScriptHash :: forall era. ConwayEraOnwards era -> ScriptHash -> StakeTarget era

data ColdVerificationKeyOrFile

Either a stake pool verification key, genesis delegate verification key, or a path to a cold verification key file.

Note that a "cold verification key" refers to either a stake pool or genesis delegate verification key.

TODO: A genesis delegate extended key should also be valid here.

Constructors

ColdStakePoolVerificationKey !(VerificationKey StakePoolKey)

ColdGenesisDelegateVerificationKey !(VerificationKey GenesisDelegateKey)

ColdVerificationKeyFile !(VerificationKeyFile 'In)

data DRepHashSource

Constructors

DRepHashSourceScript ScriptHash

DRepHashSourceVerificationKey (VerificationKeyOrHashOrFile DRepKey)

readDRepCredential :: forall t (m :: Type -> Type). MonadIOTransError (FileError InputDecodeError) t m => DRepHashSource -> t m (Credential 'DRepRole StandardCrypto)

newtype SPOHashSource

Constructors

SPOHashSourceVerificationKey (VerificationKeyOrHashOrFile StakePoolKey)

readSPOCredential :: forall t (m :: Type -> Type). MonadIOTransError (FileError InputDecodeError) t m => SPOHashSource -> t m (KeyHash 'StakePool StandardCrypto)

data SomeSigningKey

Constructors

AByronSigningKey (SigningKey ByronKey)

APaymentSigningKey (SigningKey PaymentKey)

APaymentExtendedSigningKey (SigningKey PaymentExtendedKey)

AStakeSigningKey (SigningKey StakeKey)

AStakeExtendedSigningKey (SigningKey StakeExtendedKey)

AStakePoolSigningKey (SigningKey StakePoolKey)

AGenesisSigningKey (SigningKey GenesisKey)

AGenesisExtendedSigningKey (SigningKey GenesisExtendedKey)

AGenesisDelegateSigningKey (SigningKey GenesisDelegateKey)

AGenesisDelegateExtendedSigningKey (SigningKey GenesisDelegateExtendedKey)

AGenesisUTxOSigningKey (SigningKey GenesisUTxOKey)

ADRepSigningKey (SigningKey DRepKey)

ADRepExtendedSigningKey (SigningKey DRepExtendedKey)

ACommitteeColdSigningKey (SigningKey CommitteeColdKey)

ACommitteeColdExtendedSigningKey (SigningKey CommitteeColdExtendedKey)

ACommitteeHotSigningKey (SigningKey CommitteeHotKey)

ACommitteeHotExtendedSigningKey (SigningKey CommitteeHotExtendedKey)

AVrfSigningKey (SigningKey VrfKey)

AKesSigningKey (SigningKey KesKey)

withSomeSigningKey :: SomeSigningKey -> (forall keyrole. (Key keyrole, HasTypeProxy keyrole) => SigningKey keyrole -> a) -> a

readSigningKeyFile :: SigningKeyFile 'In -> ExceptT (FileError InputDecodeError) IO SomeSigningKey