Cardano.CLI.Types.Key.VerificationKey

Documentation

data AnyVerificationKeySource

The source from which a verification key of an unspecified key role can be derived.

Constructors

AnyVerificationKeySourceOfText !AnyVerificationKeyText

AnyVerificationKeySourceOfFile !(File (VerificationKey ()) In)

newtype AnyVerificationKeyText

A bech32 text encoded verification key of an unspecified key role.

Constructors

AnyVerificationKeyText

Fields

unAnyVerificationKeyText ∷ Text