Cardano.CLI.Types.MonadWarning

Description

This module defines the MonadWarning type class, which provides a common interface for monads that support reporting warning messages without aborting the computation (unlike with exceptions, Either, or MonadFail, which either fail or return a value).

It also includes two functions that instantiate it into either a MonadIO (runWarningIO) or a StateT monad with a '[String]' as state (runWarningStateT) respectively.

In the case of MonadIO, warnings are printed to stderr. In the case of StateT, with a '[String]' state, warnings are added to the list in the state.

By using the MonadWarning type class, users can write code that remains agnostic to the specific monad in which it operates, and to easily change it at a later stage if necessary.

Example usage:

computeWithWarning :: (MonadWarning m) => Int -> m Int
computeWithWarning x = do
  when (x < 0) $ reportIssue "Input value is negative!"
  return (x * 2)

-- Using IO monad to perform computation and report warnings.
main :: IO ()
main = do
  result <- runWarningIO $ computeWithWarning (-4)
  putStrLn $ "Result: " ++ show result

Documentation

class Monad m ⇒ MonadWarning m where

Type class for monads that support reporting warnings.

Methods

reportIssue

Arguments

∷ String
The warning message to report.

→ m ()
The action that reports the warning.

Report a warning issue.

eitherToWarning ∷ MonadWarning m ⇒ a → Either String a → m a

Convert an Either into a MonadWarning. If Either is Left it returns the default value (first parameter) and reports the value as an error. -- If Either is Right it just returns the value.

runWarningIO ∷ WarningIO m a → m a

runWarningStateT ∷ WarningStateT m a → StateT [String] m a