Cardano.CLI.Types.Output

Description

Types that are used when writing to standard output or to files. These types (and their encodings) are typically consumed by users of cardano-cli.

Documentation

data PlutusScriptCostError

data QueryKesPeriodInfoOutput

Constructors

QueryKesPeriodInfoOutput

Fields

qKesOpCertIntervalInformation ∷ OpCertIntervalInformation

qKesInfoKesKeyExpiry ∷ Maybe UTCTime

Date of KES key expiry.

qKesInfoNodeStateOperationalCertNo ∷ Maybe OpCertNodeStateCounter

The latest operational certificate number in the node's state i.e how many times a new KES key has been generated.

qKesInfoOnDiskOperationalCertNo ∷ OpCertOnDiskCounter

The on disk operational certificate number.

qKesInfoMaxKesKeyEvolutions ∷ Word64

The maximum number of KES key evolutions permitted per KES period.

qKesInfoSlotsPerKesPeriod ∷ Word64

data QueryTipLocalState mode

Constructors

QueryTipLocalState

Fields

era ∷ AnyCardanoEra

eraHistory ∷ EraHistory

mSystemStart ∷ Maybe SystemStart

mChainTip ∷ Maybe ChainTip

data QueryTipLocalStateOutput

Constructors

QueryTipLocalStateOutput

Fields

localStateChainTip ∷ ChainTip

mEra ∷ Maybe AnyCardanoEra

mEpoch ∷ Maybe EpochNo

mSlotInEpoch ∷ Maybe Word64

mSlotsToEpochEnd ∷ Maybe Word64

mSyncProgress ∷ Maybe Text

data ScriptCostOutput

Constructors

ScriptCostOutput

Fields

scScriptHash ∷ ScriptHash

scExecutionUnits ∷ ExecutionUnits

scAda ∷ Coin

createOpCertIntervalInfo ∷ CurrentKesPeriod → OpCertStartingKesPeriod → OpCertEndingKesPeriod → Maybe SlotsTillKesKeyExpiry → OpCertIntervalInformation

renderScriptCosts

Arguments

∷ UTxO era

→ Prices

→ [(ScriptWitnessIndex, AnyScriptWitness era)]
Initial mapping of script witness index to actual script. We need this in order to know which script corresponds to the calculated execution units.

→ Map ScriptWitnessIndex (Either ScriptExecutionError ExecutionUnits)
Post execution cost calculation mapping of script witness index to execution units.

→ Either PlutusScriptCostError [ScriptCostOutput]