Cardano.CLI.Types.TxFeature

Documentation

data TxFeature

An enumeration of era-dependent features where we have to check that it is permissible to use this feature in this era.

Constructors

TxFeatureShelleyAddresses

TxFeatureExplicitFees

TxFeatureImplicitFees

TxFeatureValidityLowerBound

TxFeatureValidityUpperBound

TxFeatureValidityNoUpperBound

TxFeatureTxMetadata

TxFeatureAuxScripts

TxFeatureWithdrawals

TxFeatureCertificates

TxFeatureMintValue

TxFeatureMultiAssetOutputs

TxFeatureScriptWitnesses

TxFeatureShelleyKeys

TxFeatureCollateral

TxFeatureProtocolParameters

TxFeatureTxOutDatum

TxFeatureScriptValidity

TxFeatureExtraKeyWits

TxFeatureInlineDatums

TxFeatureTotalCollateral

TxFeatureReferenceInputs

TxFeatureReturnCollateral

renderFeature :: TxFeature -> Text