Test.Cardano.CLI.Util

Documentation

assertDirectoryMissing :: (MonadTest m, MonadIO m, HasCallStack) => FilePath -> m ()

Asserts that the given directory is missing.

checkTxCddlFormat

Arguments

:: (MonadTest m, MonadIO m, HasCallStack)

=> FilePath
Reference/golden file

-> FilePath
Newly created file

-> m ()

checkTextEnvelopeFormat :: (MonadTest m, MonadIO m, HasCallStack) => TextEnvelopeType -> FilePath -> FilePath -> m ()

Checks that the tvType and tvDescription are equivalent between two files.

equivalence :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m ()

Check for equivalence between two types and perform a file cleanup on failure.

execCardanoCLI

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m String
Captured stdout

Execute cardano-cli via the command line.

Waits for the process to finish and returns the stdout.

execDetailCardanoCLI

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m (ExitCode, String, String)
Captured stdout

Execute cardano-cli via the command line, expecting it to fail.

Waits for the process to finish and returns the exit code, stdout and stderr.

tryExecCardanoCLI

Arguments

:: [String]
Arguments to the CLI command

-> PropertyT IO (Either Failure String)
Captured stdout, or error in case of failures

propertyOnce :: PropertyT IO () -> Property

withSnd :: (a -> b) -> a -> (a, b)

Return the supply value with the result of the supplied function as a tuple

noteInputFile :: (MonadTest m, HasCallStack) => FilePath -> m FilePath

Return the input file path after annotating it relative to the project root directory

noteTempFile :: (MonadTest m, HasCallStack) => FilePath -> FilePath -> m FilePath

Return the test file path after annotating it relative to the project root directory

redactJsonField :: (MonadTest m, MonadIO m, HasCallStack) => Text -> Text -> FilePath -> FilePath -> m ()

bracketSem

Arguments

:: MonadBaseControl IO m

=> FileSem
a file semaphore

-> (FilePath -> m c)
an action, a file path will be extracted from the semaphore

-> m c

Run action acquiring a semaphore, and releasing afterwards. Guards against concurrent access to a block of code.

data FileSem

A file semaphore protecting against a concurrent path access

newFileSem

Arguments

:: FilePath
path to be guarded by a semaphore allowing only one concurrent to access it

-> FileSem

Create new file semaphore. Always use with NOINLINE pragma! Example: createTestnetDataOutSem :: FileSem createTestnetDataOutSem = newFileSem "testcardano-cli-goldenfilesgoldenconway/create-testnet-data.out" {-# NOINLINE createTestnetDataOutSem #-}