Cardano.CLI.Byron.Delegation

Documentation

data ByronDelegationError

Constructors

CertificateValidationErrors !FilePath ![Text]

DlgCertificateDeserialisationFailed !FilePath !Text

ByronDelegationKeyError !ByronKeyFailure

checkByronGenesisDelegation :: CertificateFile -> ProtocolMagicId -> VerificationKey -> VerificationKey -> ExceptT ByronDelegationError IO ()

Verify that a certificate signifies genesis delegation by assumed genesis key to a delegate key, for a given protocol magic. If certificate fails validation, throw an error.

issueByronGenesisDelegation :: ProtocolMagicId -> EpochNumber -> SigningKey -> VerificationKey -> Certificate

Issue a certificate for genesis delegation to a delegate key, signed by the issuer key, for a given protocol magic and coming into effect at given epoch.

renderByronDelegationError :: ByronDelegationError -> Doc ann

serialiseDelegationCert :: Certificate -> ByteString

serialiseByronWitness :: SomeByronSigningKey -> ByteString