Cardano.CLI.Byron.Genesis

Documentation

data ByronGenesisError

Constructors

ByronDelegationCertSerializationError !ByronDelegationError

ByronDelegationKeySerializationError ByronDelegationError

GenesisGenerationError !GenesisDataGenerationError

GenesisOutputDirAlreadyExists FilePath

GenesisReadError !FilePath !GenesisDataError

GenesisSpecError !Text

MakeGenesisDelegationError !GenesisDelegationError

NoGenesisDelegationForKey !Text

ProtocolParametersParseFailed !FilePath !Text

PoorKeyFailure !ByronKeyFailure

data GenesisParameters

Parameters required for generation of new genesis.

Constructors

GenesisParameters

Fields

gpStartTime ∷ !UTCTime

gpProtocolParamsFile ∷ !FilePath

gpK ∷ !BlockCount

gpProtocolMagic ∷ !ProtocolMagic

gpTestnetBalance ∷ !TestnetBalanceOptions

gpFakeAvvmOptions ∷ !FakeAvvmOptions

gpAvvmBalanceFactor ∷ !LovelacePortion

gpSeed ∷ !(Maybe Integer)

newtype NewDirectory

Constructors

NewDirectory FilePath

dumpGenesis ∷ NewDirectory → GenesisData → GeneratedSecrets → ExceptT ByronGenesisError IO ()

Write out genesis into a directory that must not yet exist. An error is thrown if the directory already exists, or the genesis has delegate keys that are not delegated to.

mkGenesis ∷ GenesisParameters → ExceptT ByronGenesisError IO (GenesisData, GeneratedSecrets)

Generate a genesis, for given blockchain start time, protocol parameters, security parameter, protocol magic, testnet balance options, fake AVVM options, AVVM balance factor and seed. Throw an error in the following cases: if the protocol parameters file can't be read or fails parse, if genesis delegation couldn't be generated, if the parameter-derived genesis specification is wrong, or if the genesis fails generation.

readGenesis ∷ GenesisFile → NetworkId → ExceptT ByronGenesisError IO Config

Read genesis from a file.

renderByronGenesisError ∷ ByronGenesisError → Text