Cardano.CLI.Byron.Tx

Documentation

data ByronTxError

Constructors

TxDeserialisationFailed !FilePath !DecoderError

ByronTxSubmitError !Text

ByronTxSubmitErrorEraMismatch !EraMismatch

data Tx era

type TxFile = File (Tx ())

newtype NewTxFile

Constructors

NewTxFile FilePath

prettyAddress :: Address ByronAddr -> Text

Pretty-print an address in its Base58 form, and also its full structure.

readByronTx :: TxFile 'In -> ExceptT ByronTxError IO (ATxAux ByteString)

normalByronTxToGenTx :: ATxAux ByteString -> GenTx ByronBlock

The GenTx is all the kinds of transactions that can be submitted and "normal" Byron transactions are just one of the kinds.

txSpendGenesisUTxOByronPBFT :: Config -> NetworkId -> SomeByronSigningKey -> Address ByronAddr -> [TxOut CtxTx ByronEra] -> ATxAux ByteString

Generate a transaction spending genesis UTxO at a given address, to given outputs, signed by the given key.

txSpendUTxOByronPBFT :: NetworkId -> SomeByronSigningKey -> [TxIn] -> [TxOut CtxTx ByronEra] -> ATxAux ByteString

Generate a transaction from given Tx inputs to outputs, signed by the given key.

nodeSubmitTx :: SocketPath -> NetworkId -> GenTx ByronBlock -> ExceptT ByronTxError IO ()

Submit a transaction to a node specified by topology info.

renderByronTxError :: ByronTxError -> Doc ann

fromCborTxAux :: ByteString -> Either DecoderError (ATxAux ByteString)

toCborTxAux :: ATxAux ByteString -> ByteString

data ScriptValidity

Constructors

ScriptInvalid

ScriptValid