Cardano.CLI.EraBased.Commands.Governance.Actions

Documentation

data GovernanceActionCmds era

Constructors

GovernanceActionCreateConstitutionCmd !(GovernanceActionCreateConstitutionCmdArgs era)

GovernanceActionUpdateCommitteeCmd !(GovernanceActionUpdateCommitteeCmdArgs era)

GovernanceActionCreateNoConfidenceCmd !(GovernanceActionCreateNoConfidenceCmdArgs era)

GovernanceActionProtocolParametersUpdateCmd !(GovernanceActionProtocolParametersUpdateCmdArgs era)

GovernanceActionTreasuryWithdrawalCmd !(GovernanceActionTreasuryWithdrawalCmdArgs era)

GovernanceActionHardforkInitCmd !(GovernanceActionHardforkInitCmdArgs era)

GovernanceActionInfoCmd !(GovernanceActionInfoCmdArgs era)

GovernanceActionViewCmd !(GovernanceActionViewCmdArgs era)

data GovernanceActionUpdateCommitteeCmdArgs era

Constructors

GovernanceActionUpdateCommitteeCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnAddress :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

oldCommitteeVkeySource :: ![VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey]

newCommitteeVkeySource :: ![(VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey, EpochNo)]

requiredThreshold :: !Rational

mPrevGovernanceActionId :: !(Maybe (TxId, Word16))

outFile :: !(File () 'Out)

data GovernanceActionCreateConstitutionCmdArgs era

Constructors

GovernanceActionCreateConstitutionCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

stakeCredential :: !StakeIdentifier

mPrevGovernanceActionId :: !(Maybe (TxId, Word16))

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

constitutionUrl :: !ConstitutionUrl

constitutionHash :: !(SafeHash StandardCrypto AnchorData)

checkConstitutionHash :: !(MustCheckHash ConstitutionUrl)

constitutionScript :: !(Maybe ScriptHash)

outFile :: !(File () 'Out)

data GovernanceActionCreateNoConfidenceCmdArgs era

Constructors

GovernanceActionCreateNoConfidenceCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnStakeAddress :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

mPrevGovernanceActionId :: !(Maybe (TxId, Word16))

outFile :: !(File () 'Out)

data GovernanceActionInfoCmdArgs era

Datatype to carry data for the create-info governance action

Constructors

GovernanceActionInfoCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnStakeAddress :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

outFile :: !(File () 'Out)

data GovernanceActionViewCmdArgs era

Constructors

GovernanceActionViewCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

actionFile :: !(ProposalFile 'In)

outFormat :: !ViewOutputFormat

mOutFile :: !(Maybe (File () 'Out))

data GovernanceActionProtocolParametersUpdateCmdArgs era

Constructors

GovernanceActionProtocolParametersUpdateCmdArgs

Fields

uppShelleyBasedEra :: !(ShelleyBasedEra era)

uppPreConway :: !(Maybe (UpdateProtocolParametersPreConway era))

uppConwayOnwards :: !(Maybe (UpdateProtocolParametersConwayOnwards era))

uppNewPParams :: !(EraBasedProtocolParametersUpdate era)

New parameters to be proposed. From Alonzo onwards, the type EraBasedProtocolParametersUpdate also contains cost models. Since all other protocol parameters are read from command line arguments, whereas the cost models are read from a file, we separate the cost models from the rest of the protocol parameters to ease parsing.

uppCostModelsFile :: !(Maybe (CostModelsFile era))

The new cost models proposed. See the comment at uppNewPParams for why this is a separate field.

uppFilePath :: !(File () 'Out)

data GovernanceActionTreasuryWithdrawalCmdArgs era

Constructors

GovernanceActionTreasuryWithdrawalCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnAddr :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

treasuryWithdrawal :: ![(StakeIdentifier, Lovelace)]

constitutionScriptHash :: !(Maybe ScriptHash)

outFile :: !(File () 'Out)

data UpdateProtocolParametersConwayOnwards era

Constructors

UpdateProtocolParametersConwayOnwards

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnAddr :: !StakeIdentifier

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

governanceActionId :: !(Maybe (TxId, Word16))

constitutionScriptHash :: !(Maybe ScriptHash)

data UpdateProtocolParametersPreConway era

Constructors

UpdateProtocolParametersPreConway

Fields

eon :: !(ShelleyToBabbageEra era)

expiryEpoch :: !EpochNo

genesisVerificationKeys :: ![VerificationKeyFile 'In]

data GovernanceActionHardforkInitCmdArgs era

Constructors

GovernanceActionHardforkInitCmdArgs

Fields

eon :: !(ConwayEraOnwards era)

networkId :: !Network

deposit :: !Lovelace

returnStakeAddress :: !StakeIdentifier

mPrevGovernanceActionId :: !(Maybe (TxId, Word16))

proposalUrl :: !ProposalUrl

proposalHash :: !(SafeHash StandardCrypto AnchorData)

checkProposalHash :: !(MustCheckHash ProposalUrl)

protVer :: !ProtVer

outFile :: !(File () 'Out)

data CostModelsFile era

Constructors

CostModelsFile

Fields

eon :: !(AlonzoEraOnwards era)

costModelsFile :: !(File CostModels 'In)

renderGovernanceActionCmds :: GovernanceActionCmds era -> Text