Cardano.CLI.EraBased.Options.Common

Documentation

command' :: String -> String -> Parser a -> Mod CommandFields a

prefixFlag :: Maybe String -> String -> String

prefixFlag Nothing bar is bar, while prefixFlag (Just "foo") bar is foo-bar. This function is used to optionally prefix some long flags

bounded :: (Bounded a, Integral a, Show a) => String -> ReadM a

parseFilePath :: String -> String -> Parser FilePath

pNetworkIdDeprecated :: Parser NetworkId

pNetworkId :: EnvCli -> Parser NetworkId

pTarget :: ShelleyBasedEra era -> Parser (Target ChainPoint)

toUnitIntervalOrErr :: Rational -> UnitInterval

pConsensusModeParams :: Parser ConsensusModeParams

defaultByronEpochSlots :: Word64

pEpochSlots :: Parser EpochSlots

pSocketPath :: EnvCli -> Parser SocketPath

readerFromParsecParser :: Parser a -> ReadM a

parseTxIn :: Parser TxIn

parseTxId :: Parser TxId

parseTxIx :: Parser TxIx

decimal :: Parser Integer

pStakeIdentifier :: Maybe String -> Parser StakeIdentifier

pStakeVerifier :: Maybe String -> Parser StakeVerifier

pStakeAddress :: Maybe String -> Parser StakeAddress

parseStakeAddress :: Parser StakeAddress

pStakeVerificationKeyOrFile :: Maybe String -> Parser (VerificationKeyOrFile StakeKey)

First argument is the optional prefix

pScriptFor :: String -> Maybe String -> String -> Parser ScriptFile

pStakeVerificationKey :: Maybe String -> Parser (VerificationKey StakeKey)

The first argument is the optional prefix.

readVerificationKey :: SerialiseAsBech32 (VerificationKey keyrole) => AsType keyrole -> ReadM (VerificationKey keyrole)

Read a Bech32 or hex-encoded verification key.

pStakeVerificationKeyFile :: Maybe String -> Parser (VerificationKeyFile 'In)

The first argument is the optional prefix.

subInfoParser :: String -> InfoMod a -> [Maybe (Parser a)] -> Maybe (Parser a)

pAnyShelleyBasedEra :: EnvCli -> Parser (EraInEon ShelleyBasedEra)

deprecationText :: String

pAnyShelleyToBabbageEra :: EnvCli -> Parser (EraInEon ShelleyToBabbageEra)

pFileOutDirection :: String -> String -> Parser (File a 'Out)

pFileInDirection :: String -> String -> Parser (File a 'In)

parseLovelace :: Parser Lovelace

pStakePoolVerificationKeyOrFile :: Maybe String -> Parser (VerificationKeyOrFile StakePoolKey)

The first argument is the optional prefix.

pStakePoolVerificationKey :: Maybe String -> Parser (VerificationKey StakePoolKey)

The first argument is the optional prefix.

pStakePoolVerificationKeyFile :: Maybe String -> Parser (VerificationKeyFile 'In)

The first argument is the optional prefix.

pOutputFile :: Parser (File content 'Out)

pMIRPot :: Parser MIRPot

pRewardAmt :: Parser Lovelace

pTransferAmt :: Parser Lovelace

pTreasuryWithdrawalAmt :: Parser Lovelace

rHexHash

Arguments

:: SerialiseAsRawBytes (Hash a)

=> AsType a

-> Maybe String
Optional prefix to the error message

-> ReadM (Hash a)

rBech32KeyHash :: SerialiseAsBech32 (Hash a) => AsType a -> ReadM (Hash a)

pGenesisDelegateVerificationKey :: Parser (VerificationKey GenesisDelegateKey)

rVerificationKey

Arguments

:: SerialiseAsRawBytes (VerificationKey a)

=> AsType a
Singleton value identifying the kind of verification keys

-> Maybe String
Optional prefix to the error message

-> ReadM (VerificationKey a)

Reader for verification keys

pColdVerificationKeyOrFile :: Maybe String -> Parser ColdVerificationKeyOrFile

The first argument is the optional prefix.

pColdVerificationKeyFile :: forall (direction :: FileDirection). Parser (VerificationKeyFile direction)

pColdSigningKeyFile :: forall keyrole (direction :: FileDirection). Parser (File (SigningKey keyrole) direction)

pVerificationKeyFileOut :: Parser (File (VerificationKey keyrole) 'Out)

pSigningKeyFileOut :: Parser (File (SigningKey keyrole) 'Out)

pOperatorCertIssueCounterFile :: forall (direction :: FileDirection). Parser (File OpCertCounter direction)

pAddCommitteeColdVerificationKeySource :: Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey)

pAddCommitteeColdVerificationKeyHash :: Parser (Hash CommitteeColdKey)

pAddCommitteeColdVerificationKeyOrFile :: Parser (VerificationKeyOrFile CommitteeColdKey)

pAddCommitteeColdVerificationKey :: Parser (VerificationKey CommitteeColdKey)

pAddCommitteeColdVerificationKeyFile :: Parser (File (VerificationKey keyrole) 'In)

pRemoveCommitteeColdVerificationKeySource :: Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeColdKey)

pScriptHash

Arguments

:: String
long option name

-> String
help text

-> Parser ScriptHash

pRemoveCommitteeColdVerificationKeyHash :: Parser (Hash CommitteeColdKey)

pRemoveCommitteeColdVerificationKeyOrFile :: Parser (VerificationKeyOrFile CommitteeColdKey)

pRemoveCommitteeColdVerificationKey :: Parser (VerificationKey CommitteeColdKey)

deserialiseColdCCKeyFromHex :: ReadM (VerificationKey CommitteeColdKey)

deserialiseColdCCKeyHashFromHex :: ReadM (Hash CommitteeColdKey)

pRemoveCommitteeColdVerificationKeyFile :: Parser (File (VerificationKey keyrole) 'In)

pCommitteeColdVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile CommitteeColdKey)

pCommitteeColdVerificationKeyOrFile :: Parser (VerificationKeyOrFile CommitteeColdKey)

pCommitteeColdVerificationKey :: Parser (VerificationKey CommitteeColdKey)

pCommitteeColdVerificationKeyHash :: Parser (Hash CommitteeColdKey)

pCommitteeColdVerificationKeyFile :: Parser (File (VerificationKey keyrole) 'In)

pVerificationKeyFileIn :: Parser (VerificationKeyFile 'In)

pAnyVerificationKeyFileIn :: String -> Parser (VerificationKeyFile 'In)

pAnyVerificationKeyText :: String -> Parser AnyVerificationKeyText

pAnyVerificationKeySource :: String -> Parser AnyVerificationKeySource

pCommitteeHotKey :: Parser (VerificationKey CommitteeHotKey)

pCommitteeHotVerificationKeyOrFile :: Parser (VerificationKeyOrFile CommitteeHotKey)

pCommitteeHotVerificationKeyHash :: Parser (Hash CommitteeHotKey)

pCommitteeHotVerificationKey :: String -> Parser (VerificationKey CommitteeHotKey)

deserialiseHotCCKeyFromHex :: ReadM (VerificationKey CommitteeHotKey)

deserialiseHotCCKeyHashFromHex :: ReadM (Hash CommitteeHotKey)

pCommitteeHotVerificationKeyFile :: String -> Parser (VerificationKeyFile 'In)

pCommitteeHotKeyHash :: Maybe String -> Parser (Hash CommitteeHotKey)

The first argument is the optional prefix.

pCommitteeHotKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile CommitteeHotKey)

pCommitteeHotVerificationKeyOrHashOrVerificationFile :: Parser (VerificationKeyOrHashOrFile CommitteeHotKey)

pCommitteeHotVerificationKeyOrHashOrVerificationFileOrScriptHash :: Parser (VerificationKeyOrHashOrFileOrScriptHash CommitteeHotKey)

catCommands :: [Parser a] -> Maybe (Parser a)

pConstitutionUrl :: Parser ConstitutionUrl

pConstitutionHash :: Parser (SafeHash StandardCrypto AnchorData)

pUrl :: String -> String -> Parser Url

pGovActionDeposit :: Parser Lovelace

pNewGovActionDeposit :: Parser Lovelace

pStakeVerificationKeyOrHashOrFile :: Maybe String -> Parser (VerificationKeyOrHashOrFile StakeKey)

First argument is the optional prefix

pStakeVerificationKeyHash :: Maybe String -> Parser (Hash StakeKey)

First argument is the optional prefix

pStakePoolVerificationKeyOrHashOrFile :: Maybe String -> Parser (VerificationKeyOrHashOrFile StakePoolKey)

The first argument is the optional prefix.

pCBORInFile :: Parser FilePath

pPollQuestion :: Parser Text

pPollAnswer :: Parser Text

pPollAnswerIndex :: Parser Word

pPollFile :: Parser (File GovernancePoll 'In)

pPollTxFile :: Parser (TxFile 'In)

pPollNonce :: Parser Word

pMintScriptFile :: Parser (File ScriptInAnyLang 'In)

pPlutusMintScriptWitnessData :: ShelleyBasedEra era -> WitCtx witctx -> BalanceTxExecUnits -> Parser (ScriptDataOrFile, ExecutionUnits)

pScriptWitnessFiles

Arguments

:: forall witctx era. ShelleyBasedEra era

-> WitCtx witctx

-> BalanceTxExecUnits
Use the execution-units flag.

-> String
Script flag prefix

-> Maybe String

-> String

-> Parser (ScriptWitnessFiles witctx)

pExecutionUnits :: String -> Parser ExecutionUnits

pScriptRedeemerOrFile :: String -> Parser ScriptDataOrFile

pScriptDatumOrFileCip69 :: String -> WitCtx witctx -> Parser (ScriptDatumOrFile witctx)

pScriptDatumOrFile :: String -> WitCtx witctx -> Parser (ScriptDatumOrFile witctx)

pScriptDataOrFile

Arguments

:: String
data flag prefix

-> String
value help text

-> String
file help text

-> Parser ScriptDataOrFile

pVoteFiles :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser [(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

pVoteFile :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))

pProposalFiles :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser [(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

pProposalFile :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))

pCurrentTreasuryValueAndDonation :: ShelleyBasedEra era -> Parser (Maybe (TxCurrentTreasuryValue, TxTreasuryDonation))

pCurrentTreasuryValue' :: Parser TxCurrentTreasuryValue

pTreasuryDonation :: ShelleyBasedEra era -> Parser (Maybe TxTreasuryDonation)

pTreasuryDonation' :: Parser TxTreasuryDonation

pPaymentVerifier :: Parser PaymentVerifier

pPaymentVerificationKeyTextOrFile :: Parser VerificationKeyTextOrFile

pPaymentVerificationKeyText :: Parser Text

pPaymentVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pScript :: Parser ScriptFile

pReferenceTxIn :: String -> String -> Parser TxIn

pReadOnlyReferenceTxIn :: Parser TxIn

pAddressKeyType :: Parser AddressKeyType

pProtocolParamsFile :: Parser ProtocolParamsFile

pTxBuildOutputOptions :: Parser TxBuildOutputOptions

pCertificateFile :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))

pPoolMetadataFile :: Parser (StakePoolMetadataFile 'In)

pTxMetadataJsonSchema :: Parser TxMetadataJsonSchema

convertTime :: String -> UTCTime

pMetadataFile :: Parser MetadataFile

pWithdrawal :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (StakeAddress, Lovelace, Maybe (ScriptWitnessFiles WitCtxStake))

pPlutusStakeReferenceScriptWitnessFilesVotingProposing

Arguments

:: String

-> BalanceTxExecUnits
Use the execution-units flag.

-> Parser (ScriptWitnessFiles WitCtxStake)

pPlutusStakeReferenceScriptWitnessFiles

Arguments

:: String

-> BalanceTxExecUnits
Use the execution-units flag.

-> Parser (ScriptWitnessFiles WitCtxStake)

pPlutusScriptLanguage :: String -> Parser AnyPlutusScriptVersion

plutusP :: IsPlutusScriptLanguage lang => String -> PlutusScriptVersion lang -> String -> Parser AnyPlutusScriptVersion

pUpdateProposalFile :: Parser UpdateProposalFile

pRequiredSigner :: Parser RequiredSigner

pVrfSigningKeyFile :: Parser (SigningKeyFile 'In)

pWhichLeadershipSchedule :: Parser EpochLeadershipSchedule

pWitnessSigningData :: Parser WitnessSigningData

pSigningKeyFileIn :: Parser (SigningKeyFile 'In)

pKesPeriod :: Parser KESPeriod

pEpochNo :: String -> Parser EpochNo

pEpochNoUpdateProp :: Parser EpochNo

pGenesisFile :: String -> Parser GenesisFile

pOperationalCertificateFile :: forall (direction :: FileDirection). Parser (File () direction)

pKeyOutputFormat :: Parser KeyOutputFormat

pPoolIdOutputFormat :: Parser IdOutputFormat

pOutputFormatJsonOrText :: String -> Parser OutputFormatJsonOrText

pOutputFormatJsonOrText kind is a parser to specify in which format to view some data (json or text). kind is the kind of data considered.

pTxViewOutputFormat :: Parser ViewOutputFormat

pGovernanceActionViewOutputFormat :: Parser ViewOutputFormat

pGovernanceVoteViewOutputFormat :: Parser ViewOutputFormat

pViewOutputFormat :: String -> Parser ViewOutputFormat

pViewOutputFormat kind is a parser to specify in which format to view some data (json or yaml). what is the kind of data considered.

pMaybeOutputFile :: Parser (Maybe (File content 'Out))

pVerificationKey :: SerialiseAsBech32 (VerificationKey keyrole) => AsType keyrole -> Parser (VerificationKey keyrole)

pVerificationKeyOrFileIn :: SerialiseAsBech32 (VerificationKey keyrole) => AsType keyrole -> Parser (VerificationKeyOrFile keyrole)

pExtendedVerificationKeyFileIn :: Parser (VerificationKeyFile 'In)

pGenesisVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pGenesisVerificationKeyHash :: Parser (Hash GenesisKey)

pGenesisVerificationKey :: Parser (VerificationKey GenesisKey)

pGenesisVerificationKeyOrFile :: Parser (VerificationKeyOrFile GenesisKey)

pGenesisVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile GenesisKey)

pGenesisDelegateVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pGenesisDelegateVerificationKeyHash :: Parser (Hash GenesisDelegateKey)

pGenesisDelegateVerificationKeyOrFile :: Parser (VerificationKeyOrFile GenesisDelegateKey)

pGenesisDelegateVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile GenesisDelegateKey)

pKesVerificationKeyOrFile :: Parser (VerificationKeyOrFile KesKey)

pKesVerificationKey :: Parser (VerificationKey KesKey)

pKesVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pTxSubmitFile :: Parser FilePath

pTxIn :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))

pTxInCollateral :: Parser TxIn

pReturnCollateral :: Parser TxOutShelleyBasedEra

pTotalCollateral :: Parser Lovelace

pWitnessOverride :: Parser Word

pNumberOfShelleyKeyWitnesses :: Parser Int

pNumberOfByronKeyWitnesses :: Parser Int

pTotalUTxOValue :: Parser Value

pTxOut :: Parser TxOutAnyEra

pTxOutShelleyBased :: Parser TxOutShelleyBasedEra

pTxOutDatum :: Parser TxOutDatumAnyEra

pRefScriptFp :: Parser ReferenceScriptAnyEra

pMintMultiAsset :: ShelleyBasedEra era -> BalanceTxExecUnits -> Parser (Value, [CliMintScriptRequirements])

pPolicyId :: Parser PolicyId

pInvalidBefore :: Parser SlotNo

pLegacyInvalidHereafter :: Parser SlotNo

pInvalidHereafter :: ShelleyBasedEra era -> Parser (TxValidityUpperBound era)

pTxFee :: Parser Lovelace

pWitnessFile :: Parser WitnessFile

pTxBodyFileIn :: Parser (TxBodyFile 'In)

pTxBodyFileOut :: Parser (TxBodyFile 'Out)

pTxFileIn :: Parser (TxFile 'In)

pTxFileOut :: Parser (TxFile 'Out)

pInputTxOrTxBodyFile :: Parser InputTxBodyOrTxFile

pTxInCountDeprecated :: Parser TxInCount

pTxOutCountDeprecated :: Parser TxOutCount

pTxShelleyWitnessCount :: Parser TxShelleyWitnessCount

pTxByronWitnessCount :: Parser TxByronWitnessCount

pQueryUTxOFilter :: Parser QueryUTxOFilter

pFilterByStakeAddress :: Parser StakeAddress

pByronAddress :: Parser (Address ByronAddr)

pAddress :: Parser Text

pStakePoolVerificationKeyHash :: Maybe String -> Parser (Hash StakePoolKey)

First argument is the prefix for the option's flag to use

pVrfVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pVrfVerificationKeyHash :: Parser (Hash VrfKey)

pVrfVerificationKey :: Parser (VerificationKey VrfKey)

pVrfVerificationKeyOrFile :: Parser (VerificationKeyOrFile VrfKey)

pVrfVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile VrfKey)

pRewardAcctVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pRewardAcctVerificationKey :: Parser (VerificationKey StakeKey)

pRewardAcctVerificationKeyOrFile :: Parser (VerificationKeyOrFile StakeKey)

pPoolOwnerVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pPoolOwnerVerificationKey :: Parser (VerificationKey StakeKey)

pPoolOwnerVerificationKeyOrFile :: Parser (VerificationKeyOrFile StakeKey)

pPoolPledge :: Parser Lovelace

pPoolCost :: Parser Lovelace

pRational :: String -> String -> Parser Rational

pPoolMargin :: Parser Rational

pPoolRelay :: Parser StakePoolRelay

pMultiHostName :: Parser StakePoolRelay

pSingleHostName :: Parser StakePoolRelay

pSingleHostAddress :: Parser StakePoolRelay

pIpV4 :: Parser IPv4

pIpV6 :: Parser IPv6

pPort :: Parser PortNumber

pStakePoolMetadataReference :: Parser StakePoolMetadataReference

pStakePoolMetadataUrl :: Parser Text

pStakePoolMetadataHash :: Parser (Hash StakePoolMetadata)

pStakePoolRegistrationParserRequirements :: EnvCli -> Parser StakePoolRegistrationParserRequirements

pProtocolParametersUpdate :: Parser ProtocolParametersUpdate

pCostModels :: Parser FilePath

pMinFeePerByteFactor :: Parser Lovelace

pMinFeeConstantFactor :: Parser Lovelace

pMinUTxOValue :: Parser Lovelace

pMinPoolCost :: Parser Lovelace

pMaxBodySize :: Parser Word32

pMaxTransactionSize :: Parser Word32

pairIntegralReader :: (Typeable a, Integral a, Bits a) => ReadM (a, a)

A parser for (Int, Int)-like expressions. In other words, integralReader-lifted to a pairs with a Haskell-like syntax.

pairIntegralParsecParser :: (Typeable a, Integral a, Bits a) => Parser (a, a)

integralReader :: (Typeable a, Integral a, Bits a) => ReadM a

integralReader is a reader for a word of type a. When it fails parsing, it provides a nice error message. This custom reader is needed to avoid the overflow issues of auto described in https://github.com/IntersectMBO/cardano-cli/issues/860.

integralParsecParser :: (Typeable a, Integral a, Bits a) => Parser a

pMaxBlockHeaderSize :: Parser Word16

pKeyRegistDeposit :: Parser Lovelace

pDrepDeposit :: Parser Lovelace

pPoolDeposit :: Parser Lovelace

pEpochBoundRetirement :: Parser EpochInterval

pNumberOfPools :: Parser Natural

pPoolInfluence :: Parser Rational

pTreasuryExpansion :: Parser Rational

pMonetaryExpansion :: Parser Rational

pDecentralParam :: Parser Rational

pExtraEntropy :: Parser (Maybe PraosNonce)

pUTxOCostPerByte :: Parser Lovelace

pExecutionUnitPrices :: Parser ExecutionUnitPrices

pMaxTxExecutionUnits :: Parser ExecutionUnits

pMaxBlockExecutionUnits :: Parser ExecutionUnits

pMaxValueSize :: Parser Natural

pCollateralPercent :: Parser Natural

pMaxCollateralInputs :: Parser Natural

pProtocolVersion :: Parser (Natural, Natural)

pPoolVotingThresholds :: Parser PoolVotingThresholds

pDRepVotingThresholds :: Parser DRepVotingThresholds

pMinCommitteeSize :: Parser Natural

pCommitteeTermLength :: Parser EpochInterval

pGovActionLifetime :: Parser EpochInterval

pDRepDeposit :: Parser Lovelace

pDRepActivity :: Parser EpochInterval

parseTxOutShelleyBasedEra :: Parser (TxOutDatumAnyEra -> ReferenceScriptAnyEra -> TxOutShelleyBasedEra)

parseShelleyAddress :: Parser (Address ShelleyAddr)

parseTxOutAnyEra :: Parser (TxOutDatumAnyEra -> ReferenceScriptAnyEra -> TxOutAnyEra)

pVoteChoice :: Parser Vote

pVoterType :: Parser VType

pVotingCredential :: Parser (VerificationKeyOrFile StakePoolKey)

pVoteDelegationTarget :: Parser VoteDelegationTarget

pAlwaysAbstain :: Parser ()

pVoteAnchor :: Parser (VoteUrl, SafeHash StandardCrypto AnchorData)

pVoteAnchorDataHash :: Parser (SafeHash StandardCrypto AnchorData)

pAlwaysNoConfidence :: Parser ()

pDrepRefund :: Parser (DRepHashSource, Lovelace)

pDepositRefund :: Parser Lovelace

pDRepHashSource :: Parser DRepHashSource

pSPOHashSource :: Parser SPOHashSource

pDRepScriptHash :: Parser ScriptHash

pSPOScriptHash :: Parser ScriptHash

pConstitutionScriptHash :: Parser ScriptHash

pDRepVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile DRepKey)

pSPOVerificationKeyOrHashOrFile :: Parser (VerificationKeyOrHashOrFile StakePoolKey)

pDRepVerificationKeyOrHashOrFileOrScriptHash :: Parser (VerificationKeyOrHashOrFileOrScriptHash DRepKey)

pAllOrOnlyDRepHashSource :: Parser (AllOrOnly DRepHashSource)

pAllOrOnlySPOHashSource :: Parser (AllOrOnly SPOHashSource)

pAllOrOnlyGovActionIds :: ConwayEraOnwards era -> Parser (AllOrOnly (GovActionId StandardCrypto))

pGovActionIds :: ConwayEraOnwards era -> Parser [GovActionId StandardCrypto]

pDRepVerificationKeyHash :: Parser (Hash DRepKey)

pDRepVerificationKey :: Parser (VerificationKey DRepKey)

pDRepVerificationKeyOrFile :: Parser (VerificationKeyOrFile DRepKey)

pDRepVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pSPOVerificationKeyHash :: Parser (Hash StakePoolKey)

pSPOVerificationKey :: Parser (VerificationKey StakePoolKey)

pSPOVerificationKeyOrFile :: Parser (VerificationKeyOrFile StakePoolKey)

pSPOVerificationKeyFile :: Parser (VerificationKeyFile 'In)

pAnchorUrl :: Parser ProposalUrl

pExpectedAnchorDataHash :: Parser (SafeHash StandardCrypto AnchorData)

pExpectedHash :: (SafeHash StandardCrypto AnchorData -> a) -> String -> Parser a

pAnchorDataHash :: Parser (SafeHash StandardCrypto AnchorData)

pMustCheckHash :: String -> String -> String -> String -> Parser (MustCheckHash anchorData)

pPotentiallyCheckedAnchorData :: Parser (MustCheckHash anchorType) -> Parser anchor -> Parser (PotentiallyCheckedAnchor anchorType anchor)

pMustCheckProposalHash :: Parser (MustCheckHash ProposalUrl)

pMustCheckConstitutionHash :: Parser (MustCheckHash ConstitutionUrl)

pMustCheckMetadataHash :: Parser (MustCheckHash DRepMetadataUrl)

pMustCheckStakeMetadataHash :: Parser (MustCheckHash StakePoolMetadataReference)

pMustCheckVoteUrl :: Parser (MustCheckHash VoteUrl)

pMustCheckResignationMetadataHash :: Parser (MustCheckHash ResignationMetadataUrl)

pPreviousGovernanceAction :: Parser (Maybe (TxId, Word16))

pGovernanceActionId :: Parser (TxId, Word16)

pWord16 :: String -> String -> Parser Word16

pTxId :: String -> String -> Parser TxId

pNetworkIdForTestnetData :: EnvCli -> Parser NetworkId

pReferenceScriptSize :: Parser ReferenceScriptSize

pFeatured :: forall (eon :: Type -> Type) f era a. (Eon eon, ToCardanoEra f) => f era -> Parser a -> Parser (Maybe (Featured eon era a))

hiddenSubParser :: String -> ParserInfo a -> Parser a