Cardano.CLI.EraBased.Query.Run

Documentation

runQueryCmds :: QueryCmds era -> CIO e ()

runQueryKesPeriodInfoCmd :: QueryKesPeriodInfoCmdArgs -> CIO e ()

runQueryLeadershipScheduleCmd :: QueryLeadershipScheduleCmdArgs -> CIO e ()

runQueryLedgerStateCmd :: QueryLedgerStateCmdArgs -> CIO e ()

runQueryLedgerPeerSnapshot :: QueryLedgerPeerSnapshotCmdArgs -> CIO e ()

runQueryPoolStateCmd :: QueryPoolStateCmdArgs -> CIO e ()

Query the current and future parameters for a stake pool, including the retirement date. Any of these may be empty (in which case a null will be displayed).

runQueryProtocolParametersCmd :: QueryProtocolParametersCmdArgs -> CIO e ()

runQueryProtocolStateCmd :: QueryProtocolStateCmdArgs -> CIO e ()

runQuerySlotNumberCmd :: QuerySlotNumberCmdArgs -> CIO e ()

runQueryStakeAddressInfoCmd :: QueryStakeAddressInfoCmdArgs -> CIO e ()

Query the current delegations and reward accounts, filtered by a given set of addresses, from a Shelley node via the local state query protocol.

runQueryStakeDistributionCmd :: QueryStakeDistributionCmdArgs -> CIO e ()

runQueryStakePoolsCmd :: QueryStakePoolsCmdArgs -> CIO e ()

runQueryStakeSnapshotCmd :: QueryStakeSnapshotCmdArgs -> CIO e ()

Obtain stake snapshot information for a pool, plus information about the total active stake. This information can be used for leader slot calculation, for example, and has been requested by SPOs. Obtaining the information directly is significantly more time and memory efficient than using a full ledger state dump.

runQueryTipCmd :: QueryTipCmdArgs -> CIO e ()

runQueryTxMempoolCmd :: QueryTxMempoolCmdArgs -> CIO e ()

Query the local mempool state

runQueryUTxOCmd :: QueryUTxOCmdArgs -> CIO e ()

Query the UTxO, filtered by a given set of addresses, from a Shelley node via the local state query protocol.

newtype DelegationsAndRewards

A mapping of Shelley reward accounts to both the stake pool that they delegate to and their reward account balance.

Constructors

DelegationsAndRewards (Map StakeAddress Coin, Map StakeAddress PoolId)

renderQueryCmdError :: QueryCmdError -> Doc ann

renderOpCertIntervalInformation :: FilePath -> OpCertIntervalInformation -> Doc AnsiStyle

percentage

Arguments

:: RelativeTime
tolerance. If b - a < tolerance, then 100% is reported. This even if we are tolerance seconds behind, we are still considered fully synced.

-> RelativeTime
tipTime The time of the most recently synced block.

-> RelativeTime
nowTime The time of the tip of the block chain to which we need to sync.

-> Text

Calculate the percentage sync rendered as text: min 1 (tipTime/nowTime)