Cardano.CLI.EraBased.StakeAddress.Run

Documentation

runStakeAddressCmds :: StakeAddressCmds era -> CIO e ()

runStakeAddressBuildCmd :: StakeVerifier -> NetworkId -> Maybe (File () 'Out) -> CIO e ()

runStakeAddressKeyGenCmd :: Vary '[FormatBech32, FormatTextEnvelope] -> VerificationKeyFile 'Out -> SigningKeyFile 'Out -> CIO e (VerificationKey StakeKey, SigningKey StakeKey)

runStakeAddressKeyHashCmd :: VerificationKeyOrFile StakeKey -> Maybe (File () 'Out) -> CIO e ()

runStakeAddressStakeDelegationCertificateCmd

Arguments

:: IsEra era

=> Era era

-> StakeIdentifier
Delegator stake verification key, verification key file or script file.

-> StakePoolKeyHashSource
Delegatee stake pool verification key or verification key file or verification key hash.

-> File () 'Out

-> CIO e ()

runStakeAddressDeregistrationCertificateCmd

Arguments

:: Era era

-> StakeIdentifier

-> Lovelace
Deposit required in conway era

-> File () 'Out

-> CIO e ()

runStakeAddressRegistrationCertificateCmd

Arguments

:: Era era

-> StakeIdentifier

-> Lovelace
Deposit required in conway era

-> File () 'Out

-> CIO e ()

runStakeAddressRegistrationAndDelegationCertificateCmd

Arguments

:: Era era

-> StakeIdentifier

-> StakePoolKeyHashSource
Delegatee stake pool verification key or verification key file or id

-> Lovelace

-> File () 'Out

-> CIO e ()

runStakeAddressRegistrationStakeAndVoteDelegationCertificateCmd :: Era era -> StakeIdentifier -> StakePoolKeyHashSource -> VoteDelegationTarget -> Lovelace -> File () 'Out -> CIO e ()