Cardano.CLI.EraIndependent.Debug.Run

Documentation

data DebugCmdError

Constructors

DebugCmdFailed

DebugNodeConfigGenesisDataCmdError !FilePath !GenesisDataError
DebugNodeConfigGenesisDataCmdError filepath error represents an error when reading the node configuration at filepath

DebugNodeConfigWrongGenesisHashCmdError
DebugNodeConfigWrongGenesisHashCmdError filepath eraGenesisPath actualHash expectedHash represents a user error that the hash of eraGenesisPath in filepath is actualHash, whereas it should be expectedHash

Fields

!FilePath

The file path of the node configuration file

!FilePath

The file path of the era configuration file whose hash is wrong

!Text

The actual hash (the hash found by hashing the genesis file)

!Text

The expected hash (the hash mentioned in the configuration file)

DebugTxCmdError !TxCmdError

runLogEpochStateCmd :: LogEpochStateCmdArgs -> IO ()

runDebugCmds :: DebugCmds -> CIO e ()