Cardano.CLI.Legacy.Commands.Transaction

Documentation

data LegacyTransactionCmds

Constructors

TransactionBuildRawCmd

Fields

AnyCardanoEra

(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

[(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

[TxIn]

Read only reference inputs

[TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

(Maybe TxOutShelleyBasedEra)

Return collateral

(Maybe Coin)

Total collateral

[RequiredSigner]

Required signers

[TxOutAnyEra]

(Maybe (Value, [ScriptWitnessFiles WitCtxMint]))

Multi-Asset value with script witness

(Maybe SlotNo)

Transaction lower bound

(Maybe SlotNo)

Transaction upper bound

Coin

Tx fee

[(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

[(StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))]

TxMetadataJsonSchema

[ScriptFile]

Auxiliary scripts

[MetadataFile]

(Maybe ProtocolParamsFile)

(Maybe UpdateProposalFile)

(TxBodyFile 'Out)

TransactionBuildCmd
Like TransactionBuildRaw but without the fee, and with a change output.

Fields

SocketPath

(EraInEon ShelleyBasedEra)

ConsensusModeParams

NetworkId

(Maybe ScriptValidity)

Mark script as expected to pass or fail validation

(Maybe Word)

Override the required number of tx witnesses

[(TxIn, Maybe (ScriptWitnessFiles WitCtxTxIn))]

Transaction inputs with optional spending scripts

[TxIn]

Read only reference inputs

[RequiredSigner]

Required signers

[TxIn]

Transaction inputs for collateral, only key witnesses, no scripts.

(Maybe TxOutShelleyBasedEra)

Return collateral

(Maybe Coin)

Total collateral

[TxOutAnyEra]

Normal outputs

TxOutChangeAddress

A change output

(Maybe (Value, [ScriptWitnessFiles WitCtxMint]))

Multi-Asset value with script witness

(Maybe SlotNo)

Transaction lower bound

(Maybe SlotNo)

Transaction upper bound

[(CertificateFile, Maybe (ScriptWitnessFiles WitCtxStake))]

Certificates with potential script witness

[(StakeAddress, Coin, Maybe (ScriptWitnessFiles WitCtxStake))]

Withdrawals with potential script witness

TxMetadataJsonSchema

[ScriptFile]

Auxiliary scripts

[MetadataFile]

(Maybe UpdateProposalFile)

[(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

[(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))]

(Maybe TxTreasuryDonation)

TxBuildOutputOptions

TransactionSignCmd InputTxBodyOrTxFile [WitnessSigningData] (Maybe NetworkId) (TxFile 'Out)

TransactionWitnessCmd (TxBodyFile 'In) WitnessSigningData (Maybe NetworkId) (File () 'Out)

TransactionSignWitnessCmd (TxBodyFile 'In) [WitnessFile] (File () 'Out)

TransactionSubmitCmd SocketPath ConsensusModeParams NetworkId FilePath

TransactionPolicyIdCmd ScriptFile

TransactionCalculateMinFeeCmd (TxBodyFile 'In) ProtocolParamsFile TxShelleyWitnessCount TxByronWitnessCount ReferenceScriptSize (Maybe OutputFormatJsonOrText) (Maybe (File () 'Out))
The total size in bytes of the transaction reference scripts.

TransactionCalculateMinValueCmd (EraInEon ShelleyBasedEra) ProtocolParamsFile TxOutShelleyBasedEra

TransactionHashScriptDataCmd ScriptDataOrFile

TransactionTxIdCmd InputTxBodyOrTxFile

TransactionViewCmd

renderLegacyTransactionCmds :: LegacyTransactionCmds -> Text