Cardano.CLI.Read

Metadata

data MetadataError

Constructors

MetadataErrorFile (FileError ())

MetadataErrorJsonParseError !FilePath !String

MetadataErrorConversionError !FilePath !TxMetadataJsonError

MetadataErrorValidationError !FilePath ![(Word64, TxMetadataRangeError)]

MetadataErrorDecodeError !FilePath !DecoderError

renderMetadataError :: MetadataError -> Doc ann

readFileTxMetadata :: TxMetadataJsonSchema -> MetadataFile -> ExceptT MetadataError IO TxMetadata

readTxMetadata :: ShelleyBasedEra era -> TxMetadataJsonSchema -> [MetadataFile] -> IO (Either MetadataError (TxMetadataInEra era))

Script

data ScriptWitnessError

Constructors

ScriptWitnessErrorFile (FileError ScriptDecodeError)

ScriptWitnessErrorScriptLanguageNotSupportedInEra AnyScriptLanguage AnyCardanoEra

ScriptWitnessErrorExpectedSimple !FilePath !AnyScriptLanguage

ScriptWitnessErrorExpectedPlutus !FilePath !AnyScriptLanguage

ScriptWitnessErrorReferenceScriptsNotSupportedInEra !AnyShelleyBasedEra

ScriptWitnessErrorScriptData ScriptDataError

renderScriptWitnessError :: ScriptWitnessError -> Doc ann

readScriptDataOrFile :: forall (m :: Type -> Type). MonadIO m => ScriptDataOrFile -> ExceptT ScriptDataError m HashableScriptData

readScriptWitness :: ShelleyBasedEra era -> ScriptWitnessFiles witctx -> ExceptT ScriptWitnessError IO (ScriptWitness witctx era)

readScriptWitnessFiles :: ShelleyBasedEra era -> [(a, Maybe (ScriptWitnessFiles ctx))] -> ExceptT ScriptWitnessError IO [(a, Maybe (ScriptWitness ctx era))]

readScriptWitnessFilesTuple :: ShelleyBasedEra era -> [(a, b, Maybe (ScriptWitnessFiles ctx))] -> ExceptT ScriptWitnessError IO [(a, b, Maybe (ScriptWitness ctx era))]

data ScriptDecodeError

Constructors

ScriptDecodeTextEnvelopeError TextEnvelopeError

ScriptDecodeSimpleScriptError JsonDecodeError

ScriptDecodeUnknownPlutusScriptVersion Text

deserialiseScriptInAnyLang :: ByteString -> Either ScriptDecodeError ScriptInAnyLang

readFileScriptInAnyLang :: forall t (m :: Type -> Type). MonadIOTransError (FileError ScriptDecodeError) t m => FilePath -> t m ScriptInAnyLang

Read a script file. The file can either be in the text envelope format wrapping the binary representation of any of the supported script languages, or alternatively it can be a JSON format file for one of the simple script language versions.

readFileSimpleScript :: forall t (m :: Type -> Type). MonadIOTransError (FileError ScriptDecodeError) t m => FilePath -> t m (Script SimpleScript')

data AnyPlutusScript where

Constructors

AnyPlutusScript :: forall lang. IsPlutusScriptLanguage lang => PlutusScriptVersion lang -> PlutusScript lang -> AnyPlutusScript

data PlutusScriptDecodeError

Constructors

PlutusScriptDecodeErrorUnknownVersion !Text

PlutusScriptJsonDecodeError !JsonDecodeError

PlutusScriptDecodeTextEnvelopeError !TextEnvelopeError

PlutusScriptDecodeErrorVersionMismatch

Fields

!Text

Script version

!AnyPlutusScriptVersion

Attempted to decode with version

readFilePlutusScript :: forall t (m :: Type -> Type). MonadIOTransError (FileError PlutusScriptDecodeError) t m => FilePath -> t m AnyPlutusScript

Script data (datums and redeemers)

data ScriptDataError

Constructors

ScriptDataErrorFile (FileError ())

ScriptDataErrorJsonParse !FilePath !String

ScriptDataErrorConversion !FilePath !ScriptDataJsonError

ScriptDataErrorValidation !FilePath !ScriptDataRangeError

ScriptDataErrorMetadataDecode !FilePath !DecoderError

ScriptDataErrorJsonBytes !ScriptDataJsonBytesError

readScriptDatumOrFile :: ScriptDatumOrFile witctx -> ExceptT ScriptDataError IO (ScriptDatum witctx)

readScriptRedeemerOrFile :: ScriptRedeemerOrFile -> ExceptT ScriptDataError IO ScriptRedeemer

renderScriptDataError :: ScriptDataError -> Doc ann

Tx

data CddlError

Constructors

CddlErrorTextEnv !(FileError TextEnvelopeError) !(FileError TextEnvelopeCddlError)

CddlIOError (FileError TextEnvelopeError)

newtype CddlTx

Constructors

CddlTx

Fields

unCddlTx :: InAnyShelleyBasedEra Tx

newtype IncompleteCddlTxBody

Constructors

IncompleteCddlTxBody

Fields

unIncompleteCddlTxBody :: InAnyShelleyBasedEra TxBody

readFileTx :: FileOrPipe -> IO (Either (FileError TextEnvelopeCddlError) (InAnyShelleyBasedEra Tx))

readFileTxBody :: FileOrPipe -> IO (Either (FileError TextEnvelopeCddlError) IncompleteCddlTxBody)

readCddlTx :: FileOrPipe -> IO (Either (FileError TextEnvelopeCddlError) CddlTx)

txTextEnvelopeTypes :: [Text]

Tx witnesses

data ReadWitnessSigningDataError

Constructors

ReadWitnessSigningDataSigningKeyDecodeError !(FileError InputDecodeError)

ReadWitnessSigningDataScriptError !(FileError JsonDecodeError)

ReadWitnessSigningDataSigningKeyAndAddressMismatch
A Byron address was specified alongside a non-Byron signing key.

renderReadWitnessSigningDataError :: ReadWitnessSigningDataError -> Doc ann

Render an error message for a ReadWitnessSigningDataError.

data SomeSigningWitness

Constructors

AByronSigningWitness (SigningKey ByronKey) (Maybe (Address ByronAddr))

APaymentSigningWitness (SigningKey PaymentKey)

APaymentExtendedSigningWitness (SigningKey PaymentExtendedKey)

AStakeSigningWitness (SigningKey StakeKey)

AStakeExtendedSigningWitness (SigningKey StakeExtendedKey)

AStakePoolSigningWitness (SigningKey StakePoolKey)

AGenesisSigningWitness (SigningKey GenesisKey)

AGenesisExtendedSigningWitness (SigningKey GenesisExtendedKey)

AGenesisDelegateSigningWitness (SigningKey GenesisDelegateKey)

AGenesisDelegateExtendedSigningWitness (SigningKey GenesisDelegateExtendedKey)

AGenesisUTxOSigningWitness (SigningKey GenesisUTxOKey)

ADRepSigningWitness (SigningKey DRepKey)

ADRepExtendedSigningWitness (SigningKey DRepExtendedKey)

ACommitteeColdSigningWitness (SigningKey CommitteeColdKey)

ACommitteeColdExtendedSigningWitness (SigningKey CommitteeColdExtendedKey)

ACommitteeHotSigningWitness (SigningKey CommitteeHotKey)

ACommitteeHotExtendedSigningWitness (SigningKey CommitteeHotExtendedKey)

data ByronOrShelleyWitness

Some kind of Byron or Shelley witness.

Constructors

AByronWitness !ShelleyBootstrapWitnessSigningKeyData

AShelleyKeyWitness !ShelleyWitnessSigningKey

data ShelleyBootstrapWitnessSigningKeyData

Data required for constructing a Shelley bootstrap witness.

Constructors

ShelleyBootstrapWitnessSigningKeyData

Fields

!(SigningKey ByronKey)

Byron signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

data CddlWitnessError

Constructors

CddlWitnessErrorTextEnv (FileError TextEnvelopeError) (FileError TextEnvelopeCddlError)

CddlWitnessIOError (FileError TextEnvelopeError)

readFileTxKeyWitness :: FilePath -> IO (Either CddlWitnessError (InAnyShelleyBasedEra KeyWitness))

readWitnessSigningData :: WitnessSigningData -> IO (Either ReadWitnessSigningDataError SomeSigningWitness)

txWitnessTextEnvelopeTypes :: [Text]

Required signer

data RequiredSignerError

Constructors

RequiredSignerErrorFile (FileError InputDecodeError)

RequiredSignerErrorByronKey (SigningKeyFile 'In)

categoriseSomeSigningWitness :: SomeSigningWitness -> ByronOrShelleyWitness

readRequiredSigner :: RequiredSigner -> IO (Either RequiredSignerError (Hash PaymentKey))

Governance related

data ConstitutionError

Constructors

ConstitutionErrorFile (FileError TextEnvelopeError)

ConstitutionNotSupportedInEra AnyCardanoEra

ConstitutionNotUnicodeError UnicodeException

data ProposalError

Constructors

ProposalErrorFile (FileError TextEnvelopeError)

ProposalNotSupportedInEra AnyCardanoEra

ProposalNotUnicodeError UnicodeException

ProposalErrorScriptWitness ScriptWitnessError

data VoteError

Constructors

VoteErrorFile (FileError TextEnvelopeError)

VoteErrorTextNotUnicode UnicodeException

VoteErrorScriptWitness ScriptWitnessError

readTxGovernanceActions :: ShelleyBasedEra era -> [(ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))] -> IO (Either ProposalError [(Proposal era, Maybe (ScriptWitness WitCtxStake era))])

constitutionHashSourceToHash :: ConstitutionHashSource -> ExceptT ConstitutionError IO (SafeHash StandardCrypto AnchorData)

readProposal :: ConwayEraOnwards era -> (ProposalFile 'In, Maybe (ScriptWitnessFiles WitCtxStake)) -> IO (Either ProposalError (Proposal era, Maybe (ScriptWitness WitCtxStake era)))

data CostModelsError

Constructors

CostModelsErrorReadFile (FileError ())

CostModelsErrorJSONDecode FilePath String

CostModelsErrorEmpty FilePath

readCostModels :: File CostModels 'In -> ExceptT CostModelsError IO CostModels

FileOrPipe

data FileOrPipe

We need a type for handling files that may be actually be things like pipes. Currently the CLI makes no guarantee that a "file" will only be read once. This is a problem for a user who who expects to be able to pass a pipe. To handle this, we have a type for representing either files or pipes where the contents will be saved in memory if what we're reading is a pipe (so it can be re-read later). Unfortunately this means we can't easily stream data from pipes, but at present that's not an issue.

fileOrPipe :: FilePath -> IO FileOrPipe

fileOrPipePath :: FileOrPipe -> FilePath

Get the path backing a FileOrPipe. This should primarily be used when generating error messages for a user. A user should not call directly call a function like readFile on the result of this function

fileOrPipeCache :: FileOrPipe -> IO (Maybe ByteString)

readFileOrPipe :: FileOrPipe -> IO ByteString

Get the contents of a file or pipe. This function reads the entire contents of the file or pipe, and is blocking.

Stake credentials

getStakeCredentialFromVerifier :: StakeVerifier -> ExceptT StakeCredentialError IO StakeCredential

getStakeCredentialFromIdentifier :: StakeIdentifier -> ExceptT StakeCredentialError IO StakeCredential

getStakeAddressFromVerifier :: NetworkId -> StakeVerifier -> ExceptT StakeCredentialError IO StakeAddress

readVotingProceduresFiles :: ConwayEraOnwards era -> [(VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake))] -> IO (Either VoteError [(VotingProcedures era, Maybe (ScriptWitness WitCtxStake era))])

readSingleVote :: ConwayEraOnwards era -> (VoteFile 'In, Maybe (ScriptWitnessFiles WitCtxStake)) -> IO (Either VoteError (VotingProcedures era, Maybe (ScriptWitness WitCtxStake era)))

DRep credentials

getDRepCredentialFromVerKeyHashOrFile :: forall t (m :: Type -> Type). MonadIOTransError (FileError InputDecodeError) t m => VerificationKeyOrHashOrFile DRepKey -> t m (Credential 'DRepRole StandardCrypto)

data ReadSafeHashError

Constructors

ReadSafeHashErrorNotHex ByteString String

ReadSafeHashErrorInvalidHash Text

readHexAsSafeHash :: Text -> Either ReadSafeHashError (SafeHash StandardCrypto AnchorData)

readSafeHash :: ReadM (SafeHash StandardCrypto AnchorData)

scriptHashReader :: ReadM ScriptHash

Update proposals

readTxUpdateProposal :: ShelleyToBabbageEra era -> UpdateProposalFile -> ExceptT (FileError TextEnvelopeError) IO (TxUpdateProposal era)

Vote related

readVoteDelegationTarget :: VoteDelegationTarget -> ExceptT DelegationError IO (DRep StandardCrypto)

readVerificationKeyOrHashOrFileOrScript :: forall t (m :: Type -> Type) keyrole (kr :: KeyRole). (MonadIOTransError (Either (FileError ScriptDecodeError) (FileError InputDecodeError)) t m, Key keyrole) => AsType keyrole -> (Hash keyrole -> KeyHash kr StandardCrypto) -> VerificationKeyOrHashOrFileOrScript keyrole -> t m (Credential kr StandardCrypto)

readVerificationKeySource :: forall t (m :: Type -> Type) keyrole (kr :: KeyRole). (MonadIOTransError (Either (FileError ScriptDecodeError) (FileError InputDecodeError)) t m, Key keyrole) => AsType keyrole -> (Hash keyrole -> KeyHash kr StandardCrypto) -> VerificationKeySource keyrole -> t m (Credential kr StandardCrypto)

Genesis hashes

readShelleyOnwardsGenesisAndHash

Arguments

:: MonadIO m

=> FilePath
The file to read

-> m (Hash Blake2b_256 ByteString)