Cardano.CLI.Read

Metadata

readFileTxMetadata :: TxMetadataJsonSchema -> MetadataFile -> CIO e TxMetadata

readTxMetadata :: Era era -> TxMetadataJsonSchema -> [MetadataFile] -> CIO e (TxMetadataInEra era)

Script

data ScriptDecodeError

Constructors

ScriptDecodeTextEnvelopeError TextEnvelopeError

ScriptDecodeSimpleScriptError JsonDecodeError

readAnyScript :: (MonadIO m, IsEra era) => FilePath -> m (AnyScript (LedgerEra era))

deserialiseScriptInAnyLang :: ByteString -> Either ScriptDecodeError ScriptInAnyLang

readFileScriptInAnyLang :: MonadIO m => FilePath -> m ScriptInAnyLang

Read a script file. The file can either be in the text envelope format wrapping the binary representation of any of the supported script languages, or alternatively it can be a JSON format file for one of the simple script language versions.

data PlutusScriptDecodeError

Constructors

PlutusScriptDecodeErrorUnknownVersion !Text

PlutusScriptJsonDecodeError !JsonDecodeError

PlutusScriptDecodeTextEnvelopeError !TextEnvelopeError

PlutusScriptDecodeErrorVersionMismatch

Fields

!Text

Script version

!AnyPlutusScriptVersion

Attempted to decode with version

Script data (datums and redeemers)

data ScriptDataError

Constructors

ScriptDataErrorFile (FileError ())

ScriptDataErrorJsonParse !FilePath !String

ScriptDataErrorConversion !FilePath !ScriptDataJsonError

ScriptDataErrorValidation !FilePath !ScriptDataRangeError

ScriptDataErrorMetadataDecode !FilePath !DecoderError

ScriptDataErrorJsonBytes !ScriptDataJsonBytesError

renderScriptDataError :: ScriptDataError -> Doc ann

Tx

newtype IncompleteTxBody

Constructors

IncompleteTxBody

Fields

unIncompleteTxBody :: InAnyShelleyBasedEra TxBody

readFileTx :: FileOrPipe -> IO (Either (FileError TextEnvelopeError) (InAnyShelleyBasedEra Tx))

readFileTxBody :: FileOrPipe -> IO (Either (FileError TextEnvelopeError) IncompleteTxBody)

readTx :: FileOrPipe -> IO (Either (FileError TextEnvelopeError) (InAnyShelleyBasedEra Tx))

Tx witnesses

data ReadWitnessSigningDataError

Constructors

ReadWitnessSigningDataSigningKeyDecodeError !(FileError InputDecodeError)

ReadWitnessSigningDataScriptError !(FileError JsonDecodeError)

ReadWitnessSigningDataSigningKeyAndAddressMismatch
A Byron address was specified alongside a non-Byron signing key.

renderReadWitnessSigningDataError :: ReadWitnessSigningDataError -> Doc ann

Render an error message for a ReadWitnessSigningDataError.

data SomeSigningWitness

Constructors

AByronSigningWitness (SigningKey ByronKey) (Maybe (Address ByronAddr))

APaymentSigningWitness (SigningKey PaymentKey)

APaymentExtendedSigningWitness (SigningKey PaymentExtendedKey)

AStakeSigningWitness (SigningKey StakeKey)

AStakeExtendedSigningWitness (SigningKey StakeExtendedKey)

AStakePoolSigningWitness (SigningKey StakePoolKey)

AStakePoolExtendedSigningWitness (SigningKey StakePoolExtendedKey)

AGenesisSigningWitness (SigningKey GenesisKey)

AGenesisExtendedSigningWitness (SigningKey GenesisExtendedKey)

AGenesisDelegateSigningWitness (SigningKey GenesisDelegateKey)

AGenesisDelegateExtendedSigningWitness (SigningKey GenesisDelegateExtendedKey)

AGenesisUTxOSigningWitness (SigningKey GenesisUTxOKey)

ADRepSigningWitness (SigningKey DRepKey)

ADRepExtendedSigningWitness (SigningKey DRepExtendedKey)

ACommitteeColdSigningWitness (SigningKey CommitteeColdKey)

ACommitteeColdExtendedSigningWitness (SigningKey CommitteeColdExtendedKey)

ACommitteeHotSigningWitness (SigningKey CommitteeHotKey)

ACommitteeHotExtendedSigningWitness (SigningKey CommitteeHotExtendedKey)

data ByronOrShelleyWitness

Some kind of Byron or Shelley witness.

Constructors

AByronWitness !ShelleyBootstrapWitnessSigningKeyData

AShelleyKeyWitness !ShelleyWitnessSigningKey

mkShelleyBootstrapWitness :: ShelleyBasedEra era -> Maybe NetworkId -> TxBody (ShelleyLedgerEra era) -> ShelleyBootstrapWitnessSigningKeyData -> Either BootstrapWitnessError (KeyWitness era)

Construct a Shelley bootstrap witness (i.e. a Byron key witness in the Shelley era).

data ShelleyBootstrapWitnessSigningKeyData

Data required for constructing a Shelley bootstrap witness.

Constructors

ShelleyBootstrapWitnessSigningKeyData

Fields

!(SigningKey ByronKey)

Byron signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

readFileTxKeyWitness :: FilePath -> IO (Either (FileError TextEnvelopeError) (InAnyShelleyBasedEra KeyWitness))

readWitnessSigningData :: WitnessSigningData -> IO (Either ReadWitnessSigningDataError SomeSigningWitness)

txWitnessTextEnvelopeTypes :: [Text]

Required signer

data RequiredSignerError

Constructors

RequiredSignerErrorFile (FileError InputDecodeError)

RequiredSignerErrorByronKey (SigningKeyFile 'In)

categoriseSomeSigningWitness :: SomeSigningWitness -> ByronOrShelleyWitness

readRequiredSigner :: RequiredSigner -> IO (Either RequiredSignerError (Hash PaymentKey))

Governance related

newtype ConstitutionError

Constructors

ConstitutionNotUnicodeError UnicodeException

newtype VoteError

Constructors

VoteErrorFile (FileError CliScriptWitnessError)

data CostModelsError

Constructors

CostModelsErrorReadFile (FileError ())

CostModelsErrorJSONDecode FilePath String

CostModelsErrorEmpty FilePath

readCostModels :: File CostModels 'In -> ExceptT CostModelsError IO CostModels

FileOrPipe

data FileOrPipe

We need a type for handling files that may be actually be things like pipes. Currently the CLI makes no guarantee that a "file" will only be read once. This is a problem for a user who who expects to be able to pass a pipe. To handle this, we have a type for representing either files or pipes where the contents will be saved in memory if what we're reading is a pipe (so it can be re-read later). Unfortunately this means we can't easily stream data from pipes, but at present that's not an issue.

fileOrPipe :: FilePath -> IO FileOrPipe

fileOrPipePath :: FileOrPipe -> FilePath

Get the path backing a FileOrPipe. This should primarily be used when generating error messages for a user. A user should not call directly call a function like readFile on the result of this function

fileOrPipeCache :: FileOrPipe -> IO (Maybe ByteString)

readFileOrPipe :: FileOrPipe -> IO ByteString

Get the contents of a file or pipe. This function reads the entire contents of the file or pipe, and is blocking.

readFileOrPipeTextEnvelopeAnyOf :: [FromSomeType HasTextEnvelope b] -> FileOrPipe -> IO (Either (FileError TextEnvelopeError) b)

Stake credentials

getStakeCredentialFromVerifier :: StakeVerifier -> CIO e StakeCredential

getStakeCredentialFromIdentifier :: StakeIdentifier -> CIO e StakeCredential

getStakeAddressFromVerifier :: NetworkId -> StakeVerifier -> CIO e StakeAddress

Stake pool credentials

getHashFromStakePoolKeyHashSource :: MonadIO m => StakePoolKeyHashSource -> m (Hash StakePoolKey)

Get the hash from a stake pool key hash source

getVerificationKeyFromStakePoolVerificationKeySource :: MonadIO m => StakePoolVerificationKeySource -> m AnyStakePoolVerificationKey

Get the verification key from a stake pool verification key source

DRep credentials

data ReadSafeHashError

Constructors

ReadSafeHashErrorNotHex ByteString String

ReadSafeHashErrorInvalidHash Text

readHexAsSafeHash :: Text -> Either ReadSafeHashError (SafeHash AnchorData)

readSafeHash :: ReadM (SafeHash AnchorData)

scriptHashReader :: ReadM ScriptHash

Update proposals

readTxUpdateProposal :: ShelleyToBabbageEra era -> UpdateProposalFile -> ExceptT (FileError TextEnvelopeError) IO (TxUpdateProposal era)

Vote related

readVoteDelegationTarget :: VoteDelegationTarget -> CIO e DRep

readVerificationKeySource :: forall keyrole (kr :: KeyRole) e. Key keyrole => (Hash keyrole -> KeyHash kr) -> VerificationKeySource keyrole -> CIO e (Credential kr)

Genesis hashes

readShelleyOnwardsGenesisAndHash

Arguments

:: MonadIO m

=> FilePath
The file to read

-> m (Hash Blake2b_256 ByteString)

readFileCli :: (HasCallStack, MonadIO m) => FilePath -> m ByteString

Plutus

readFilePlutusScript :: forall e era. IsEra era => FilePath -> CIO e (AnyPlutusScript (LedgerEra era))

utilities

readerFromParsecParser :: Parser a -> ReadM a