Cardano.CLI.Shelley.Key

Description

Shelley CLI option data types and functions for cryptographic keys.

Documentation

data VerificationKeyOrFile keyrole

Either a verification key or path to a verification key file.

Constructors

VerificationKeyValue !(VerificationKey keyrole)
A verification key.

VerificationKeyFilePath !VerificationKeyFile
A path to a verification key file. Note that this file hasn't been validated at all (whether it exists, contains a key of the correct type, etc.)

readVerificationKeyOrFile ∷ (HasTextEnvelope (VerificationKey keyrole), SerialiseAsBech32 (VerificationKey keyrole)) ⇒ AsType keyrole → VerificationKeyOrFile keyrole → IO (Either (FileError InputDecodeError) (VerificationKey keyrole))

Read a verification key or verification key file and return a verification key.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrTextEnvFile ∷ HasTextEnvelope (VerificationKey keyrole) ⇒ AsType keyrole → VerificationKeyOrFile keyrole → IO (Either (FileError InputDecodeError) (VerificationKey keyrole))

Read a verification key or verification key file and return a verification key.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data VerificationKeyTextOrFile

Either an unvalidated text representation of a verification key or a path to a verification key file.

Constructors

VktofVerificationKeyText !Text

VktofVerificationKeyFile !VerificationKeyFile

data VerificationKeyTextOrFileError

An error in deserialising a VerificationKeyTextOrFile to a VerificationKey.

Constructors

VerificationKeyTextError !InputDecodeError

VerificationKeyFileError !(FileError InputDecodeError)

readVerificationKeyTextOrFileAnyOf ∷ VerificationKeyTextOrFile → IO (Either VerificationKeyTextOrFileError SomeAddressVerificationKey)

Deserialise a verification key from text or a verification key file. If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

renderVerificationKeyTextOrFileError ∷ VerificationKeyTextOrFileError → Text

Render an error message for a VerificationKeyTextOrFileError.

data VerificationKeyOrHashOrFile keyrole

Verification key, verification key hash, or path to a verification key file.

Constructors

VerificationKeyOrFile !(VerificationKeyOrFile keyrole)
Either a verification key or path to a verification key file.

VerificationKeyHash !(Hash keyrole)
A verification key hash.

readVerificationKeyOrHashOrFile ∷ (Key keyrole, SerialiseAsBech32 (VerificationKey keyrole)) ⇒ AsType keyrole → VerificationKeyOrHashOrFile keyrole → IO (Either (FileError InputDecodeError) (Hash keyrole))

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, the file can either be formatted as Bech32, hex, or text envelope.

readVerificationKeyOrHashOrTextEnvFile ∷ Key keyrole ⇒ AsType keyrole → VerificationKeyOrHashOrFile keyrole → IO (Either (FileError InputDecodeError) (Hash keyrole))

Read a verification key or verification key hash or verification key file and return a verification key hash.

If a filepath is provided, it will be interpreted as a text envelope formatted file.

data PaymentVerifier

Constructors

PaymentVerifierKey VerificationKeyTextOrFile

PaymentVerifierScriptFile ScriptFile

data StakeVerifier

Constructors

StakeVerifierKey (VerificationKeyOrFile StakeKey)

StakeVerifierScriptFile ScriptFile

StakeVerifierAddress StakeAddress

generateKeyPair ∷ Key keyrole ⇒ AsType keyrole → IO (VerificationKey keyrole, SigningKey keyrole)