Cardano.CLI.Types

Documentation

data BalanceTxExecUnits

A flag that differentiates between automatically and manually balancing a tx.

Constructors

AutoBalance

ManualBalance

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile ∷ FilePath

newtype CurrentKesPeriod

Constructors

CurrentKesPeriod

Fields

unCurrentKesPeriod ∷ Word64

data EpochLeadershipSchedule

Which leadership schedule we are interested in. TODO: Implement Previous and Next epochs

Constructors

CurrentEpoch

NextEpoch

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile ∷ FilePath

newtype OpCertEndingKesPeriod

Constructors

OpCertEndingKesPeriod

Fields

unOpCertEndingKesPeriod ∷ Word64

data OpCertIntervalInformation

Constructors

OpCertWithinInterval OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod SlotsTillKesKeyExpiry

OpCertStartingKesPeriodIsInTheFuture OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertExpired OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertSomeOtherError OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod
Shouldn't be possible

newtype OpCertOnDiskCounter

Constructors

OpCertOnDiskCounter

Fields

unOpCertOnDiskCounter ∷ Word64

data OpCertNodeAndOnDiskCounterInformation

Constructors

OpCertOnDiskCounterEqualToNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is equal to its corresponding counter in the node state. The on disk operational certificate therefore has a valid counter.

OpCertOnDiskCounterAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by 1. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterTooFarAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is less than the counter in the node state. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterBehindNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The corresponding counter for operational certificate was not found in the node state. This means the relevant stake pool has not minted a block yet. When the stake pool has minted a block the corresponding operational certificate's counter will be present in the node state.

OpCertNoBlocksMintedYet OpCertOnDiskCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by more than 1. The on disk operational certificate is invalid in this case.

newtype OpCertNodeStateCounter

Constructors

OpCertNodeStateCounter

Fields

unOpCertNodeStateCounter ∷ Word64

newtype OpCertStartingKesPeriod

Constructors

OpCertStartingKesPeriod

Fields

unOpCertStartingKesPeriod ∷ Word64

data OutputFormat

The desired output format.

Constructors

OutputFormatHex

OutputFormatBech32

data TxBuildOutputOptions

Specify whether to render the script cost as JSON in the cli's build command.

Constructors

OutputScriptCostOnly FilePath

OutputTxBodyOnly TxBodyFile

data ReferenceScriptAnyEra

Constructors

ReferenceScriptAnyEraNone

ReferenceScriptAnyEra FilePath

newtype SigningKeyFile

Constructors

SigningKeyFile

Fields

unSigningKeyFile ∷ FilePath

newtype ScriptFile

Constructors

ScriptFile

Fields

unScriptFile ∷ FilePath

data ScriptDataOrFile

Constructors

ScriptDataCborFile FilePath
By reference to a CBOR file

ScriptDataJsonFile FilePath
By reference to a JSON file

ScriptDataValue HashableScriptData
By value

type ScriptRedeemerOrFile = ScriptDataOrFile

data ScriptWitnessFiles witctx where

This type is like ScriptWitness, but the file paths from which to load the script witness data representation.

It is era-independent, but witness context-dependent.

Constructors

SimpleScriptWitnessFile ∷ ScriptFile → ScriptWitnessFiles witctx

PlutusScriptWitnessFiles ∷ ScriptFile → ScriptDatumOrFile witctx → ScriptRedeemerOrFile → ExecutionUnits → ScriptWitnessFiles witctx

PlutusReferenceScriptWitnessFiles

Fields

∷ TxIn

→ AnyScriptLanguage

→ ScriptDatumOrFile witctx

→ ScriptRedeemerOrFile

→ ExecutionUnits

→ Maybe PolicyId

For minting reference scripts

→ ScriptWitnessFiles witctx

SimpleReferenceScriptWitnessFiles

Fields

∷ TxIn

→ AnyScriptLanguage

→ Maybe PolicyId

For minting reference scripts

→ ScriptWitnessFiles witctx

data ScriptDatumOrFile witctx where

Constructors

ScriptDatumOrFileForTxIn ∷ ScriptDataOrFile → ScriptDatumOrFile WitCtxTxIn

InlineDatumPresentAtTxIn ∷ ScriptDatumOrFile WitCtxTxIn

NoScriptDatumOrFileForMint ∷ ScriptDatumOrFile WitCtxMint

NoScriptDatumOrFileForStake ∷ ScriptDatumOrFile WitCtxStake

newtype SlotsTillKesKeyExpiry

Constructors

SlotsTillKesKeyExpiry

Fields

unSlotsTillKesKeyExpiry ∷ SlotNo

data TransferDirection

Determines the direction in which the MIR certificate will transfer ADA.

Constructors

TransferToReserves

TransferToTreasury

newtype TxBodyFile

Constructors

TxBodyFile FilePath

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value TxOutDatumAnyEra ReferenceScriptAnyEra

newtype TxOutChangeAddress

A partially-specified transaction output indented to use as a change output.

It does not specify a value, since this will be worked out automatically.

It does not use any script data hash, since that's generally not used for change outputs.

Constructors

TxOutChangeAddress AddressAny

data TxOutDatumAnyEra

Constructors

TxOutDatumByHashOnly (Hash ScriptData)

TxOutDatumByHashOf ScriptDataOrFile

TxOutDatumByValue ScriptDataOrFile

TxOutInlineDatumByValue ScriptDataOrFile

TxOutDatumByNone

newtype TxFile

Constructors

TxFile FilePath

data TxMempoolQuery

Constructors

TxMempoolQueryTxExists TxId

TxMempoolQueryNextTx

TxMempoolQueryInfo

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile ∷ FilePath

newtype VerificationKeyFile

Constructors

VerificationKeyFile

Fields

unVerificationKeyFile ∷ FilePath

data Params crypto

This data structure is used to allow nicely formatted output in the query pool-params command. params are the current pool parameter settings, futureparams are new parameters, retiringEpoch is the epoch that has been set for pool retirement. Any of these may be Nothing.

Constructors

Params

Fields

poolParameters ∷ Maybe (PoolParams crypto)

futurePoolParameters ∷ Maybe (PoolParams crypto)

retiringEpoch ∷ Maybe EpochNo

data RequiredSigner

Plutus script required signers

Constructors

RequiredSignerSkeyFile SigningKeyFile

RequiredSignerHash (Hash PaymentKey)

data AllOrOnly a

Constructors

All

Only a