Cardano.CLI.Types

Documentation

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile ∷ FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile ∷ FilePath

data OutputFormat

The desired output format.

Constructors

OutputFormatHex

OutputFormatBech32

data QueryFilter

UTxO query filtering options.

Constructors

FilterByAddress !(Set AddressAny)

NoFilter

newtype SigningKeyFile

Constructors

SigningKeyFile

Fields

unSigningKeyFile ∷ FilePath

data SigningKeyOrScriptFile

Constructors

ScriptFileForWitness FilePath

SigningKeyFileForWitness FilePath

newtype SocketPath

Constructors

SocketPath

Fields

unSocketPath ∷ FilePath

newtype ScriptFile

Constructors

ScriptFile

Fields

unScriptFile ∷ FilePath

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile ∷ FilePath

newtype VerificationKeyFile

Constructors

VerificationKeyFile

Fields

unVerificationKeyFile ∷ FilePath