Cardano.CLI.Types.Common

Documentation

data AllOrOnly a

Constructors

All

Only [a]

data AddressKeyType

Constructors

AddressKeyShelley

AddressKeyShelleyExtended

AddressKeyByron

data AnchorScheme

The different schemes that can be used to fetch anchor data.

Constructors

FileScheme

HttpScheme

HttpsScheme

IpfsScheme

data AnyPlutusScriptVersion where

Constructors

AnyPlutusScriptVersion :: forall lang. IsPlutusScriptLanguage lang => PlutusScriptVersion lang -> AnyPlutusScriptVersion

data BalanceTxExecUnits

A flag that differentiates between automatically and manually balancing a tx.

Constructors

AutoBalance

ManualBalance

newtype BlockId

Constructors

BlockId String

data ByronKeyFormat

Constructors

NonLegacyByronKeyFormat

LegacyByronKeyFormat

data ByronKeyType

Constructors

ByronPaymentKey ByronKeyFormat

ByronGenesisKey ByronKeyFormat

ByronDelegateKey ByronKeyFormat

data CardanoAddressKeyType

The type of cardano-address key.

Constructors

CardanoAddressShelleyPaymentKey

CardanoAddressShelleyStakeKey

CardanoAddressIcarusPaymentKey

CardanoAddressByronPaymentKey

CardanoAddressCommitteeColdKey

CardanoAddressCommitteeHotKey

CardanoAddressDRepKey

data CBORObject

Specify what the CBOR file is i.e a block, a tx, etc

Constructors

CBORBlockByron EpochSlots

CBORDelegationCertificateByron

CBORTxByron

CBORUpdateProposalByron

CBORVoteByron

newtype CertificateFile

Constructors

CertificateFile

Fields

unCertificateFile :: FilePath

data ConstitutionHashSource

Constructors

ConstitutionHashSourceFile (File ConstitutionText 'In)

ConstitutionHashSourceText Text

ConstitutionHashSourceHash (SafeHash StandardCrypto AnchorData)

newtype ConstitutionText

Constructors

ConstitutionText

Fields

unConstitutionText :: Text

newtype ConstitutionUrl

Constructors

ConstitutionUrl

Fields

unConstitutionUrl :: Url

data CredentialGenerationMode

Constructors

OnDisk
Write credentials to disk

Transient
Don't write them to disk (process them in memory)

newtype CurrentKesPeriod

Constructors

CurrentKesPeriod

Fields

unCurrentKesPeriod :: Word64

data DRepCredentials

Constructors

DRepCredentials

Fields

dRepCredentialGenerationMode :: !CredentialGenerationMode

Whether to write them to disk

numOfDRepCredentials :: !Word

The number of DRep credentials to generate

data EpochLeadershipSchedule

Which leadership schedule we are interested in. TODO: Implement Previous and Next epochs

Constructors

CurrentEpoch

NextEpoch

newtype File content (direction :: FileDirection)

Constructors

File

Fields

unFile :: FilePath

data FileDirection

Constructors

In

Out

InOut

newtype GenesisDir

Constructors

GenesisDir FilePath

newtype GenesisFile

Constructors

GenesisFile

Fields

unGenesisFile :: FilePath

newtype GenesisKeyFile

Constructors

GenesisKeyFile FilePath

data IncludeStake

Whether to include the stake, as queried by drep-stake-distribution, in the output of drep-state. This is (computationally) expensive, but sometimes convenient.

Constructors

WithStake

NoStake

data InputTxBodyOrTxFile

Constructors

InputTxBodyFile (TxBodyFile 'In)

InputTxFile (TxFile 'In)

data KeyOutputFormat

Constructors

KeyOutputFormatTextEnvelope

KeyOutputFormatBech32

data MetadataFile

Constructors

MetadataFileJSON (File () 'In)

MetadataFileCBOR (File () 'In)

data MustCheckHash a

Constructors

CheckHash

TrustHash

data OpCertCounter

type OpCertCounterFile = File OpCertCounter

newtype OpCertEndingKesPeriod

Constructors

OpCertEndingKesPeriod

Fields

unOpCertEndingKesPeriod :: Word64

data OpCertIntervalInformation

Constructors

OpCertWithinInterval OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod SlotsTillKesKeyExpiry

OpCertStartingKesPeriodIsInTheFuture OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertExpired OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod

OpCertSomeOtherError OpCertStartingKesPeriod OpCertEndingKesPeriod CurrentKesPeriod
Shouldn't be possible

data OpCertNodeAndOnDiskCounterInformation

Constructors

OpCertOnDiskCounterEqualToNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is equal to its corresponding counter in the node state. The on disk operational certificate therefore has a valid counter.

OpCertOnDiskCounterAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by 1. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterTooFarAheadOfNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The on disk operational certificate has a counter that is less than the counter in the node state. The on disk operational certificate is invalid in this case.

OpCertOnDiskCounterBehindNodeState OpCertOnDiskCounter OpCertNodeStateCounter
The corresponding counter for operational certificate was not found in the node state. This means the relevant stake pool has not minted a block yet. When the stake pool has minted a block the corresponding operational certificate's counter will be present in the node state.

OpCertNoBlocksMintedYet OpCertOnDiskCounter
The on disk operational certificate has a counter that is ahead of the counter in the node state by more than 1. The on disk operational certificate is invalid in this case.

newtype OpCertNodeStateCounter

Constructors

OpCertNodeStateCounter

Fields

unOpCertNodeStateCounter :: Word64

newtype OpCertOnDiskCounter

Constructors

OpCertOnDiskCounter

Fields

unOpCertOnDiskCounter :: Word64

newtype OpCertStartingKesPeriod

Constructors

OpCertStartingKesPeriod

Fields

unOpCertStartingKesPeriod :: Word64

data Params crypto

This data structure is used to allow nicely formatted output in the query pool-params command. params are the current pool parameter settings, futureparams are new parameters, retiringEpoch is the epoch that has been set for pool retirement. Any of these may be Nothing.

Constructors

Params

Fields

poolParameters :: Maybe (PoolParams crypto)

futurePoolParameters :: Maybe (PoolParams crypto)

retiringEpoch :: Maybe EpochNo

data ParserFileDirection

Constructors

Input

Output

data IdOutputFormat

Some entities such as stake pools and dreps have a notion of an ID and that id can be encoded as either a bech32 or hex string. This type is used to specify which encoding to use.

Constructors

IdOutputFormatHex

IdOutputFormatBech32

newtype PrivKeyFile

Constructors

PrivKeyFile FilePath

data ProposalBinary

Tag for tracking proposals submitted as Bytestring

type ProposalFile = File ()

data ProposalText

Tag for tracking proposals submitted as Text

newtype ProposalUrl

Constructors

ProposalUrl

Fields

unProposalUrl :: Url

newtype ProtocolParamsFile

Constructors

ProtocolParamsFile FilePath

data OutputFormatJsonOrText

Constructors

OutputFormatJson

OutputFormatText

data ReferenceScriptAnyEra

Constructors

ReferenceScriptAnyEraNone

ReferenceScriptAnyEra FilePath

newtype ReferenceScriptSize

Constructors

ReferenceScriptSize

Fields

unReferenceScriptSize :: Int

data RequiredSigner

Plutus script required signers

Constructors

RequiredSignerSkeyFile (SigningKeyFile 'In)

RequiredSignerHash (Hash PaymentKey)

data ScriptDataOrFile

Constructors

ScriptDataCborFile FilePath
By reference to a CBOR file

ScriptDataJsonFile FilePath
By reference to a JSON file

ScriptDataValue HashableScriptData
By value

data ScriptDatumOrFile witctx where

Constructors

ScriptDatumOrFileForTxIn :: Maybe ScriptDataOrFile -> ScriptDatumOrFile WitCtxTxIn

InlineDatumPresentAtTxIn :: ScriptDatumOrFile WitCtxTxIn

NoScriptDatumOrFileForMint :: ScriptDatumOrFile WitCtxMint

NoScriptDatumOrFileForStake :: ScriptDatumOrFile WitCtxStake

type ScriptFile = File ScriptInAnyLang 'In

type ScriptRedeemerOrFile = ScriptDataOrFile

data ScriptWitnessFiles witctx where

This type is like ScriptWitness, but the file paths from which to load the script witness data representation.

It is era-independent, but witness context-dependent. NB: This is in the process of being deprecated because it is difficult to accomodate for changes for specific plutus script purposes. As an example when minting a multi-asset with a plutus script we need the policy id of the said script. This is fine when we have access to the plutus script however in the case of a reference script we demand the user provides the policy id. Enshrining that change in the ScriptWitnessFiles is difficult because only minting scripts require this but not the other kinds of plutus scripts (spending, certifying etc.) Another example is CIP-69 where datums are no longer required for spending scripts. This is further complicated by the fact at the parsing level we make user facing simplifications e.g `--mint-script-file` which says nothing about the script type (simple vs plutus) or script version. As a result need to separate the different script purposes into their own separate data definitions where we can make changes specific to that script purpose more easily without affecting the rest of the api.

Constructors

SimpleScriptWitnessFile :: forall witctx. ScriptFile -> ScriptWitnessFiles witctx

PlutusScriptWitnessFiles :: forall witctx. ScriptFile -> ScriptDatumOrFile witctx -> ScriptRedeemerOrFile -> ExecutionUnits -> ScriptWitnessFiles witctx

PlutusReferenceScriptWitnessFiles

Fields

:: forall witctx. TxIn

-> AnyPlutusScriptVersion

-> ScriptDatumOrFile witctx

-> ScriptRedeemerOrFile

-> ExecutionUnits

For minting reference scripts

-> ScriptWitnessFiles witctx

SimpleReferenceScriptWitnessFiles :: forall witctx. TxIn -> AnyScriptLanguage -> ScriptWitnessFiles witctx

type SigningKeyFile = File (SigningKey ())

newtype SlotsTillKesKeyExpiry

Constructors

SlotsTillKesKeyExpiry

Fields

unSlotsTillKesKeyExpiry :: SlotNo

data SomeKeyFile (direction :: FileDirection)

Either a verification or signing key, used for conversions and other commands that make sense for both.

Constructors

AVerificationKeyFile (VerificationKeyFile direction)

ASigningKeyFile (SigningKeyFile direction)

data StakeDelegators

Constructors

StakeDelegators

Fields

stakeDelegatorsGenerationMode :: !CredentialGenerationMode

Whether to write them to disk

numOfStakeDelegators :: !Word

The number of stake credentials to generate

type StakePoolMetadataFile = File StakePoolMetadata

type SupportedSchemes = [AnchorScheme]

Specifies the schemes that are allowed to fetch anchor data.

data TransferDirection

Determines the direction in which the MIR certificate will transfer ADA.

Constructors

TransferToReserves

TransferToTreasury

type TxBodyFile = File (TxBody ())

data TxBuildOutputOptions

Specify whether to render the script cost as JSON in the cli's build command.

Constructors

OutputScriptCostOnly (File () 'Out)

OutputTxBodyOnly (TxBodyFile 'Out)

newtype TxByronWitnessCount

Constructors

TxByronWitnessCount Int

type TxFile = File (Tx ())

newtype TxSubmissionResult

Type used for serialization when printing the hash of a transaction after having submitted it.

Constructors

TxSubmissionResult

Fields

txhash :: TxId

newtype TxTreasuryDonation

Constructors

TxTreasuryDonation

Fields

unTxTreasuryDonation :: Lovelace

newtype TxInCount

Constructors

TxInCount Int

data TxMempoolQuery

Constructors

TxMempoolQueryTxExists TxId

TxMempoolQueryNextTx

TxMempoolQueryInfo

data TxOutAnyEra

A TxOut value that is the superset of possibilities for any era: any address type and allowing multi-asset values. This is used as the type for values passed on the command line. It can be converted into the era-dependent TxOutValue type.

Constructors

TxOutAnyEra AddressAny Value TxOutDatumAnyEra ReferenceScriptAnyEra

data TxOutShelleyBasedEra

Constructors

TxOutShelleyBasedEra !(Address ShelleyAddr) Value TxOutDatumAnyEra ReferenceScriptAnyEra

newtype TxOutChangeAddress

A partially-specified transaction output indented to use as a change output.

It does not specify a value, since this will be worked out automatically.

It does not use any script data hash, since that's generally not used for change outputs.

Constructors

TxOutChangeAddress AddressAny

newtype TxOutCount

Constructors

TxOutCount Int

data TxOutDatumAnyEra

Constructors

TxOutDatumByHashOnly (Hash ScriptData)

TxOutDatumByHashOf ScriptDataOrFile

TxOutDatumByValue ScriptDataOrFile

TxOutInlineDatumByValue ScriptDataOrFile

TxOutDatumByNone

newtype TxShelleyWitnessCount

Constructors

TxShelleyWitnessCount Int

newtype UpdateProposalFile

Constructors

UpdateProposalFile

Fields

unUpdateProposalFile :: FilePath

newtype VerificationKeyBase64

A raw verification key given in Base64, and decoded into a ByteString.

Constructors

VerificationKeyBase64 String

type VerificationKeyFile = File (VerificationKey ())

data ViewOutputFormat

Constructors

ViewOutputFormatJson

ViewOutputFormatYaml

newtype VoteUrl

Constructors

VoteUrl

Fields

unVoteUrl :: Url

newtype VoteText

Constructors

VoteText

Fields

unVoteText :: Text

data VoteHashSource

Constructors

VoteHashSourceFile (File VoteText 'In)

VoteHashSourceText Text

VoteHashSourceHash (SafeHash StandardCrypto AnchorData)

newtype WitnessFile

Constructors

WitnessFile FilePath

data WitnessSigningData

Data required to construct a witness.

Constructors

KeyWitnessSigningData

Fields

!(SigningKeyFile 'In)

Path to a file that should contain a signing key.

!(Maybe (Address ByronAddr))

An optionally specified Byron address.

If specified, both the network ID and derivation path are extracted from the address and used in the construction of the Byron witness.

type DRepMetadataFile = File DRepMetadata

data DRepMetadataUrl

Tag for differentiating between DRep metadata sources and sources for other types of anchor data

data ResignationMetadataUrl

Tag for differentiating between resignation metadatata sources and sources for other types of anchor data

data PotentiallyCheckedAnchor anchorType anchor

Constructors

PotentiallyCheckedAnchor

Fields

pcaAnchor :: anchor

The anchor data whose hash is to be checked

pcaMustCheck :: MustCheckHash anchorType

Whether to check the hash or not (CheckHash for checking or TrustHash for not checking)