Cardano.CLI.Types.Errors.QueryCmdLocalStateQueryError

Documentation

newtype QueryCmdLocalStateQueryError

An error that can occur while querying a node's local state.

Constructors

EraMismatchError EraMismatch
A query from a certain era was applied to a ledger from a different era.

mkEraMismatchError :: NodeEraMismatchError -> QueryCmdLocalStateQueryError