Cardano.CLI.Types.Errors.TxValidationError

Documentation

data TxAuxScriptsValidationError

Constructors

TxAuxScriptsNotSupportedInEra AnyCardanoEra

TxAuxScriptsLanguageError ScriptLanguageValidationError

newtype TxGovDuplicateVotes era

Constructors

TxGovDuplicateVotes (VotesMergingConflict era)

data TxNotSupportedInEraValidationError era

Constructors

TxNotSupportedInAnyCardanoEraValidationError Text AnyCardanoEra
First argument is the kind of data that is not supported. Second argument is the era that doesn't support the data.

TxNotSupportedInShelleyBasedEraValidationError Text (ShelleyBasedEra era)
First argument is the kind of data that is not supported. Second argument is the Shelley era that doesn't support the data.

validateScriptSupportedInEra :: ShelleyBasedEra era -> ScriptInAnyLang -> Either ScriptLanguageValidationError (ScriptInEra era)

validateTxAuxScripts :: ShelleyBasedEra era -> [ScriptInAnyLang] -> Either TxAuxScriptsValidationError (TxAuxScripts era)

validateRequiredSigners :: ShelleyBasedEra era -> [Hash PaymentKey] -> Either (TxNotSupportedInEraValidationError era) (TxExtraKeyWitnesses era)

validateTxReturnCollateral :: ShelleyBasedEra era -> Maybe (TxOut CtxTx era) -> Either (TxNotSupportedInEraValidationError era) (TxReturnCollateral CtxTx era)

validateTxScriptValidity :: ShelleyBasedEra era -> Maybe ScriptValidity -> Either (TxNotSupportedInEraValidationError era) (TxScriptValidity era)

validateTxTotalCollateral :: ShelleyBasedEra era -> Maybe Lovelace -> Either (TxNotSupportedInEraValidationError era) (TxTotalCollateral era)

validateTxValidityLowerBound :: ShelleyBasedEra era -> Maybe SlotNo -> Either (TxNotSupportedInEraValidationError era) (TxValidityLowerBound era)

validateUpdateProposalFile :: CardanoEra era -> Maybe UpdateProposalFile -> Either (TxNotSupportedInEraValidationError era) (Maybe (Featured ShelleyToBabbageEra era (Maybe UpdateProposalFile)))

validateTxCurrentTreasuryValue :: ShelleyBasedEra era -> Maybe TxCurrentTreasuryValue -> Either (TxNotSupportedInEraValidationError era) (Maybe (Featured ConwayEraOnwards era (Maybe Lovelace)))

validateTxTreasuryDonation :: ShelleyBasedEra era -> Maybe TxTreasuryDonation -> Either (TxNotSupportedInEraValidationError era) (Maybe (Featured ConwayEraOnwards era Lovelace))