Test.Cardano.CLI.Util

Documentation

assertDirectoryMissing :: (MonadTest m, MonadIO m, HasCallStack) => FilePath -> m ()

Asserts that the given directory is missing.

checkTxCddlFormat

Arguments

:: (MonadTest m, MonadIO m, HasCallStack)

=> FilePath
Reference/golden file

-> FilePath
Newly created file

-> m ()

checkTextEnvelopeFormat

Arguments

:: (MonadTest m, MonadIO m, HasCallStack)

=> TextEnvelopeType

-> FilePath
On-disk reference/golden file

-> FilePath
Newly created file

-> m ()

Checks that the tvType and tvDescription are equivalent between two files.

equivalence :: (MonadTest m, Eq a, Show a, HasCallStack) => a -> a -> m ()

Check for equivalence between two types and perform a file cleanup on failure.

execCardanoCLI

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m String
Captured stdout

Execute cardano-cli via the command line.

Waits for the process to finish and returns the stdout.

execCardanoCLIWithEnvVars

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [(String, String)]
Environment variables to set

-> [String]
Arguments to the CLI command

-> m String

Execute cardano-cli via the command line but set environment variables. Fails if the process returns a non-zero exit code.

Waits for the process to finish and returns the stdout.

execDetailCardanoCLI

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> m (ExitCode, String, String)
exit code, stdout, stderr

Execute cardano-cli via the command line, expecting it to fail.

Waits for the process to finish and returns the exit code, stdout and stderr.

execDetailConfigCardanoCLI

Arguments

:: (MonadTest m, MonadCatch m, MonadIO m, HasCallStack)

=> ExecConfig
Configuration for the execution

-> [String]
Arguments to the CLI command

-> m (ExitCode, String, String)
Exit code, stdout, stderr

Execute cardano-cli via the command line, expecting it to fail, and accepting custom config.

Waits for the process to finish and returns the exit code, stdout and stderr.

tryExecCardanoCLI

Arguments

:: forall (m :: Type -> Type). (MonadCatch m, MonadIO m, HasCallStack)

=> [String]
Arguments to the CLI command

-> PropertyT m (Either Failure String)
Captured stdout, or error in case of failures

propertyOnce :: PropertyT (ResourceT IO) () -> Property

withSnd :: (a -> b) -> a -> (a, b)

Return the supply value with the result of the supplied function as a tuple

noteInputFile :: (MonadTest m, HasCallStack) => FilePath -> m FilePath

Return the input file path after annotating it relative to the project root directory

noteTempFile :: (MonadTest m, HasCallStack) => FilePath -> FilePath -> m FilePath

Return the test file path after annotating it relative to the project root directory

redactJsonField :: (MonadTest m, MonadIO m, HasCallStack) => Text -> Text -> FilePath -> FilePath -> m ()

watchdogProp :: HasCallStack => Property -> Property